% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPeakCalling.R
\docType{data}
\name{ChIP_bed}
\alias{ChIP_bed}
\title{Peak calling}
\format{An object of class \code{character} of length 2.}
\usage{
ChIP_bed
}
\arguments{
\item{ChIP_bed}{Vector of CHIP bed files path}

\item{Input_bed}{Vector of INPUT bed files path. default= NULL uses local enrichment. Otherwise, 1 input file can be specified or as many files as for ChIP}

\item{binsize}{size of the bins used to call peaks}

\item{Nbins_test}{Number of INPUT bins used to average INPUT signal and perform Fishwer}

\item{BSgenome}{BSgenome object used for binning the data}

\item{collapse_touching_peaks}{If set to FALSE, return all bins with related padj and OR. Else, returns collapsed reads with max OR and -log10(padj) (Default).}
}
\value{
Collapse coor data.table
}
\description{
Compute peaks from bed files
}
\examples{
ex <- GRanges(c("chr3L", "chr3L", "chr3L", "chr3L", "chr2R"), 
IRanges(c(1000, 2000, 2100, 2200, 2000), 
c(1500, 2099, 2199, 2299, 3000)),
strand= c("+","+","+","-","+"))
vl_collapse_DT_ranges(ex, mingap= 1, stranded= F)
vl_collapse_DT_ranges(ex, mingap= 1, stranded= T)

}
\keyword{datasets}
