% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{plot.bw_heatmap}
\alias{plot.bw_heatmap}
\title{bw heatmap plot only}
\usage{
\method{plot}{bw_heatmap}(obj, center_label = "TSS", col = c("blue", "yellow", "white"))
}
\arguments{
\item{obj}{An object returned by the vl_heatmap_bw_track() function. The object can easily be modified to change colors and so on...}

\item{center_label}{Label for the center of the heatmaps}

\item{col}{Color vector to be used for heatmaps. col= c("blue", "yellow", "white")}
}
\description{
Plotting method for bw_heatmap retuned by vl_heatmap_bw_track output object
}
