% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot.vl_heatmap_obj}
\alias{plot.vl_heatmap_obj}
\title{Plot heatmap object}
\usage{
plot.vl_heatmap_obj(obj, add, add_inherit_row_order = T,
  add_inherit_col_order = F, breaks = seq(min(obj$x, na.rm = T),
  max(obj$x, na.rm = T), length.out = length(col)), show_rownames = T,
  show_colnames = T, col = c("cornflowerblue", "white", "red"),
  main = NA, auto_margins = T, cutree_rows = 1,
  show_row_dendrogram = T, show_row_clusters = T, cutree_cols = 1,
  show_col_dendrogram = T, show_col_clusters = T,
  display_numbers = F, display_numbers_FUN = function(x) round(x, 2),
  display_numbers_cex = 0.5, legend_title = NA)
}
\arguments{
\item{obj}{Object of class vl_heatmap_obj such as returned by vl_heatmap}

\item{add}{Matrix of new values to plot}

\item{add_inherit_row_order}{Should add data inherit from row order? Default= T}

\item{add_inherit_col_order}{Should add data inherit from col order? Default= F}

\item{breaks}{Breaks for colors}

\item{show_rownames}{show labels}

\item{show_colnames}{show labels}

\item{col}{Color scale. default= c("cornflowerblue", "white", "red")}

\item{main}{Title. Default= NA}

\item{auto_margins}{Should auto margins be computed and used?}

\item{cutree_rows}{Number of cuts for rows tree. Default= 1L}

\item{show_row_dendrogram}{Plot row dendrogram?}

\item{show_row_clusters}{Plot row cluster names/lines?}

\item{cutree_cols}{Number of cuts for cols tree. default= 1L}

\item{show_col_dendrogram}{Plot col dendrogram?}

\item{show_col_clusters}{Plot col cluster names/lines?}

\item{display_numbers}{Display numbers on heatmap? Default= F}

\item{display_numbers_FUN}{Function to apply before displaying numbers on heatmap.}

\item{display_numbers_cex}{cex display numbers}

\item{legend_title}{Character to plot as title legend}
}
\description{
Plot heatmap object
}
