% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vboxplot.R
\name{vboxplot}
\alias{vboxplot}
\title{violin plot}
\usage{
vboxplot(x, formula, outline = F, xlab = "", ylab = "", xlim = c(1,
  length(unique(obj$variable))), ylim = if (outline)
  range(unlist(obj$value)) else range(c(obj$Q1, obj$Q5)))
}
\arguments{
\item{x}{Matrix, data.atble or data.frame containing data}

\item{formula}{Formula to apply to the data}

\item{outline}{plot outlier values? default= F}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}
}
\description{
This function plots a violin plot from either matrix,
data.table, data.frame or using formula.
}
