% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_plot.R
\name{vl_alluvial_plot}
\alias{vl_alluvial_plot}
\title{bw Average tracks}
\usage{
vl_alluvial_plot(
  dat,
  col = vl_palette_many_categ(length(unique(unlist(dat)))),
  alpha.f = 0.5,
  ...
)
}
\arguments{
\item{dat}{Data.table or matrix containing 1 column/condition}

\item{col}{Colors to use for bars. Wrapped with colorRampPalette}

\item{alpha.f}{Alpha value for transition colors}

\item{...}{Extra parameters to be passed to barplot}
}
\value{
An object that can be used with the vl_average_bw_track_plot_only() function.
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\examples{
test <- data.table(c("A", "A", "A", "A", "A", "B", "B", "B", "C", "C", "C"),
c("A", "A", "A", "B", "B", "B", "B", "C", "C", "C", "C"),
c("A", "A", "B", "B", "B", "B", "B", "B", "B", "C", "C"))
vl_alluvial_plot(test)
}
