% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_average_bw_track}
\alias{vl_average_bw_track}
\title{bw Average tracks}
\usage{
vl_average_bw_track(
  bed,
  set_IDs = NULL,
  tracks,
  extend = c(-5000, 5000),
  stranded = F,
  nbins = 501,
  names = NULL,
  plot = T,
  xlab = "genomic distance",
  ylab = "Enrichment",
  ylim = NULL,
  col = NULL,
  axes = T,
  legend = T
)
}
\arguments{
\item{bed}{Regions to plot. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{extend}{How much should the bed regions be extended (starting from the center). dEfault= c(-5000, 5000)}

\item{stranded}{Should the average track be stranded? If yes, - features are reversed :D}

\item{nbins}{Number of bins spanning the extended regions. Default= 501}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}

\item{plot}{Should the average track be ploted? default= T}

\item{xlab}{X label. default= "genomic distance"}

\item{ylab}{Y labels. default= "Enrichment"}

\item{ylim}{ylim for plotting. default= range(data)}

\item{col}{Colors to use. If specified, should match length(tracks)*length(unique(set_IDs)). default= NULL (redirects to rainbow)}

\item{axes}{Should the x axis be plotted? Default to T}

\item{legend}{Should the legen be plotted? default to T}
}
\value{
An object that can be used with the vl_average_bw_track_plot_only() function.
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\examples{
Unstranded 
STARR <- fread("/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/peaks/DSCP_600bp_gw_cut_merged.peaks.txt", select = c(1,2))
STARR <- STARR[, .(seqnames= V1, start= V2, end= V2)]
obj1 <- vl_average_bw_track(bed= STARR,
                            set_IDs = c(rep("High", 2000), rep("Low", 2781)),
                            #' extend = c(-5000, 5000),
                            stranded = F,
                            tracks= "/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/bw/DSCP_600bp_gw_cut_merged.bw", 
                            col= c("lightgrey", "red"), 
                            names = c("STARR-Seq"))

Stranded 
prom <- readRDS("/groups/stark/hendy/Projects/Chromatin_remodeler_specificity/Analyses/Rdata/unique_proms_merged.df.RDS")
prom <- as.data.table(GRanges(unique(prom$Flybase.TSS)))
obj2 <- vl_average_bw_track(bed= prom,
                            extend = c(-5000, 5000),
                            stranded = T,
                            tracks= "/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE81795_H3K4me3_rep1_uniq.bw")
}
