% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vAverageTrackbwPlotOnly.R
\name{vl_average_bw_track_plot_only}
\alias{vl_average_bw_track_plot_only}
\title{bw Average tracks plot only}
\usage{
vl_average_bw_track_plot_only(
  obj,
  ylim = NULL,
  xlab = "genomic distance",
  ylab = "Enrichment",
  axes = T,
  legend = T
)
}
\arguments{
\item{obj}{An object returned by the vl_average_bw_track() function. The object can easily be modified to change colors and so on...}

\item{ylim}{ylim for plotting. default= range(data)}

\item{xlab}{X label. default= "genomic distance"}

\item{ylab}{Y labels. default= "Enrichment"}

\item{axes}{Should the x axis be plotted? Default to T}

\item{legend}{Should the legend be plotted? default to T}
}
\description{
Plots average tracks from a vl_average_bw_track output object
}
\examples{
Unstranded 
STARR <- fread("/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/peaks/DSCP_600bp_gw_cut_merged.peaks.txt", select = c(1,2))
STARR <- STARR[, .(seqnames= V1, start= V2, end= V2)]
obj1 <- vl_average_bw_track(bed= STARR,
                            set_IDs = c(rep("High", 2000), rep("Low", 2781)),
                            #' extend = c(-5000, 5000),
                            stranded = F,
                            tracks= "/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/bw/DSCP_600bp_gw_cut_merged.bw", 
                            col= c("lightgrey", "red"), 
                            names = c("STARR-Seq"), 
                            plot=F)
vl_average_bw_track_plot_only(obj = obj1)

Stranded 
prom <- readRDS("/groups/stark/hendy/Projects/Chromatin_remodeler_specificity/Analyses/Rdata/unique_proms_merged.df.RDS")
prom <- as.data.table(GRanges(unique(prom$Flybase.TSS)))
obj2 <- vl_average_bw_track(bed= prom,
                            extend = c(-5000, 5000),
                            stranded = T,
                            tracks= "/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE81795_H3K4me3_rep1_uniq.bw",
                            plot= F)
vl_average_bw_track_plot_only(obj = obj2)
}
