% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloons_plot.R
\name{vl_balloons_plot}
\alias{vl_balloons_plot}
\alias{vl_balloons_plot.matrix}
\title{Ballons plot}
\usage{
vl_balloons_plot(x, ...)

\method{vl_balloons_plot}{matrix}(
  x,
  color.var,
  x.breaks,
  color.breaks,
  col = c("#440154FF", "#472D7BFF", "#3B528BFF", "#2C728EFF", "#21908CFF", "#27AD81FF",
    "#5DC863FF", "#AADC32FF", "#FDE725FF"),
  cex.balloons = 1,
  main,
  xaxt = "s",
  yaxt = "s",
  balloon.size.legend = "Size",
  balloon.col.legend = "Color",
  legend.left.pos,
  gap.axis = 0
)
}
\arguments{
\item{x}{Matrix specifyiung size of the balloons.}

\item{color.var}{Matrix specifyiung color of the balloons.}

\item{color.breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{cex.balloons}{Expansion factor for balloons. default= 1.}

\item{main}{Title. Default= NA}

\item{xaxt}{see ?par()}

\item{yaxt}{see ?par()}

\item{balloon.size.legend}{Title size legend}

\item{balloon.col.legend}{Title color legend}

\item{legend.left.pos}{If specified, overrides default legend left user coordinates}
}
\value{
Balloon plot
}
\description{
Starting from two matrices (dot size and colors, respectively), generates a balloon plot
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_balloons_plot(matrix)}: matrix_method

}}
\examples{
x <- color.var <- matrix(-3:5, ncol= 3)
vl_balloons_plot(x= x,
                 color.var= color.var)

}
