% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_balloonskey}
\alias{vl_balloonskey}
\title{Heatkey plot}
\usage{
vl_balloonskey(sizes, labels, left = par("usr")[2],
  top = par("usr")[4], ticks.cex = 0.6, main.cex = 1, main = NA)
}
\arguments{
\item{sizes}{Balloons sizes (cex)}

\item{labels}{Labels corresponding to each size}

\item{left}{left pos}

\item{top}{top pos}

\item{ticks.cex}{cex expansion factor for ticks labels}

\item{main.cex}{cex expansion factor for legend title}

\item{main}{Title}

\item{height}{height}
}
\value{
Plots heatkey
}
\description{
Heatkey plot
}
