% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{vl_barplot}
\alias{vl_barplot}
\title{Title}
\usage{
vl_barplot(
  height,
  sd = NULL,
  xlim = NULL,
  ylim = NULL,
  individual.var = NULL,
  ind.pch = 16,
  ind.col = adjustcolor("lightgrey", 0.7),
  ind.jitter = 0.2,
  compute.diff = NULL,
  diff.cex = 0.8,
  xpd = NA,
  horiz = F,
  ...
)
}
\arguments{
\item{height}{A vector of values describing the bars which make up the plot.}

\item{sd}{sd values to be plotted as whiskers around bars}

\item{xlim}{}

\item{ylim}{}

\item{individual.var}{List of individual variables to be plotted as points.}

\item{ind.pch}{pch value for individual variables}

\item{ind.col}{color value for individual variables}

\item{ind.jitter}{jitter value for individual variables}

\item{compute.diff}{List of pairwise bar pairs to be compared (V2/V11 ratio)}

\item{diff.cex}{cex value for pairwise comparisons}

\item{xpd}{xpd value for pairwaise comparisons}

\item{horiz}{Not supported atm}

\item{...}{Extra arguments to be passed to barplot}
}
\value{
A simple barplot
}
\description{
Title
}
\examples{
vl_barplot(1:3, rep(.2, 3))
}
