% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwrappers.R
\name{vl_bash_wrap_windows}
\alias{vl_bash_wrap_windows}
\title{Wrap commands to send to bash on windows}
\usage{
vl_bash_wrap_windows(cmd)
}
\description{
Wrap commands to send to bash on windows
}
