% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_bedEnrichment}
\alias{vl_bedEnrichment}
\title{Compute Enrichment}
\usage{
vl_bedEnrichment(bins, ChIP_bed, Input_bed = NULL)
}
\arguments{
\item{bins}{Data.table containing bins (seqnames, start and end columns)}

\item{ChIP_bed}{Either a character vector poiting to bed file path OR a data.table containing ChIP reads}

\item{Input_bed}{Either a character vector poiting to bed file path OR a data.table containing Input reads. 
If set to NULL (default), use shuffled version of the ChIP reads as input.}
}
\value{
original bins file with OR, pval and padj corresponding to fisher result
}
\description{
Compute enrichment for a list of regions using fisher test to compare ChIP and Input
}
\examples{
ChIP_bed <- "/mnt/d/_R_data/projects/epigenetic_cancer/db/bed/cutnrun/merge/H3K27me3_PH18_merge_uniq.bed"
res <- vl_bedEnrichment(ChIP_bed)

}
