% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_bowtie1_align}
\alias{vl_bowtie1_align}
\title{BOWTIE Alignment to custom genome}
\usage{
vl_bowtie1_align(
  fq1,
  index_prefix,
  sam_output,
  o = NA,
  e = NA,
  execute = F,
  cores = 6,
  d = NA
)
}
\description{
BOWTIE Alignment to custom genome
}
