% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwrappers.R
\name{vl_bowtie2_align}
\alias{vl_bowtie2_align}
\title{BOWTIE2 Alignment to custom genome}
\usage{
vl_bowtie2_align(fq1, fq2 = NA, index_prefix, sam_output, cores = 6)
}
\description{
BOWTIE2 Alignment to custom genome
}
