% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\alias{vl_boxplot.formula}
\alias{vl_boxplot.data.table}
\title{violin plot}
\usage{
vl_boxplot(x, ...)

\method{vl_boxplot}{formula}(formula, x, ...)

\method{vl_boxplot}{data.table}(x, ...)
}
\arguments{
\item{x}{Matrix, data.atble or data.frame containing data}

\item{formula}{Formula to apply to the data}

\item{outline}{plot outlier values? default= F}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{plot_labels}{Should box labels be plotted? Default= T}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}
}
\description{
This function plots a violin plot from either matrix,
data.table, data.frame or using formula.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: method for formula

\item \code{data.table}: method for data.table
}}

