% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_bsub}
\alias{vl_bsub}
\title{Stark lab wrappers}
\usage{
vl_bsub(
  cmd,
  cores = 1L,
  name = "vl",
  d = NA,
  m = NA,
  o = NA,
  e = NA,
  t = NA,
  execute = T
)
}
\arguments{
\item{cmd}{Command to be submitted}

\item{cores}{Number of cores. default= 1L}

\item{name}{Name of the job. Default= "vl}

\item{d}{Name of te dependent job}

\item{m}{Memory in Go. Example= 16}

\item{o}{Std output directory}

\item{e}{Std error directory}

\item{t}{Time parameter for slurm in the format hh:mm:ss. Example 2 days:'2-00:00:00'}

\item{execute}{If FALSE, simply returns the full command. Default= TRUE}
}
\description{
bsub submits to the cluster
}
