% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_coverage}
\alias{vl_bw_coverage}
\title{bw coverage}
\usage{
vl_bw_coverage(bed, bw, na.value = 0)
}
\arguments{
\item{bed}{Regions to quantify. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{bw}{Path to target bw file (character vector)}

\item{na.value}{Value to use for NAs. default to 0}
}
\description{
Quantify bw file at a set of intervals. returns mean value/region
}
\details{
Note that strand-specific overlap is not implemented! bw files do not contain strand info!?
}
\examples{
bins <- vl_binBSgenome(genome= "dm3",
                       restrict.seqnames = "chr3R")
cov <- vl_bw_coverage(bins, track)

# GRanges method
binned_average_function <- function(gr, file_bw_path){
file_bw <- import.bw(file_bw_path, as="RleList")
seqlevels(gr) <- names(file_bw)
t1 <- Sys.time()
bins1 <- binnedAverage(gr,file_bw, "average_score")
return(bins1$average_score)
}
cov <- binned_average_function(GRanges(bins), track)

}
