% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_coverage_bins}
\alias{vl_bw_coverage_bins}
\title{Bins bed file and compute bw signal}
\usage{
vl_bw_coverage_bins(
  bed,
  set.IDs,
  tracks,
  upstream,
  downstream,
  ignore.strand,
  genome,
  nbins,
  names
)
}
\arguments{
\item{bed}{Regions to plot. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{set.IDs}{Set IDs specifying the groups as subsets of the bed file}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{upstream}{Upstream  extension of bed regions (centered on center)}

\item{downstream}{Downstream  extension of bed regions (centered on center)}

\item{ignore.strand}{Should the strand be ignored?}

\item{genome}{BSgenome used to check limits of extended regions. Out of limit ranges will be resized accordingly}

\item{nbins}{Number of bins spanning the extended regions. Default= 500}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}
}
\description{
Compared to vl_bw_coverage, this functions bins bed file before quantifying the signal. Useful to make heatmaps/average tracks
}
