% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_enrich}
\alias{vl_bw_enrich}
\title{Compute ChIP enrichment}
\usage{
vl_bw_enrich(regions, ChIP_bw, Input_bw, read_length = 50)
}
\arguments{
\item{regions}{Regions to analyse. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{ChIP_bw}{ChIP bw files paths. If several provided, cat}

\item{Input_bw}{Input bw files paths. If several provided, cat}

\item{read_length}{Used to estimate total reads. default to 50}
}
\value{
original bins file with OR, pval and padj corresponding to fisher result
}
\description{
compared to enrichBed, uses bw files as Input_bw
}
