% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCalling.R
\name{vl_bw_totalReads}
\alias{vl_bw_totalReads}
\title{bw total reads}
\usage{
vl_bw_totalReads(bw, read_length)
}
\arguments{
\item{bw}{Path to target bw file(s) (character vector)}

\item{read_length}{length of the reads used to generate bw file}
}
\description{
IN DEVELOPMENT!! -> Estimate total reads within bw file
}
