% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPSeqPipeline.R
\name{vl_ChIP_pipeline}
\alias{vl_ChIP_pipeline}
\title{Process ChIP-Seq/ATAC-Seq data}
\usage{
vl_ChIP_pipeline(fq1, fq2 = NULL, chrom_sizes, basename_prefix = "",
  Rsubread_index_prefix = NULL, bam_folder = NULL, bw_folder,
  extend = 300, max_frag_size_pe = 600, maxMismatches = 2,
  nTrim3 = 0, nTrim5 = 0, gzip_sam_output = F)
}
\arguments{
\item{fq1}{fastq file path (read1)}

\item{fq2}{fastq file path (read2). Default is null -> fq1 treated as single end!}

\item{chrom_sizes}{data.table containing "seqnames" and "seqlengths" columns.}

\item{basename_prefix}{prefix to add at the beginning of all output files}

\item{Rsubread_index_prefix}{Rsubread index to use for alignment. Attached seqnames and seqlenths will be used.}

\item{bam_folder}{Folder for output bam files.}

\item{bw_folder}{Folder for output bw files.}

\item{extend}{length to which single end reads should be extended (bed and bw)}

\item{maxMismatches}{default 2}

\item{nTrim3}{default 0}

\item{nTrim5}{default 0}

\item{gzip_sam_output}{Should sam alignment file be gzipped? default= F}

\item{use_samtools}{If FALSE, avoids using samtools by saving alignments in a non-binary sam file. default= TRUE.}
}
\value{
bam , bam_stats.txt, uniq.bed and .bw files
}
\description{
Starts from fastq files, outputs bam, bam statistics, bed containing collapsed reads and bw
}
\examples{
require(data.table)
require(GenomicRanges)
require(Rsubread)
vl_ChIP_pipeline(fq1= "/groups/stark/vloubiere/projects/available_data_dm3/test/SRR1042411.fastq.gz",
basename_prefix = "test_single_end",
Rsubread_index_prefix = "/groups/stark/vloubiere/genomes/Rsubread_genome/dm3/dm3_genome/index/dm3_genome_idx",
bam_folder = "/groups/stark/vloubiere/projects/available_data_dm3/test/",
bw_folder = "/groups/stark/vloubiere/projects/available_data_dm3/test/")
vl_ChIP_pipeline(fq1= "/groups/stark/vloubiere/projects/available_data_dm3/db/fastq/SRR7813063_1.fastq.gz",
fq2= "/groups/stark/vloubiere/projects/available_data_dm3/db/fastq/SRR7813063_2.fastq.gz",
basename_prefix = "test_paired_end",
Rsubread_index_prefix = "/groups/stark/vloubiere/genomes/Rsubread_genome/dm3/dm3_genome/index/dm3_genome_idx",
bam_folder = "/groups/stark/vloubiere/projects/available_data_dm3/test/",
bw_folder = "/groups/stark/vloubiere/projects/available_data_dm3/test/")
}
