% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_collapseBed}
\alias{vl_collapseBed}
\title{Collapse ranges with a certain min dist}
\usage{
vl_collapseBed(bed, mingap = 1, return_idx_only = F)
}
\arguments{
\item{bed}{bed file to be collapsed. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames','start', 'end' columns. see ?vl_importBed()}

\item{mingap}{min gap ditance for merging. default is 1 (that is, touching coordinates)}

\item{return_idx_only}{If set to T, does not collapse regions but returns idx as an extra columns. default= F}
}
\value{
Collapse coor data.table
}
\description{
Collapse ranges with a certain min dist between them
}
\examples{
bed <- data.table(seqnames= c("chr3L", "chr3L", "chr3L", "chr3L", "chr3L", "chr2R", "chr2R", "chr2R"),
start= c(1000, 2000, 2100, 2200, 2500, 2000, 2500, 5000),
end= c(1500, 2099, 2199, 2299, 3000, 3000, 3500, 6000))
vl_collapseBed(bed, mingap= 1)
}
