% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vComputeEnrichment.R
\name{vl_computeEnrichment}
\alias{vl_computeEnrichment}
\title{Compyte enrichment}
\usage{
vl_computeEnrichment(
  ChIP_bed,
  Input_bed = NULL,
  peaks,
  ext_peaks = 10000,
  BSgenome = BSgenome.Dmelanogaster.UCSC.dm3
)
}
\arguments{
\item{ChIP_bed}{Vector of CHIP bed files path}

\item{Input_bed}{Vector of INPUT bed files path. default= NULL uses local enrichment. Otherwise, 1 input file can be specified or as many files as for ChIP}

\item{peaks}{DT containing peaks}

\item{ext_peaks}{size extension to compute enrichment at peak}

\item{BSgenome}{BSgenome object used for binning the data}
}
\value{
enrichment DT
}
\description{
Compute peaks from bed files
}
\examples{
ChIP_bed <- c("../available_data_dm3/db/bed//GSE119708_ATAC_rep1_uniq.bed", "../available_data_dm3/db/bed//GSE119708_ATAC_rep2_uniq.bed")

}
