% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_contrib_addMotif}
\alias{vl_contrib_addMotif}
\title{Add motif columns to contribution object}
\usage{
vl_contrib_addMotif(
  contrib,
  pwm_log_odds = vl_motifs_DB_v2[collection == "jaspar", pwms_log_odds],
  p.cutoff = 5e-05,
  genome,
  bg = "genome"
)
}
\arguments{
\item{contrib}{A contribution object generated by ?vl_importContrib()}

\item{pwm_log_odds}{A list of pwms (log_odds) for which matchin positions should be returned.}

\item{p.cutoff}{The cutoff to be used for calling motifs. Default= 5e-5}

\item{genome}{The genome to be used}

\item{bg}{background. Default= "genome".}

\item{motifDB}{The motifDB to be used. Default= vl_motifs_DB_v2}
}
\value{
A contribution data.table containing, for each motif_ID, a vector corresponding to its matches.
}
\description{
Add motif columns to contribution object
}
\examples{
dat <- vl_importContrib(h5= list.files("db/model_PH18/contributions/", "h5$", recursive = TRUE, full.names = TRUE))
mot <- vl_contrib_addMotif(dat, p.cutoff= 1e-4, genome= "dm6")

}
