% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_contrib_enrich}
\alias{vl_contrib_enrich}
\title{Compute enrichment for each motif instance}
\usage{
vl_contrib_enrich(obj = vl_PH18_contrib, mot_columns = NULL)
}
\arguments{
\item{obj}{}
}
\description{
Compute enrichment for each motif instance
}
