% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSgenomeTools.R
\name{vl_control_regions_BSgenome}
\alias{vl_control_regions_BSgenome}
\title{Generate control regions}
\usage{
vl_control_regions_BSgenome(bed, genome, no.overlap = F)
}
\arguments{
\item{bed}{Bed file used to produce similar control}

\item{genome}{BSgneome object to use. ex: "dm3", "dm6"}

\item{no.overlap}{If set to TRUE, avoids overlap between control sequences and the original bed file. Default= FALSE.}
}
\value{
data.table containing control regions
}
\description{
Generate regions with similar dispersion and widths ditrib than provided bed file
}
\examples{
vl_control_regions_BSgenome(vl_SUHW_top_peaks, "dm3")

}
