% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_covBed}
\alias{vl_covBed}
\title{Compute bed coverage}
\usage{
vl_covBed(bins, bed)
}
\arguments{
\item{bins}{bins for which enrichment has to be computed. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{bed}{bed file. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}
}
\value{
numeric vector of the number of overlapping reads
}
\description{
For each bin, computes the number of overlapping reads from a bed file
}
