% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_CP_motifs_DB}
\alias{vl_CP_motifs_DB}
\title{Core Promoter motifs data base}
\format{
An object containing Motifs and related metadata
\describe{
  \item{Pwms_log_odds}{PWM expressed as log odds}
  \item{Pwms_perc}{PWM expressed as percentage}
  ...
}
}
\source{
{"/groups/stark/almeida/data/motifs/motif_collection_v7_no_transfac_SteinAerts/TF_clusters_PWMs.RData"}
}
\usage{
Can be used with the matchMotifs function from the motifmatchr package (see example)
}
\description{
Contains 19 core promoter motifs (Vanja + Bernie)
}
\examples{
hit <- matchMotifs(vl_CP_motifs_DB$Pwms_log_odds, GRanges("chrX", IRanges(10000000, 10000500)), genome= "dm3", p.cutoff= 5e-4, bg= "even", out= "scores")
counts <- as.matrix(motifCounts(hit))
colnames(counts) <- name(vl_CP_motifs_DB$Pwms_log_odds)
counts <- as.data.table(counts)

}
\keyword{datasets}
