% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIP_CUT&RUN_pipeline.R
\name{vl_CUTNRUN_peakCalling}
\alias{vl_CUTNRUN_peakCalling}
\title{Title}
\usage{
vl_CUTNRUN_peakCalling(
  processed_metadata,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  extsize = 300,
  cores = 8,
  mem = 32,
  overwrite = F,
  submit = F,
  wdir = getwd(),
  logs = "db/logs/CUTNRUN/peak_calling"
)
}
\arguments{
\item{processed_metadata}{Processed metadata file generated by vl_CUTNRUN_processing}

\item{Rpath}{The path to the Rscript executable to use. Default= /software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript}

\item{cores}{Number of cores per job. Default= 8}

\item{mem}{Memory per job (in Go). Default= 32.}

\item{overwrite}{Should existing files be overwritten?}

\item{submit}{Should the command be submitted? default= G}

\item{wdir}{The working directory to use. defaut= getwd().}

\item{logs}{Path to save logs. Default= "db/logs/CUTNRUN/peak_calling"}

\item{extsise}{The extsize to be used, meaning that the building of the peak model will be skipped (--nomodel). To build the model, set extize= NA. Default= 300}
}
\description{
Title
}
