% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdigest.R
\name{vl_digest}
\alias{vl_digest}
\title{Digest sequence}
\usage{
vl_digest(seq, enzyme)
}
\arguments{
\item{seq}{character sequence to digest}

\item{enzyme}{Character vector containing the enzymes to use}
}
\value{
A vector containing the digested pieces
}
\description{
Simulate DNA digest from thermofisher "simple" enzymes
}
\examples{
vl_digest(seq= "AAAAAAAAGGTACCTTTTTTTTTTTTGCGGCCGCAAAAAAAAAA",
enzyme= c("KpnI", "NotI"))
}
