% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_Dmel_GO_FB2020_05}
\alias{vl_Dmel_GO_FB2020_05}
\title{Dm6 GO database}
\format{
data.table containing GO metadata
}
\usage{
Can be used with vl_GO_cluster. see ?vl_GO_cluster()
}
\description{
Contains GO annotation from FB2020_05
}
\examples{
GO: GO id
name: GO full name
type: GO type (biological_process, molecular_function, cellular_component)
FBgn: gene ID
Symbol: gene symbol

Object was generated using the following code:
tmp <- tempfile(fileext = ".gz")
download.file("http://ftp.flybase.net/releases/FB2020_05/precomputed_files/go/gene_association.fb.gz",
destfile = tmp)
go <- fread(tmp,
skip= 5,
select = c(2,3,5),
col.names = c("FBgn", "Symbol", "GO"))
tmp <- tempfile(fileext = ".gz")
download.file("http://ftp.flybase.net/releases/FB2020_05/precomputed_files/ontologies/go-basic.obo.gz",
destfile = tmp)
details <- ontologyIndex::get_ontology(tmp, extract_tags = "everything")
details <- data.table(GO= details$id,
name= details$name,
type= details$namespace)
vl_Dmel_GO_FB2020_05 <- merge(details, unique(go))
vl_Dmel_GO_FB2020_05 <- vl_Dmel_GO_FB2020_05[!(name \%in\% type)]
vl_Dmel_GO_FB2020_05[, type:= unlist(type)]
save(vl_Dmel_GO_FB2020_05,
file= "/groups/stark/vloubiere/vlfunction/data/vl_Dmel_GO_FB2020_05.RData")
}
\keyword{datasets}
