% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_Dmel_motifs_DB}
\alias{vl_Dmel_motifs_DB}
\title{Dmel motifs data base}
\format{
An object containing 13899 motifs and related metadata
\describe{
  \item{Motif_cluster}{901 clusters containing redunsant motifs. Used collections: bergman, cisbp, flyfactorsurvey, homer, jaspar, stark, idmmpmm.}
  \item{Motif_cluster_name}{Cluster names with the respective motif types.}
  \item{Pwms_log_odds}{PWM expressed as log odds}
  \item{Pwms_perc}{PWM expressed as percentage}
}
}
\source{
{"/groups/stark/almeida/data/motifs/motif_collection_v7_no_transfac_SteinAerts/TF_clusters_PWMs.RData"}
}
\usage{
Can be used with the matchMotifs function from the motifmatchr package (see example)
}
\description{
Contains 13899 motifs usable for motif enrichments in Dmel
}
\examples{
# Select convenient set of Drosophila TFs
sel <- vl_Dmel_motifs_DB$metadata[!is.na(vl_Dmel_motifs_DB$metadata$Dmel) & # Associated to a known TF
vl_Dmel_motifs_DB$metadata$X..motif_collection_name \%in\% # From a relevant DB
c("flyfactorsurvey", "bergman", "jaspar", "idmmpmm", "cisbp"), "motif_name"]
sel <- which(name(vl_Dmel_motifs_DB$All_pwms_log_odds) \%in\% sel)
hit <- matchMotifs(vl_Dmel_motifs_DB$All_pwms_log_odds[sel], GRanges("chrX", IRanges(10000000, 10000500)), genome= "dm3", p.cutoff= 5e-4, bg= "even", out= "scores")
counts <- as.matrix(motifCounts(hit))
colnames(counts) <- name(vl_Dmel_motifs_DB$All_pwms_log_odds[sel])
counts <- as.data.table(counts)

}
\keyword{datasets}
