% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_Dmel_motifs_DB_full}
\alias{vl_Dmel_motifs_DB_full}
\title{Dmel motifs data base}
\format{
An object containing 13899 motifs and related metadata
\describe{
  \item{motif_name}{Motif ID from TF_clusters_PWMs.RData}
  \item{FBgn}{Curated FBgn symbols. sep= "/"}
  \item{motif_cluster}{Motif cluster from Bernardo Almeida}
  \item{collection}{}
  \item{collection_version}{}
  \item{species}{PWM expressed as log odds}
  \item{pwms_log_odds}{list of PWM matrices}
  \item{pwms_perc}{list of PWM matrices}
}
}
\source{
{"/groups/stark/almeida/data/motifs/motif_collection_v7_no_transfac_SteinAerts/TF_clusters_PWMs.RData"}
}
\usage{
Can be used with the ?vl_motif_counts() function to count motifs
}
\description{
Contains 13899 motifs usable for motif enrichments in Dmel
}
\keyword{datasets}
