% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_Dmel_STRING_DB}
\alias{vl_Dmel_STRING_DB}
\title{Dm6 STRING database}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2171887 rows and 13 columns.}
\usage{
Can be used with ?vl_STRING_interaction()
}
\description{
Contains STRING annotations v11.5
}
\examples{
Code used to create DB
# Download STRING db v11.5 from Dmel
link <- "https://stringdb-static.org/download/protein.links.detailed.v11.5/7227.protein.links.detailed.v11.5.txt.gz"
tmp <- tempfile(fileext = ".txt.gz")
download.file(link, tmp)
DB <- fread(tmp)
# Download correspondance Dmel symbols
link <- "https://stringdb-static.org/download/protein.info.v11.5/7227.protein.info.v11.5.txt.gz"
download.file(link, tmp)
sym <- fread(tmp)
DB[sym, protein1_symbol:= i.preferred_name, on= "protein1==`#string_protein_id`"]
DB[sym, protein2_symbol:= i.preferred_name, on= "protein2==`#string_protein_id`"]
# Collapse unique interactions
DB[, merge:= paste0(sort(c(protein1_symbol, protein2_symbol)), collapse= ""), .(protein1_symbol, protein2_symbol)]
vl_Dmel_STRING_DB <- DB[, .SD[1], merge]
# SAVE
save(vl_Dmel_STRING_DB,
file = "/groups/stark/vloubiere/vlfunction/data/vl_Dmel_STRING_DB.RData", 
compress = "bzip2")
}
\keyword{datasets}
