% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRRTools.R
\name{vl_download_SRA}
\alias{vl_download_SRA}
\title{Title}
\usage{
vl_download_SRA(
  SRR,
  layout,
  output_folder,
  tmp_folder,
  threads = 8,
  mem = 2048
)
}
\arguments{
\item{SRR}{}

\item{layout}{}

\item{output_folder}{}

\item{tmp_folder}{}

\item{threads}{}

\item{mem}{}
}
\description{
Title
}
