% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_enrichBed}
\alias{vl_enrichBed}
\title{Compute ChIP enrichment}
\usage{
vl_enrichBed(regions, ChIP_bed, Input_bed)
}
\arguments{
\item{regions}{Regions to analyse. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{ChIP_bed}{ChIP bed file. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{Input_bed}{Input bed file.}
}
\value{
original bins file with OR, pval and padj corresponding to fisher result
}
\description{
Given a set of ChIP and Input counts, performs hypergeometric test to know if ChIP is enriched compared to Input
}
