% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_exportBed}
\alias{vl_exportBed}
\alias{vl_exportBed.GRanges}
\alias{vl_exportBed.data.table}
\title{Export bed file}
\usage{
vl_exportBed(bed, ...)

\method{vl_exportBed}{GRanges}(bed, filename)

\method{vl_exportBed}{data.table}(bed, filename)
}
\arguments{
\item{bed}{Either a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}
}
\description{
Export bed file
}
\section{Methods (by class)}{
\itemize{
\item \code{GRanges}: for GRanges

\item \code{data.table}: for data.table
}}

