% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_fb_go_table_dm6_FB2020_05}
\alias{vl_fb_go_table_dm6_FB2020_05}
\title{Dm6 GO database}
\format{
An object containing Motifs and related metadata
\describe{
  \item{GO}{GO ID}
  \item{name}{GO name}
  \item{type}{GO category}
  \item{partents}{parent GOs}
  \item{children}{children GOs}
  \item{Fbgn}{Gene flybase ID}
  \item{symbol}{Gene symbol}
  ...
}
}
\source{
{"../../vlfunction/data/vl_fb_go_table_dm6_FB2020_05.RData"}
}
\usage{
Can be used with vl_GO_cluster. see ?vl_GO_cluster()
}
\description{
Contains GO annotation from FB2020_05
}
\examples{
Object was generated using the followin code:

}
\keyword{datasets}
