% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfiglabel.R
\name{vl_fig_label}
\alias{vl_fig_label}
\title{figure label}
\usage{
vl_fig_label(text, region = "figure", pos = "topleft", cex = NULL,
  ...)
}
\arguments{
\item{text}{Text to plot.}

\item{region}{Can be "figure", "plot", "device"}

\item{cex}{Scaling factor}

\item{...}{Extra args for text function}
}
\description{
Plots a convenient label for paper figures
}
