% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_blur.R
\name{vl_gaussian_blur}
\alias{vl_gaussian_blur}
\title{applies a gasussian blur}
\usage{
vl_gaussian_blur(x)
}
\arguments{
\item{x}{A numeric vector of the signal to be blurred}
}
\value{
Smoothed signal
}
\description{
radius= +/- 10 bins
}
