% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_genes_set}
\alias{vl_genes_set}
\title{Set of genes}
\format{FBgn, symbol and GO (RpL/HOX)}
\usage{
see ?vl_GO_enrich()

Can be used with ?vl_STRING_interaction()
}
\description{
Example set that can be used for GO enrichment. Contain RpL and HOX genes

Contains STRING annotations v11.5
}
\examples{
Code used to create DB
# Download STRING db v11.5 from Dmel
link <- "https://stringdb-static.org/download/protein.links.detailed.v11.5/7227.protein.links.detailed.v11.5.txt.gz"
tmp <- tempfile(fileext = ".txt.gz")
download.file(link, tmp)
DB <- fread(tmp)
# Download correspondance Dmel symbols
link <- "https://stringdb-static.org/download/protein.info.v11.5/7227.protein.info.v11.5.txt.gz"
download.file(link, tmp)
sym <- fread(tmp)
DB[sym, protein1_symbol:= i.preferred_name, on= "protein1==`#string_protein_id`"]
DB[sym, protein2_symbol:= i.preferred_name, on= "protein2==`#string_protein_id`"]
# Collapse unique interactions
DB[, merge:= paste0(sort(c(protein1_symbol, protein2_symbol)), collapse= ""), .(protein1_symbol, protein2_symbol)]
vl_Dmel_STRING_DB <- DB[, .SD[1], merge]
# SAVE
save(vl_Dmel_STRING_DB, file = "../../vlfunction/data/vl_Dmel_STRING_DB.Rdata")
}
\keyword{datasets}
