% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSgenomeTools.R
\name{vl_getSequence}
\alias{vl_getSequence}
\title{Get genomic sequence}
\usage{
vl_getSequence(bed, genome)
}
\arguments{
\item{bed}{Bed file for which regions have to be returned}

\item{genome}{BSgenome ID ("dm3", "dm6"...)}
}
\value{
Character vector of sequences
}
\description{
Returns the sequences of a DTranges
}
\examples{
vl_getSequence(vl_SUHW_top_peaks, "dm3")
}
