% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_clusters_plot_only}
\alias{vl_GO_clusters_plot_only}
\title{GO plotting function}
\usage{
vl_GO_clusters_plot_only(obj, padj_cutoff = 1e-05, N_top = Inf,
  auto_margins = T, cex.balloons = 1, main = NA)
}
\arguments{
\item{obj}{An object as the one produced by ?vl_GO_clusters()}

\item{padj_cutoff}{padjust cutoff applied to GOs}

\item{N_top}{Select only N top GOs/cluster}

\item{auto_margins}{Compute and apply optimal margins. Default= T}

\item{cex.balloons}{cex factor for balloons}

\item{main}{Title}
}
\description{
This function compares GO enrichments for several groups
}
