% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_enrich}
\alias{vl_GO_enrich}
\title{GO analysis}
\usage{
vl_GO_enrich(
  geneIDs,
  geneUniverse_IDs = NULL,
  species,
  plot = T,
  padj_cutoff = 0.05,
  top_enrich = Inf,
  order = "padj",
  breaks = NULL
)
}
\arguments{
\item{geneIDs}{A vector (barplot) of a list (clusters) of gene IDs}

\item{geneUniverse_IDs}{Vector of FBgn IDs to which the analysis will be restricted. If NULL, uses all genes.}

\item{species}{Chose between "Dm" and "Mm"}

\item{plot}{Plot result?}

\item{padj_cutoff}{cutoff for plotting}

\item{top_enrich}{Show only n top enriched motifs}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{breaks}{Color breaks to be used. Defaults to range of filtered padj.}
}
\description{
This function compute GO enrichment in a group of genes
}
\examples{
RpL <- vl_genes_set[GO=="RpL_genes", FBgn]
Hox <- vl_genes_set[GO=="HOX_genes", FBgn]
par(mar= c(4,20,2,6))
vl_GO_enrich(RpL, species= "Dm")
vl_GO_enrich(Hox, species= "Dm")
vl_GO_enrich(list(RpL, Hox), species= "Dm", cex.balloons = 0.3, padj_cutoff= 1e-5, top_enrich = 20)
}
