% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_heatkey}
\alias{vl_heatkey}
\title{Title}
\usage{
vl_heatkey(var, Cc, x, y, main = NA)
}
\arguments{
\item{var}{The variable to which heatkey is referring to}

\item{Cc}{Function taking values within var ranges as input and returning color. Can be created using ?circlize::colorRamp2()}

\item{x}{x position in npc}

\item{y}{y position in npc}

\item{main}{Title}
}
\value{
Plots heatkey
}
\description{
Title
}
