% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{vl_heatmap_add.vl_heatmap}
\alias{vl_heatmap_add.vl_heatmap}
\title{Title}
\usage{
vl_heatmap_add.vl_heatmap(
  obj,
  add,
  inherit_row_cl = T,
  inherit_col_cl = T,
  ...
)
}
\arguments{
\item{obj}{vl_heatmap object containing the parent clustering}

\item{add}{matrix that will inhiherit obj clustering}

\item{inherit_row_cl}{Should row clustering be inherited? Default is TRUE if rownames match}

\item{inherit_col_cl}{Should col clustering be inherited? Default is TRUE if colnames match}

\item{...}{Extra arguments to be passed to vl_heatmap}
}
\value{
a vl_heatmap object whose clustering is inherited from obj
}
\description{
Title
}
