% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_heatmap_bw_track}
\alias{vl_heatmap_bw_track}
\title{bw Average tracks}
\usage{
vl_heatmap_bw_track(
  bed,
  set_IDs = NULL,
  tracks,
  extend = c(-5000, 5000),
  stranded = F,
  nbins = 101,
  names = NULL,
  plot = T,
  center_label = "TSS",
  max_FUN = function(x) quantile(x, 0.995, na.rm = T),
  orderTrackIdx = 1,
  order_FUN = function(x) mean(x, na.rm = T),
  col = c("blue", "yellow", "white")
)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{set_IDs}{Vector specifying group of regions.}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{extend}{How much should the bed regions be extended (starting from the center). dEfault= c(-5000, 5000)}

\item{stranded}{Should the average track be stranded? If yes, - features are reversed :D}

\item{nbins}{Number of bins spanning the extended regions. Default= 501}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}

\item{plot}{Should the heatmap be ploted? default= T}

\item{center_label}{Label for the center of the heatmaps}

\item{max_FUN}{Function used to compute clipping max for each track. default= function(x) quantile(x, 0.995, na.rm =T)}

\item{orderTrackIdx}{yIdx of the tack to be used for ordering the heatmaps. Default=1 (1st left heatmap)}

\item{col}{Color vector to be used for heatmaps. col= c("blue", "yellow", "white")}

\item{oder_FUN}{Function used for regions ordering. Default= function(x) mean(x, na.rm= T)}
}
\value{
An object that can be used with the vl_heatmap_bw_track_plot_only() function.
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
