% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_heatmap_bw_track_plot_only}
\alias{vl_heatmap_bw_track_plot_only}
\title{bw heatmap plot only}
\usage{
vl_heatmap_bw_track_plot_only(obj, center_label = "TSS",
  col = c("blue", "yellow", "white"))
}
\arguments{
\item{obj}{An object returned by the vl_heatmap_bw_track() function. The object can easily be modified to change colors and so on...}

\item{center_label}{Label for the center of the heatmaps}

\item{col}{Color vector to be used for heatmaps. col= c("blue", "yellow", "white")}
}
\description{
Plots heatmap from a vl_heatmap_bw_track output object
}
