% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_import_bam}
\alias{vl_import_bam}
\title{Import bam file with all fields}
\usage{
vl_import_bam(bam_path, ncores = data.table::getDTthreads() - 1)
}
\arguments{
\item{bam_path}{}
}
\description{
Import bam file with all fields
}
