% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_import_bamToBed}
\alias{vl_import_bamToBed}
\title{Import paired bam file as bedpe}
\usage{
vl_import_bamToBed(bam_path, bedpe = F, ed = F)
}
\arguments{
\item{bam_path}{}

\item{bedpe}{Imports bam as BEDPE format. Requires BAM to be grouped or sorted by query.}

\item{ed}{Use BAM edit distance (NM tag) for BED score. Default for BED is to use mapping quality. Default for BEDPE is to use the minimum of the two mapping qualities for the pair. When -ed is used with bedpe= T, the total edit distance from the two mates is reported.}
}
\description{
Import paired bam file as bedpe
}
