% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamTools.R
\name{vl_importBam}
\alias{vl_importBam}
\title{Title}
\usage{
vl_importBam(
  file,
  sel = c("qname", "flag", "rname", "strand", "pos", "qwidth", "mapq", "mrnm", "mpos",
    "isize"),
  col.names = c("readID", "samFlag", "seqnames", "strand", "leftStart", "width", "mapq",
    "mateSeqnames", "mateLeftStart", "isize")
)
}
\arguments{
\item{file}{Path to bam file}

\item{sel}{columns to be imported (see Rsamtools::scanBamWhat()). Default= c("qname", "flag", "rname", "strand", "pos", "qwidth", "mapq", "mrnm", "mpos", "isize")}

\item{col.names}{Names of the columns to be imported.}
}
\value{
Returns a bam file content as data.table
}
\description{
Title
}
\examples{
vl_importBam("path/to_bam_file")
}
