% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_importBed.data.table}
\alias{vl_importBed.data.table}
\title{Check if data.table file contains all bed canonical fields}
\usage{
\method{vl_importBed}{data.table}(bed)
}
\description{
Check if data.table file contains all bed canonical fields
}
