% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_importBed.default}
\alias{vl_importBed.default}
\title{Default uses data.table format and does sanity checks}
\usage{
\method{vl_importBed}{default}(bed)
}
\description{
Default uses data.table format and does sanity checks
}
