% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_importBed}
\alias{vl_importBed}
\alias{vl_importBed.character}
\alias{vl_importBed.GRanges}
\title{Import bed file}
\usage{
vl_importBed(bed, ...)

\method{vl_importBed}{character}(bed)

\method{vl_importBed}{GRanges}(bed)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a character vector corresponding to ranges (chr:start, chr:start:strand, chr:start-end or chr:start-end:strand), a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}
}
\value{
DT ranges
}
\description{
Imports bed as data.table and check formats
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_importBed(character)}: for bed paths or character ranges

\item \code{vl_importBed(GRanges)}: for GRanges

}}
\examples{
bed <- vl_importBed("/path/to/file.bed")
test <- data.table(seqnames= "chr3R",
                   start= 100000,
                   end= 110000)
bed <- vl_importBed(test)
vl_importBed("chrX:17470075")
vl_importBed("chrX:17470075:+")
vl_importBed("chrX:17470075-17870970")
vl_importBed("chrX:17470075-17870970:-")

}
