% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_importContrib}
\alias{vl_importContrib}
\title{Title}
\usage{
vl_importContrib(
  h5,
  bed = list.files(dirname(h5), ".bed$", full.names = TRUE),
  selection
)
}
\arguments{
\item{h5}{Path to an h5 file containing the contribution scores.}

\item{bed}{Bed file containing the regions for which contributions were computed. By default, the file is searched in the same folder as the h5 file.}

\item{selection}{An optional bed file containing a selected set of regions. If specified, only the regions in bed overlapping with the selection will be returned}
}
\value{
A contribution data.table containing, for each nucleotide, the corresponding base and its associated contribution score.
}
\description{
Title
}
