% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_intersectBed}
\alias{vl_intersectBed}
\title{Find closestBed regions}
\usage{
vl_intersectBed(a, b, ignore.strand = TRUE, min.overlap = 1L, invert = F)
}
\arguments{
\item{a}{Granges or data.table from which regions overlapping b have to be returned.}

\item{b}{Set of regions of interest.}

\item{ignore.strand}{Should strand be ignored? Default= TRUE.}

\item{min.overlap}{A vector integer of length 1 or nrow(a) specifying the minimum overlap (bp) require to count overlaps. Default= 1L.}

\item{invert}{If set to true, returns non-overlapping features. Default= FALSE.}
}
\value{
Return 'a' ranges that overlap with any range in 'b'
}
\description{
Return regions "a" overlapping with regions "b"
}
\examples{
a <- data.table(seqnames= "chr2L",
                start= c(1000, 3000, 5000),
                end= c(2000, 4000, 6000),
                strand= c("+", "+", "-"))
b <- data.table(seqnames= "chr2L",
                start= c(1000, 5000),
                end= c(2000, 6000),
                strand= c("+", "+"))
vl_intersectBed(a, b)

}
