% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_last_err}
\alias{vl_last_err}
\title{See most recent error file in a directory}
\usage{
vl_last_err(dir)
}
\arguments{
\item{dir}{URL where the reports is to be found}
}
\description{
See most recent error file in a directory
}
