% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_plot.R
\name{vl_MAplot}
\alias{vl_MAplot}
\title{Plots a nice MA plot}
\usage{
vl_MAplot(DESeq2_dat, main = NA, padj.cutoff = 0.05, log2FC.cutoff = log2(1.5))
}
\arguments{
\item{DESeq2_dat}{A data.table containing a Deseq2 output sheet with log2FoldChange and baseMean.}

\item{main}{Title}

\item{padj.cutoff}{padjust cutoff to be used to call up/down genes. Default= 0.05.}

\item{log2FC.cutoff}{log2FoldChange cutoff to be used to call up/down genes. Default= log2(1.5).}
}
\value{
MA plot
}
\description{
Plots a nice MA plot
}
\examples{
vl_MAplot(fread("/groups/stark/vloubiere/projects/epigenetic_cancer/db/FC_tables/RNA/epiCancer_GFP_PH29_vs_W29.txt"),
          main= "PH29")
}
