% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_metadata_CUTNRUN}
\alias{vl_metadata_CUTNRUN}
\title{Example metadata CUTNRUN}
\format{
data.table
#' \describe{
  \item{user}{User name}
  \item{method}{CutNRun, ChIP, ....}
  \item{experiment}{exp1, exp2, ...}
  \item{antibody}{V5, H3K27Ac, ...}
  \item{cell_line}{AID.Hcfc1, parental, ...}
  \item{treatment}{noTreatment, IAA, ...}
  \item{replicate}{rep1, rep2, ...}
  \item{condition}{Should be the concatenation of ab, cell_line, treatment and experiment  (separated by '_'). Therefore, is is similar to sampleID but without the replicate information, and will be used to call peaks on the merge data.}
  \item{sampleID}{sampleID should be the concatenation of ab, cell_line, treatment, experiment and replicate  (separated by '_'), and be unique for each sample/replicate. Samples with the same sampleID will be merged (re-sequencing...).}
  \item{barcode}{experimental barcode. e.g. ACAGTG, GCCAAT...}
  \item{i5}{i5 index, such as ACGTCCTG...}
  \item{genome}{For now, only 'mm10' and 'hg38' are supported}
  \item{input}{Sample ID of the input data to be used for callling peaks}
  \item{layout}{'PAIRED' or 'SINGLE'. Will be used in combination with 'sequencer' to know in which column barcodes should be found}
  \item{sequencer}{'NextSeq or HiSeq. Will be used in combination with 'layout' to know in which column barcodes should be found}
  \item{bam_path}{The path to the VBC bam file containing the reads}
  \item{comment}{Extra columns are tolerated}
  ...
}
}
\usage{
See ?vl_CUTNRUN_processing()
}
\description{
Metadata table example for the CUTNRUN pipeline. See ?vl_CUTNRUN_processing() for further details
}
\keyword{datasets}
