% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_metadata_ORFtag}
\alias{vl_metadata_ORFtag}
\title{Example metadata ORFtag}
\format{
data.table
#' \describe{
  \item{user}{User name}
  \item{batch}{batch}
  \item{screen}{screen name}
  \item{condition}{Typically one of 'sort' or 'input'}
  \item{replicate}{replicate}
  \item{barcodes}{Typically two barcodes, e.g. 'GCCTCTTC|ATTGATTC'}
  \item{sampleID}{sampleID. Should correspond to the catenation of screen, condtion and replicate  (separated by '_'), and be unique for each biological sample/replicate. Samples with the same sampleID will be merged (re-sequencing...).}
  \item{expName}{Experiment name}
  \item{genome}{For now, only 'mm10' and 'hg38' are supported}
  \item{layout}{'PAIRED' or 'SINGLE'. Will be used in combination with 'sequencer' to know in which column barcodes should be found}
  \item{sequencer}{'NextSeq or HiSeq. Will be used in combination with 'layout' to know in which column barcodes should be found}
  \item{bam_path}{The path to the VBC bam file containing the reads}
  \item{comment}{Extra columns are tolerated}
  ...
}
}
\usage{
See ?ORFtag_pipeline()
}
\description{
Metadata table example for the ORFtag pipeline. See ?ORFtag_pipeline() for further details
}
\keyword{datasets}
