% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_metadata_PROseq}
\alias{vl_metadata_PROseq}
\title{Example metadata PROseq}
\format{
data.table
#' \describe{
  \item{user}{User name}
  \item{experiment}{HCFC1, HCFC1_rescue, ...}
  \item{cell_line}{AID-Hcfc1-cl4, AID-Hcfc1-cl17, ...}
  \item{treatment}{0hrIAA, 3hrIAA, ...}
  \item{replicate}{rep1, rep2, ...}
  \item{sampleID}{sampleID should be the concatenation of cell_line, treament and replicate (separated by '_'), and be unique for each sample/replicate. Samples with the same sampleID will be merged (re-sequencing...).}
  \item{DESeq2_name}{Simplified sample name to be used within the DESeq2 object. Should be unique for each sample/replicate.}
  \item{DESeq2_condition}{Condition name to be used in DESeq2. Ideally, it will be similar to DESeq2_name but with no replicate information.}
  \item{DESeq2_control}{Should correspond to the DESeq2_condition of the control condition to be used for differential analysis.}
  \item{barcode}{Index barcode. e.g. ACAGTG, ...}
  \item{eBC}{Sample barcode found at the beginning of the read. e.g. ATCG, GCTA...}
  \item{genome}{For now, only 'mm10' and 'hg38' are supported}
  \item{spikein_genome}{For now, only 'dm3' is supported}
  \item{layout}{'PAIRED' or 'SINGLE'. Will be used in combination with 'sequencer' to know in which column barcodes should be found}
  \item{sequencer}{'NextSeq or HiSeq. Will be used in combination with 'layout' to know in which column barcodes should be found}
  \item{bam_path}{The path to the VBC bam file containing the reads}
  \item{comment}{Extra columns are tolerated}
  ...
}
}
\usage{
See ?vl_CUTNRUN_processing()
}
\description{
Metadata table example for the CUTNRUN pipeline. See ?vl_PROseq_processing() for further details
}
\keyword{datasets}
