% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionTools.R
\name{vl_model_expVar}
\alias{vl_model_expVar}
\title{Compute explained variance}
\usage{
vl_model_expVar(model)
}
\arguments{
\item{model}{An object containing the results returned by a model fitting function (e.g., lm or glm).}
}
\value{
% of explained variance
}
\description{
From Bernardo, computes the % explained variance for each predictor
}
