% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich(
  counts_matrix,
  cl_IDs,
  collapse_clusters = colnames(counts_matrix),
  control_cl = unique(cl_IDs),
  plot = T,
  padj_cutoff = 1e-05,
  top_enrich = Inf,
  x_breaks,
  color_breaks,
  cex.balloons = 1,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  add_motifs = F,
  cex.width = 1,
  cex.height = 1
)
}
\arguments{
\item{counts_matrix}{A matrix containing motif counts (rows= sequences, cols= motifs)}

\item{cl_IDs}{vector of cluster IDs (used to split the data.table)}

\item{collapse_clusters}{A vector of clusters of lenght ncol(counts_matrix) used to collapse motifs (only the most enriched motif/cluster will be returned). Default= colnames(counts_matrix), meaning no collapsing}

\item{control_cl}{IDs of clusters to be used as background. default to unique(cl_IDs), meaning all clusters are used except the one being tested}

\item{plot}{Should the result be plot using balloons plot?}

\item{padj_cutoff}{cutoff for ballons to be plotted}

\item{top_enrich}{Select top enriched motifs/cluster}

\item{x_breaks}{Breaks used for ballon's sizes}

\item{color_breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{add_motifs}{Show only n top enriched motifs}

\item{cex.width}{expansion factor for motif widths}

\item{cex.height}{expansion factor for motif heights}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
\examples{
# Sets of peaks
top_SUHW <- vl_resizeBed(vl_SUHW_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
top_STARR <- vl_resizeBed(vl_STARR_DSCP_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
rdm <- vl_random_regions_BSgenome("dm3", 2000, width= 500)

# Compute enrichment SUHW vs STARR-Seq
set <- rbind(top_SUHW,
             top_STARR, fill= T)
counts <- vl_motif_counts(set, "dm3")
par(mar= c(4,20,4,4), las= 1)
pl <- vl_motif_cl_enrich(counts, 
                         cl_IDs = c(rep(1, nrow(top_SUHW)),
                                    rep(2, nrow(top_STARR))),
                         add_motifs= T)
pl <- plot(enr, padj_cutoff= 1e-5)
vl_add_motifs(pl$DT)

Compute enrichment of SUHW & STARR-Seq over random regions
counts <- vl_motif_counts(rbind(top_SUHW,
                                top_STARR, 
                                rdm, fill= T), "dm3")
cl_IDs <- c(rep("suhw", nrow(top_SUHW)),
            rep("STARR", nrow(top_STARR)),
            rep("rdm", nrow(rdm)))
enr <- vl_motif_cl_enrich(counts, 
                          cl_IDs = cl_IDs,
                          control_cl = "rdm",
                          plot=F)
par(las= 1)
plot(enr, padj_cutoff= 1e-20, top_enrich= 20)
}
