% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_motif_cl_enrich_plot_only}
\alias{vl_motif_cl_enrich_plot_only}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich_plot_only(obj, padj_cutoff = 1e-05,
  log2OR_cutoff = 0, N_top = Inf, color_breaks,
  col = c("cornflowerblue", "lightgrey", "tomato"), main = NA,
  cex.balloons = 4, auto_margins = T)
}
\arguments{
\item{obj}{A fisher test matrix similar to ?vl_motif_cl_enrich() ouput.}

\item{padj_cutoff}{cutoff for ballons to be ploted}

\item{log2OR_cutoff}{cutoff for ballons to be ploted}

\item{N_top}{Select top enriched motifs/cluster}

\item{color_breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{cex.balloons}{scaling factor for balloons size}

\item{auto_margins}{Use auto margins? Default= T}
}
\value{
ballons plot with padj and log2OR
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
