% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R,
%   R/plot_methods_enrichment_analyses.R
\name{vl_motif_enrich}
\alias{vl_motif_enrich}
\alias{plot.vl_enr}
\title{Motif enrichment analysis}
\usage{
vl_motif_enrich(
  counts,
  control.counts,
  names = vl_Dmel_motifs_DB_full[colnames(counts), motif_cluster, on = "motif_ID"],
  plot = F,
  padj.cutoff = 0.05,
  top.enrich = Inf,
  min.counts = 3L,
  order = "padj",
  breaks = NULL,
  col = c("blue", "red"),
  xlab = "Odd Ratio (log2)",
  add.motifs = F,
  cex.width = 1,
  cex.height = 1
)

\method{plot}{vl_enr}(
  obj,
  padj.cutoff = 0.05,
  top.enrich = Inf,
  min.counts = 3L,
  order = "padj",
  xlab = "Odd Ratio (log2)",
  col = c("blue", "red"),
  breaks = NULL
)
}
\arguments{
\item{counts}{data.table containing counts for the regions of interest.}

\item{control.counts}{data.table containing counts for control regions. Should have the same number of columns as 'counts' table.}

\item{names}{Convenient names to be used for aggregating and plotting. By default, returns motif_cluster matching motif_ID in vl_Dmel_motifs_DB_full.}

\item{plot}{Plot result?}

\item{padj.cutoff}{padjust cutoff to be applied before plotting}

\item{top.enrich}{Top enrichments to plot (based on padj). Default is to show all enriched features}

\item{min.counts}{The minimum number of counts required to call a hit. Default= 3L}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{breaks}{Color breaks to be used. Defaults to range of filtered padj.}

\item{col}{Color scale}

\item{xlab}{Default to "Odd Ratio (log2)"}

\item{add.motifs}{Should motif pwms be plotted?}

\item{cex.width}{Expansion factor for motif widths.}

\item{cex.height}{Expansion factor for motif heights.}

\item{obj}{An object of class vl_enr}
}
\value{
DT of enrichment values which can be plot using ?plot.vl_GO_enr
}
\description{
Compute motif enrichment between a set of regions and control regions
}
\section{Functions}{
\itemize{
\item \code{plot(vl_enr)}: method to plot enrichment objects (containing variable, log2OR and padj)

}}
\examples{
# Resize example peaks
SUHW <- vl_resizeBed(vl_SUHW_top_peaks, genome = "dm3")
STARR <- vl_resizeBed(vl_STARR_DSCP_top_peaks, genome = "dm3")

# Generate same number of random regions
random <- vl_control_regions_BSgenome(bed= STARR, genome= "dm3")

# Count JAPSPAR motifs (see below to use custom list of PWMs)
suhw <- vl_motif_counts(SUHW, genome= "dm3", pwm_log_odds= vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_log_odds])
starr <- vl_motif_counts(top_STARR, genome= "dm3", pwm_log_odds= vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_log_odds])
ctl <- vl_motif_counts(random, genome= "dm3", pwm_log_odds= vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_log_odds])

# Starting from sequence instead of bed file
seq <- vl_getSequence(SUHW, genome= "dm3")
seq_suhw <- vl_motif_counts(seq, genome= "dm3", pwm_log_odds= vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_log_odds])
identical(suhw, seq_suhw)

# Motifs can also be counted using a custom PWMatrixList, for example for promoter motifs:
prom_db <- readRDS("/groups/stark/almeida/data/motifs/CP_motifs/CP_motifs_PWM.rds")
prom <- prom_db$Pwms_log_odds
for(i in seq(prom))
  prom[[i]]@profileMatrix <- vl_pwm_perc_to_log2(prom_db$Pwms_perc[[i]]@profileMatrix)

prom_motifs <- vl_motif_counts(STARR,
                               pwm_log_odds= prom,
                               genome= "dm3")

# Compute enrichment at SUHW peaks, using random controls as background
pl <- vl_motif_enrich(suhw,
                      ctl,
                      plot= F)
pl[order(padj)][1:3] # Top motif is su(Hw)
# Plot
plot(pl,
     top.enrich= 3)


# Compute enrichment at STARR peaks and plot at the same time
enr <- vl_motif_enrich(starr,
                       ctl,
                       plot= T,
                       order= "log2OR",
                       padj.cutoff= 1e-5)
# Positive enrichments identify typical S2 enhancer motifs
plot(enr[log2OR>.5])

}
