% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_motifs_DB_v2}
\alias{vl_motifs_DB_v2}
\title{Motifs DB noin-redundant version (BA)}
\format{
An object containing 6502 motifs and related metadata
\describe{
  \item{motif_name}{Motif ID from TF_clusters_PWMs.RData}
  \item{motif_cluster}{Motif cluster from Bernardo Almeida}
  \item{collection}{}
  \item{collection_version}{}
  \item{species}{PWM expressed as log odds}
  \item{pwms_log_odds}{list of PWM matrices}
  \item{pwms_perc}{list of PWM matrices}
}
}
\source{
{"/groups/stark/almeida/Papers/DeepSTARR/Code/TF_motif_database/TF_clusters_PWMs.RData"}
}
\usage{
This curated list of non-redundant motifs was assembled by Bernardo P. de almeida (https://github.com/bernardo-de-almeida/motif-clustering). It can be used with the ?vl_motif_counts() function to count motifs
}
\description{
Contains 6502 non-redundant motifs
}
\keyword{datasets}
