% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_mPCC}
\alias{vl_mPCC}
\title{Compute Matthew's PCC}
\usage{
vl_mPCC(predicted, label)
}
\arguments{
\item{predicted}{Predicted values from the model (ranging from 0 to 1).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}
}
\description{
Compute Matthew's PCC
}
