% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMolBioTools.R
\name{vl_oligo_Tm}
\alias{vl_oligo_Tm}
\title{vOligoCalculator}
\usage{
vl_oligo_Tm(seq)
}
\arguments{
\item{seq}{Character vector containing DNA sequence}
}
\value{
pritns result
}
\description{
Calculates Oligo Tm and GC %
}
\examples{
vl_oligo_Tm("GCCGATTCTCGGGCAGTTCCTC")
}
