% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFeome_pipeline.R
\name{vl_ORFeome_MAGeCK_auto}
\alias{vl_ORFeome_MAGeCK_auto}
\alias{vl_ORFeome_MAGeCK_auto.character}
\alias{vl_ORFeome_MAGeCK_auto.default}
\title{MAGeCK analysis ORFeome screen}
\usage{
vl_ORFeome_MAGeCK_auto(processed_metadata, ...)

\method{vl_ORFeome_MAGeCK_auto}{character}(processed_metadata, ...)

\method{vl_ORFeome_MAGeCK_auto}{default}(
  processed_metadata,
  FC_metadata_output = gsub("_processed.rds", "_FC_tables.rds"),
  screen_group_columns = c("screen", "cell_line"),
  condition_group_columns = c("condition", "sort"),
  output_folder = "db/FC_tables/ORFeome/",
  input.cutoff.FUN = function(x) sum(x) >= 0,
  sample.cutoff.FUN = function(x) sum(x) >= 3,
  row.cutoff.FUN = function(x) sum(x) >= 0,
  input.pseudocount = 0,
  sample.pseudocount = 0,
  logFC.cutoff = 1,
  FDR.cutoff = 0.05,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 2,
  mem = 8,
  overwrite = FALSE,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/ORFeome/MAGeCK",
  time = "01:00:00"
)
}
\description{
MAGeCK analysis ORFeome screen
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_ORFeome_MAGeCK_auto(character)}: for processed_metadata file path

\item \code{vl_ORFeome_MAGeCK_auto(default)}: default method

}}
