% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{vl_palette_blueWhiteRed}
\alias{vl_palette_blueWhiteRed}
\title{Title}
\usage{
vl_palette_blueWhiteRed(x)
}
\arguments{
\item{x}{number of colours ro return}
}
\value{
vector of colors of length x
}
\description{
Title
}
