% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_par}
\alias{vl_par}
\title{Sets up my favorite par parameters. Optimized for 3x3 inch square.}
\usage{
vl_par(
  mai = c(0.9, 0.9, 0.9, 0.9),
  las = 1,
  tcl = -0.1,
  mgp = c(1.5, 0.35, 0),
  cex = 1,
  cex.lab = 9/12,
  cex.axis = 7/12,
  bty = "n",
  lend = 2,
  ...
)
}
\arguments{
\item{mai}{See ?par()}

\item{tcl}{}

\item{mgp}{}

\item{cex}{}

\item{cex.lab}{}

\item{cex.axis}{}

\item{bty}{}

\item{...}{Extra arguments to be passed to ?par()}
}
\value{
Set up nice plotting parameters
}
\description{
Sets up my favorite par parameters. Optimized for 3x3 inch square.
}
\examples{
# example code
pdf("testbc1558fcb6ef.pdf", 3, 3)
plot(1, 1, main= "before")
vl_par()
plot(1, 1, main= "after")
dev.off()

file.show("testbc1558fcb6ef.pdf")
file.remove("testbc1558fcb6ef.pdf")

 path.R
}
