% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_peakCalling}
\alias{vl_peakCalling}
\title{Compute peak calling}
\usage{
vl_peakCalling(
  ChIP_bed,
  Input_bed,
  average_function = function(x) ceiling(mean(x)),
  gaussian_blur = F,
  BSgenome,
  bins_width = 100,
  steps_width = round(bins_width/2),
  bins_pval_cutoff = 0.05,
  bins_OR_cutoff = 2
)
}
\arguments{
\item{ChIP_bed}{ChIP bed file. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{Input_bed}{Input bed file.}

\item{average_function}{Function applied to ChIP read counts to define candidate peaks}

\item{gaussian_blur}{Applies gaussian blur to ChIP signal to identify peak candidates. Useful for noisy signal. default= FALSE}

\item{BSgenome}{A BSgenome object used for gw binning}

\item{bins_width}{bins width used for peak calling, Default to 100L (narrow Peaks). Use larger bins to call domains}

\item{steps_width}{Step between bins. default to round(bins_width/2)}

\item{bins_pval_cutoff}{The cutoff that will be applied to find candidate bins and merge them into peaks. Less stringent cutoffs means broader regions. default to 0.05}

\item{bins_OR_cutoff}{Enrichment cutoff that will be applied to find candidate bins and merge them into peaks. Less stringent cutoffs means broader regions. default to 2}
}
\value{
significantly enriched peaks
}
\description{
Compute peak calling using ChIP and Input
}
