% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_coeff}
\alias{vl_plot_coeff}
\title{Plots Rsq or PCC coeff}
\usage{
vl_plot_coeff(
  x = "topleft",
  value,
  type = "rsq",
  adjusted = F,
  bty = "n",
  digits = 2,
  ...
)
}
\arguments{
\item{x}{}

\item{value}{Rsquare or r value. If numeric, will be rounded, while characters will be printed as is}

\item{type}{Type of value. one between "rsq" and "pcc"}

\item{adjusted}{Is the Rsq adjusted?}

\item{bty}{Box around the legend? default= F}

\item{digits}{Rounding digits. Default= 2}

\item{...}{}
}
\value{
Add Rsquare value as a legend on aplot
}
\description{
Plots Rsq or PCC coeff
}
\examples{
plot(1,1)
vl_plot_R2(value= 0.01)
vl_plot_R2("topright", value= 0.01, adjusted= T)
}
