% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_plot_contrib_logo}
\alias{vl_plot_contrib_logo}
\title{Plot contribution scores matrix}
\usage{
vl_plot_contrib_logo(
  contrib.object,
  enr,
  seqlvl,
  min.count = 3L,
  best.by = "motif",
  sel,
  xlim,
  ylim,
  xlab = "nt",
  ylab = "Contribution"
)
}
\arguments{
\item{contrib.object}{A contribution object, as outputed by ?vl_importContrib()}

\item{enr}{And enrichment object, as outputed by ?vl_contrib_enrich() on the sequences of the contrib.object.}

\item{seqlvl}{A seqlvl present in in enr$sig.mot.coor$seqlvls.}

\item{min.count}{The minimum number of significant instances across (see enr$sig.inst). Default= 3L}

\item{best.by}{The group.by column for which only the instance with the best padjust should be returned.}

\item{sel}{If specified, only the motifs for which best.by %in% sel will be plotted.}

\item{xlim}{Default= sequence length.}

\item{ylim}{Default= range(contrib).}

\item{xlab}{Default= "nt"}

\item{ylab}{Default= "Contribution"}
}
\description{
Plot contribution scores matrix
}
