% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enhancer_motifs.R
\name{vl_plot_enh_motifs}
\alias{vl_plot_enh_motifs}
\alias{vl_plot_enh_motifs.data.table}
\alias{vl_plot_enh_motifs.character}
\title{Plot enhancer with its motifs}
\usage{
vl_plot_enh_motifs(sequences, ...)

\method{vl_plot_enh_motifs}{data.table}(bed, genome, ...)

\method{vl_plot_enh_motifs}{character}(
  sequences,
  p.cutoff = 5e-04,
  sel,
  xleft,
  ybottom,
  width,
  height,
  col = rainbow(length(sel)),
  col_alpha = 0.3
)
}
\arguments{
\item{sequences}{Named character vector of sequences to analyse. If provided takes over bed argument (in the case where both are specified)}

\item{bed}{Either a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{genome}{Genome to be used for coordinates ("dm6, "dm3")}

\item{p.cutoff}{Pval cutoff used for motif detection}

\item{sel}{List of motifs to compute. see vl_Dmel_motifs_DB_full$motif}

\item{xleft}{Left position for plotting}

\item{ybottom}{Bottom position for plotting}

\item{width}{Plot width}

\item{height}{Plot height}

\item{col}{Color (1 per motif)}
}
\description{
Plot enhancer with its motifs
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_plot_enh_motifs(data.table)}: Plot enhancer with its motifs

\item \code{vl_plot_enh_motifs(character)}: Plot enhancer with its motifs

}}
\examples{
plot.new()
vl_plot_enh_motifs(bed= vl_SUHW_top_peaks[1:2], 
genome= "dm3", 
sel= c("cisbp__M2328", "flyfactorsurvey__suHw_FlyReg_FBgn0003567", "jaspar__MA0533.1"), 
xleft=0, 
ybottom=c(0.25,0.5), 
width=1, 
height= 0.1)
}
