% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vFormatPval.R
\name{vl_plot_pval}
\alias{vl_plot_pval}
\title{Plot pval}
\usage{
vl_plot_pval(x, y, pval, stars_only = F, ...)
}
\arguments{
\item{x}{X position for plotting}

\item{y}{Y position for plotting}

\item{pval}{Pval to be plotted}

\item{stars_only}{If set to TRU, then only plots */N.S. Default= FALSE}

\item{...}{extra arguments for text() function (used for plotting)}
}
\value{
An object that can be used with the vl_average_bw_track_plot_only() function.
}
\description{
Plots formated pval data as text
}
\examples{
pval <- c(1e-10, 1e-5, 1e-3, 1e-2, 1e-1, 1)
plot(NA, xlim= c(0,7), ylim= c(-1,1))
vl_plot_pval(1:6, 0.5, pval)
vl_plot_pval(1:6, 0, pval, cex= 0.6)
}
