% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_R2}
\alias{vl_plot_R2}
\title{Title}
\usage{
vl_plot_R2(x = "topleft", rsquare, adjusted = F, bty = "n", digits = 2, ...)
}
\arguments{
\item{x}{}

\item{rsquare}{}

\item{adjusted}{Is the Rsq adjusted?}

\item{bty}{Box around the legend? default= F}

\item{digits}{Rounding digits. Default= 2}

\item{...}{}
}
\value{
Add Rsquare value as a legend on aplot
}
\description{
Title
}
\examples{
plot(1,1)
vl_plot_R2(rsquare= 0.01)
vl_plot_R2("topright", rsquare= 0.01, adjusted= T)
}
