% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_table}
\alias{vl_plot_table}
\title{Table plot}
\usage{
vl_plot_table(x, colnames_col = "white", row_col = "white")
}
\arguments{
\item{x}{data.table}

\item{colnames_col}{Color for col.names row}

\item{row_col}{Color for data rows}
}
\value{
Plots heatkey
}
\description{
Plots a data.table as text
}
