% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_plotLetter}
\alias{vl_plotLetter}
\title{plot seqlogo letter}
\usage{
vl_plotLetter(letter, xleft, ytop, width, height)
}
\arguments{
\item{letter}{"A", "T", "C" or "G"}

\item{xleft}{xleft position}

\item{ytop}{ytop position}
}
\description{
See function vl_seqlogo
}
