% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROseq_pipeline.R
\name{vl_PROseq_DESeq2}
\alias{vl_PROseq_DESeq2}
\alias{vl_PROseq_DESeq2.character}
\alias{vl_PROseq_DESeq2.default}
\title{Title}
\usage{
vl_PROseq_DESeq2(processed_metadata, ...)

\method{vl_PROseq_DESeq2}{character}(processed_metadata, ...)

\method{vl_PROseq_DESeq2}{default}(
  processed_metadata,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 4,
  mem = 16,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/PROseq/DESeq2",
  time = "1-00:00:00"
)
}
\arguments{
\item{processed_metadata}{}

\item{Rpath}{}

\item{cores}{}

\item{mem}{}

\item{submit}{}

\item{wdir}{}

\item{logs}{}

\item{time}{}
}
\description{
Title
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_PROseq_DESeq2(character)}: for processed_metadata file path

\item \code{vl_PROseq_DESeq2(default)}: default method

}}
