% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterScatterplot.R
\name{vl_rasterScatterplot}
\alias{vl_rasterScatterplot}
\title{Title}
\usage{
vl_rasterScatterplot(
  x,
  y = NULL,
  type = "p",
  frame = F,
  res = 600L,
  size = 2000L,
  xlab = ifelse(is.null(y), "Index", deparse1(substitute(x))),
  ylab = ifelse(is.null(y), deparse1(substitute(x)), deparse1(substitute(y))),
  add = F,
  ...
)
}
\arguments{
\item{x}{}

\item{y}{default= NULL}

\item{type}{default= "p"}

\item{frame}{Default= F}

\item{res}{Nominal resolution of png image. default= 600}

\item{size}{Size of the png image. Default= 2000}

\item{...}{Extra plotting parameters}
}
\value{
Raster scatterplot
}
\description{
Title
}
\examples{
vl_rasterScatterplot(1:3)
}
