% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repelPlot.R
\name{vl_repelLabels}
\alias{vl_repelLabels}
\title{Title}
\usage{
vl_repelLabels(
  x,
  y,
  labels,
  label.cex = 0.7,
  label.col = "black",
  xlim = range(x),
  ylim = range(y),
  rect.draw = T,
  rect.col = adjustcolor("lightgrey", 0.5),
  point_size = 0.01,
  box_padding_x = 0,
  box_padding_y = 0,
  point_padding_x = 0,
  point_padding_y = 0,
  ...
)
}
\arguments{
\item{x}{see ?plot()}

\item{y}{see ?plot()}

\item{labels}{labels to be plotted}

\item{label.cex}{Expansion factor for labels}

\item{label.col}{Labels color}

\item{xlim}{Scatterplot xlim. default to range(x)}

\item{ylim}{Scatterplot xlim. default to range(y)}

\item{rect.draw}{Draw rectangle?}

\item{rect.col}{Rectangles color}

\item{box_padding_x}{}

\item{box_padding_y}{}

\item{point_padding_x}{}

\item{point_padding_y}{}

\item{...}{Extra arguments to be passed to plot()}

\item{points_size}{}
}
\value{
Plots a ggrepel-like scatterplot. Returns plotting object as data.table
}
\description{
Title
}
\examples{
data("mtcars")
mtcars$car <- rownames(mtcars)
vl_repelLabels(x = mtcars$wt,
y = mtcars$mpg,
labels = mtcars$car,
cex = 0.7)
}
