% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repelScatterplot.R
\name{vl_repelScatterplot}
\alias{vl_repelScatterplot}
\title{Title}
\usage{
vl_repelScatterplot(
  x,
  y = NULL,
  col = "lightgrey",
  labels,
  label.sel = rep(T, length(x)),
  label.cex = 0.7,
  label.col = "black",
  xlab = ifelse(is.null(y), "Index", deparse1(substitute(x))),
  ylab = ifelse(is.null(y), deparse1(substitute(x)), deparse1(substitute(y))),
  xlim = if (add) par("usr")[1:2] else range(x),
  ylim = if (add) par("usr")[3:4] else range(y),
  rect.draw = T,
  rect.col = col,
  seg.col = col,
  noOverlap = F,
  point_size = 0,
  box_padding_x = 0.5,
  box_padding_y = 0.5,
  points_padding_x = 0,
  points_padding_y = 0,
  force_push = 1e-05,
  force_pull = 1e-05,
  add = F,
  frame = F,
  ...
)
}
\arguments{
\item{x}{see ?plot()}

\item{y}{see ?plot()}

\item{labels}{labels to be plotted}

\item{label.sel}{Numeric or boolean vector indicating which labels should be plotted}

\item{label.cex}{Expansion factor for labels}

\item{label.col}{Labels color}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{xlim}{Scatterplot xlim. default to range(x)}

\item{ylim}{Scatterplot xlim. default to range(y)}

\item{rect.draw}{Draw rectangle?}

\item{rect.col}{Segments colors}

\item{point_size}{(numeric) approximate size of points, defined as a fraction of char size. Can be small, is used in conjunction with "point_padding". default= 0.001}

\item{box_padding_x}{(numeric) extra margin around text bounding boxes, defined as a fraction of char size. Default= 0.005}

\item{box_padding_y}{(numeric) extra margin around text bounding boxes, defined as a fraction of char size. Default= 0.005}

\item{force_push}{(numeric) magnitude of the push force (defaults to 1e-5)}

\item{force_pull}{(numeric) magnitude of the pull force (defaults to 1e-5)}

\item{add}{Should the label be added to existing plot? Default= F}

\item{frame}{Should the frame be plotted? Deault= F}

\item{...}{Extra arguments to be passed to plot()}

\item{point_padding_x}{(numeric) extra margin around data points, defined as a fraction of char size. Default= 0.01}

\item{point_padding_y}{(numeric) extra margin around data points, defined as a fraction of char size. Default= 0.01}
}
\value{
Plots a ggrepel-like scatterplot. Returns plotting object as data.table
}
\description{
Title
}
\examples{
data("mtcars")
mtcars$car <- rownames(mtcars)
vl_repelScatterplot(x = mtcars$wt,
y = mtcars$mpg,
labels = mtcars$car,
cex = 0.7)
}
