% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vReverseComplement.R
\name{vl_revComp}
\alias{vl_revComp}
\title{reverseComplement}
\usage{
vl_revComp(DNA_char, complement = T, reverse = T)
}
\arguments{
\item{DNA_char}{DNA character string}
}
\value{
Reverse Complemented DNA sequence
}
\description{
Reverse complement DNA sequence character string
}
\examples{
vl_revComp("ATCG")
}
