% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_pipeline.R
\name{vl_RNAseq_processing}
\alias{vl_RNAseq_processing}
\title{RNA-seq pipeline}
\usage{
vl_RNAseq_processing(
  metadata,
  processed_metadata_output,
  fq_output_folder = "/scratch/stark/vloubiere/fq/RNAseq/",
  bam_output_folder = "/scratch/stark/vloubiere/bam/RNAseq/",
  alignment_stats_output_folder = "db/alignment_stats/RNAseq/",
  count_table_output_folder = "db/count_tables/RNAseq/",
  bw_output_folder = "db/bw/RNAseq/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 8,
  mem = 64,
  overwrite = FALSE,
  counts.exists = TRUE,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/RNAseq/processing",
  time = "2-00:00:00"
)
}
\arguments{
\item{metadata}{The path to a correctly formatted .xlsx, .txt or .rds file, or a data.table. See the template at '/groups/stark/vloubiere/projects/vl_pipelines/Rdata/metadata_RNAseq.xlsx'.}

\item{processed_metadata_output}{An .rds path where to save the processed metadata file, which contains the paths of all output files and will be used to manage them.
By default, when importing the metadata from an excel sheet, "_processed.rds" will be appended to the excel file path.}

\item{fq_output_folder}{Output folder for .fq files. Default= "/scratch/stark/vloubiere/fq/RNAseq/".}

\item{bam_output_folder}{Output folder for .bam files. Default= "/scratch/stark/vloubiere/bam/RNAseq/".}

\item{alignment_stats_output_folder}{Output folder for alignment statistics. Default= "db/alignment_stats/RNAseq/".}

\item{count_table_output_folder}{Output folder for count table files. Default= "db/count_tables/RNAseq/".}

\item{bw_output_folder}{Output folder for .bw tracks. Default= "db/bw/RNAseq/".}

\item{Rpath}{Path to an Rscript executable. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}

\item{cores}{Number of cores per job. Default= 8.}

\item{mem}{Memory per job (in Go). Default= 64.}

\item{overwrite}{Should existing files be overwritten?}

\item{counts.exists}{If the output count_table file already exists and overwrite is set to FALSE, skips demultiplexing and alignment. Default= TRUE.}

\item{submit}{Should the command be submitted? Default= FALSE.}

\item{wdir}{The working directory to use. Default= getwd(), meaning current working directory will be used.}

\item{logs}{Output folder for log files. Default= "db/logs/CUTNRUN/processing".}

\item{time}{The time required for the SLURM scheduler. Default= '2-00:00:00'.}
}
\value{
Return a data.table containing, for each sampleID, the concatenated commands that are required to process the data.
These commands can then be submitted either directly via the function, or using vl_bsub()...
}
\description{
To use this this pipeline, please install the vl_function package using install_github("vloubiere/vlfunction")
in your interactive Rstudio session AND in the local installation of R 4.3.0 ('/software/f2022/software/r/4.3.0-foss-2022b/bin/R') 
which will be used to submit R sub-scripts.
}
\details{
The pipeline is split into two main functions. The first function vl_RNAseq_processing() demultiplexes tar files, aligns the reads and filters confident alignments.
It takes as input a (correctly formatted) metadata file, saves the processed_metadata file and returns and/or submit the command lines to: \cr
1/ extract reads from VBC .tar.gz file. Output .fq.gz files are saved in fq_output_folder/ \cr
2/ trim the reads. Output .fq.gz files are saved in bam_output_folder/ \cr
3/ align the reads to the provided genome (see 'genome' column of the metadata table, only mm10 is supported for now). Output .bam files are saved in tmp_folder/bam/ and alignment statistics in the alignment_stats_output_folder/ \cr
4/ compute read counts per exon, and returns the total number of reads per gene_id. Counts statistics will be save in alignment_stats_output_folder/ (together with alignment stats) and count_tables in count_table_output_folder/  \cr
5/ output .bw tracks that are stored in bw_output_folder/  \cr

The second function, vl_RNAseq_DESeq2() can be used for differential analysis. It returns: \cr
1/ a .dds DESeq2 object for each experiment. Output .dds (rds) files are stored in the dds_output_folder/ \cr
2/ the fold-change tables corresponding to the comparison of 'DESeq2_condition' versus 'DESeq2_control' conditions (see colnames of the processed_metadata file). Output .txt files are stored in the FC_output_folder/ \cr
3/ A barplot showing the statistics of each dds file, which will be saved in the PDF_output_folder. \cr
4/ MA plots corresponding to each comparison, which will be saved in the PDF_output_folder. \cr
}
\examples{
library(vlfunctions)
# Example metadata ----
metadata <- system.file("RNAseq_pipeline", "metadata_RNAseq.xlsx", package = "vlfunctions")

# Generate command lines ----
cmd <- vl_RNAseq_processing(metadata= metadata,
                            overwrite= TRUE,
                            submit= FALSE)
# To actually submit the jobs, switch to submit= TRUE
vl_RNAseq_processing(metadata= metadata,
                     processed_metadata_output = "Rdata/metadata_RNAseq_processed.rds",
                     overwrite= FALSE,  # Existing files will not be processed again
                     submit= TRUE)

# Differential analysis
See ?vl_RNAseq_DESeq2()

}
