% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_ROC_AUC}
\alias{vl_ROC_AUC}
\title{ROC AUC}
\usage{
vl_ROC_AUC(
  predicted,
  label,
  plot = FALSE,
  xlab = "False Positive Rate",
  ylab = "True Positive Rate",
  type = "l",
  add = FALSE,
  ...
)
}
\arguments{
\item{predicted}{Predicted values from the model (ranging from 0 to 1).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}

\item{plot}{Should the ROC be plotted? Default= FALSE.}

\item{xlab}{xlab. Default= "FALSE positive rate".}

\item{ylab}{ylab. Default= "TRUE positive rate".}

\item{type}{The type of plot, when add= FALSE. Default= "l".}

\item{add}{When plot is set to TRUE, should only the line be added to an existing plot?}

\item{...}{Extra arguments to be passed to plot (when add= FALSE) or lines (when add= TRUE).}
}
\value{
ROC AUC
}
\description{
ROC AUC
}
