% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshot.R
\name{vl_screenshot}
\alias{vl_screenshot}
\title{bw screenshot}
\usage{
vl_screenshot(
  bed,
  tracks,
  genome,
  highlight_regions = NULL,
  highlight_col = "lightgrey",
  names = NULL,
  max = NULL,
  strand = NULL,
  col = NULL,
  n_genes = 4,
  gband = 0.1
)
}
\arguments{
\item{bed}{Regions to plot. Can be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start' and 'end' columns}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{genome}{"dm3", "dm6" or "mm10"}

\item{highlight_regions}{Path to a bed file OR Granges OR data.table object specifying regions to highlight}

\item{highlight_col}{Color to use for highlighting regions. default is "lightgrey"}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}

\item{max}{Max values to clip the data. If specified, must be the same length as bw vector. By default, uses max values/track.}

\item{col}{Ploting colors. If specified, must be the same length as bw vector. By default, uses "black"}

\item{n_genes}{Max number of genes to plot.}

\item{gband}{gband ratio. default is 0.1}
}
\description{
This function plots a screenshot of bw files for a set of regions.
}
\examples{
vl_screenshot(bed = GRanges("chr3R", IRanges(1+c(0, 5e6), 1e6+c(0, 5e6))),
genome= "dm3",
tracks = c("/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE81795_H3K4me3_rep1_uniq.bw", 
"/groups/stark/vloubiere/projects/pe_STARRSeq/db/peaks/H3K4me3_peaks.txt",
"/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE81795_H3K4me3_rep1_uniq.bw"),
strand = c(1,-1,-1),
highlight_regions = "/groups/stark/vloubiere/projects/pe_STARRSeq/db/peaks/H3K4me3_peaks.txt", 
n_genes = 4)

}
