% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshotSimple.R
\name{vl_screenshot_simple}
\alias{vl_screenshot_simple}
\title{Simplified version. less options, faster}
\usage{
vl_screenshot_simple(
  bed,
  tracks,
  highlight_bed = NULL,
  col = "black",
  highlight_col = "lightgrey",
  space = 30,
  widths = c(100L, 20L),
  names = NULL,
  max = NULL,
  genome
)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{bw of bed (peaks)}

\item{space}{space between regions. Default is 30 px}

\item{widths}{Length 2 integer vector specifying the width of bw and bed tracks, respectively. Default= c(100L, 20L)}

\item{names}{names for bw/bed files}

\item{gtf}{Optional gtf file(plot genes)}
}
\value{

}
\description{
Simplified version. less options, faster
}
