% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_actMatched_controls.R
\name{vl_select_act_matched}
\alias{vl_select_act_matched}
\title{Title}
\usage{
vl_select_act_matched(group, control, act.col.name)
}
\arguments{
\item{group}{Data table containing the groups for which controls have to be selected}

\item{control}{Data table containing the potential controls}

\item{act.col.name}{Name of the column containing activity values (has to be the same in the two tables)}
}
\value{
Returns a subset of the control data.table corresponding with the act.col.name values compared to group
}
\description{
Title
}
\examples{
No exmaple yet
}
