% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vseqlogo.R
\name{vl_seqlogo}
\alias{vl_seqlogo}
\title{plot seqlogo rasterImage}
\usage{
vl_seqlogo(pwm, xleft = 0, ybottom = 0, xright = 1, ytop = 1,
  add = T)
}
\arguments{
\item{pwm}{pwm matrix}

\item{xleft}{left plot limit. Default= 0}

\item{ybottom}{bottom plot limit. Default= 0}

\item{xright}{right plot limit. Default= 1}

\item{ytop}{top plot limit. Default= 1}

\item{add}{Should the pwm be plot on the top of opened device? Default= T}
}
\description{
plot seqlogo from pwm matrix
}
