% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_seqMotifs}
\alias{vl_seqMotifs}
\alias{vl_seqMotifs.data.table}
\alias{vl_seqMotifs.character}
\alias{plot.vl_seqMotifs}
\title{Find motif positions}
\usage{
vl_seqMotifs(sequences, ...)

\method{vl_seqMotifs}{data.table}(bed, genome, ...)

\method{vl_seqMotifs}{character}(
  sequences,
  p.cutoff = 5e-04,
  sel = vl_Dmel_motifs_DB_full[!is.na(vl_Dmel_motifs_DB_full$FBgn), motif],
  xleft,
  ybottom,
  width,
  height,
  col = rainbow(length(sel)),
  col_alpha = 0.3,
  plot = T
)

\method{plot}{vl_seqMotifs}(obj, xleft, ybottom, width, height, col)
}
\arguments{
\item{sequences}{Named character vector of sequences to analyse. If provided takes over bed argument (in the case where both are specified)}

\item{bed}{Either a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{genome}{Genome to be used for coordinates ("dm6, "dm3")}

\item{p.cutoff}{Pval cutoff used for motif detection}

\item{sel}{List of motifs to compute. see vl_Dmel_motifs_DB_full$motif}

\item{xleft}{Left position for plotting}

\item{ybottom}{Bottom position for plotting}

\item{width}{Plot width}

\item{height}{Plot height}

\item{col}{Color (1 per motif)}
}
\value{
An object of class "vl_seqMotifs" that can be used for plotting -> plot(obj)
}
\description{
Compute motif positions
}
\section{Methods (by class)}{
\itemize{
\item \code{data.table}: Compute motifs position within

\item \code{character}: Identify motifs in sequences

\item \code{vl_seqMotifs}: Identify motifs in sequences
}}

\examples{
plot.new()
vl_seqMotifs(data.table(seqnames="chr3R", start= 16948622, end= 16948870), 
genome= "dm6", 
sel= c("pho", "Trl/1"), 
xleft=0, 
ybottom=c(0.25,0.5), 
width=1, 
height= 0.1, 
col= c("goldenrod1", "tomato"))
}
