% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_shuffleBed}
\alias{vl_shuffleBed}
\title{Shuffle bed file}
\usage{
vl_shuffleBed(bed)
}
\arguments{
\item{bed}{bed file to shuffle. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}
}
\value{
original bins file with OR, pval and padj corresponding to fisher result
}
\description{
Shuffle bed file in respect of covered regions, total reads and reads length
}
\examples{
ChIP <- "/test/H3K27me3_PH18_merge_uniq.bed"
res <- shuffleBed(ChIP)
}
