% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vSRRgetURL.R
\name{vl_SRR_url}
\alias{vl_SRR_url}
\title{Extract SRR download fastq}
\usage{
vl_SRR_url(SRR)
}
\arguments{
\item{SRR}{SRR ID}
}
\value{
A character vector containing the SRA download link
}
\description{
Tries to extract the download link of the SRA file for given SRR
}
\examples{
vl_SRR_url("SRR13325541") 
}
