% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_STARR_DSCP_top_peaks}
\alias{vl_STARR_DSCP_top_peaks}
\title{STARR-Seq top peaks}
\format{Narrowpeaks file containing top STARR-Seq peaks (DSCP) on dm3}
\usage{
see ?vl_motif_enrich()
}
\description{
Example set that can be used for motif enrichment...
}
\keyword{datasets}
