% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRINGtools.R
\name{vl_STRING_interaction}
\alias{vl_STRING_interaction}
\alias{plot.vl_STRING}
\title{get STRING intrations DT}
\usage{
vl_STRING_interaction(
  symbols,
  species,
  plot = T,
  score_cutoff = 400,
  top_N = Inf,
  col = "tomato",
  size = 10,
  cex.label = 1
)

\method{plot}{vl_STRING}(obj, score_cutoff = 400, top_N = Inf)
}
\arguments{
\item{symbols}{vector of Dmel gene symbols}

\item{species}{either "Dm" or "Mm"}

\item{plot}{Should the graph be plotted?}

\item{score_cutoff}{If specified, only plot interactions with score>=cutoff}

\item{top_N}{If specified, only plot top N interactions}

\item{col}{For each gene symbol, the color of corresponding vertice (i.e, tomato for up, cornflowerblue for down)}

\item{size}{For each gene symbol, the size of corresponding vertice (i.e, absolute FC)}

\item{cex.label}{cex factor to be applied to correponding vertices labels}
}
\value{
An igraph network
}
\description{
Extract interactions form STRIN db (Dmel only for now)
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Method to plot STRING interaction igraphs
}}

\examples{
#USAGE
symbols <- c("Pc", "Psc", "E(z)", "RpL10", "RpL11", "RpL12")
net <- vl_STRING_interaction(symbols, species= "Dm", col= vl_palette_few_categ(6))
E <- igraph::as_data_frame(net, what= "edges") 
V <- igraph::as_data_frame(net, what= "vertices") 
V$color <- grey.colors(6)
.g <- igraph::graph_from_data_frame(d = E, 
                                    vertices = V,
                                    directed = F)
plot(.g)
}
