% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_SUHW_top_peaks}
\alias{vl_SUHW_top_peaks}
\title{SUHW top peaks}
\format{
Narrowpeaks file containing top SUHW peaks on dm3 chanonical chromosomes
}
\usage{
see ?vl_motif_enrich()
}
\description{
Example set that can be used for motif enrichment...
}
\keyword{datasets}
