% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_thermofisher_restriction_enzymes_table}
\alias{vl_thermofisher_restriction_enzymes_table}
\title{Thermofisher enzymes}
\format{
A table containing thermofisher enzymes and related info
\describe{
  \item{name1}{enzyme name}
  \item{name2}{Alternative names}
  \item{fastdigest}{FastDigest?}
  \item{buffer}{Buffer to be used}
  \item{temperature}{Temeprature for digest}
  \item{cutsite}{cutsite pattern}
  \item{consensus_F}{consensus motif}
  ...
}
}
\source{
{"https://www.thermofisher.com/at/en/home/brands/thermo-scientific/molecular-biology/thermo-scientific-restriction-modifying-enzymes/restriction-enzymes-thermo-scientific/conventional-restriction-enzymes-thermo-scientific.html"}
}
\usage{
Can be used with vl_digest. see ?vl_digest()
}
\description{
Contains cutting sites for most commercial enzymes
}
\keyword{datasets}
