% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_tilt_xaxis}
\alias{vl_tilt_xaxis}
\title{Tilted x axis}
\usage{
vl_tilt_xaxis(
  x,
  y = rep(par("usr")[3], length(labels)) - diff(grconvertY(c(0, par("mgp")[2]), "line",
    "user")),
  labels,
  srt = 45,
  offset = 0.25,
  pos = 2,
  xpd = NA,
  cex = par("cex.axis"),
  ...
)
}
\arguments{
\item{x}{x position}

\item{y}{Y position. Defaults to y axis position}

\item{labels}{labels to be plotted}

\item{srt}{Rotation. Default to 45}

\item{offset}{offset from the x pos. default= 0.25}

\item{pos}{Position. defaults to 2}

\item{xpd}{Default to NA}

\item{cex}{Default to par("cex.axis")}

\item{...}{Extra arguments to be passed to text}
}
\description{
Tilted x axis
}
