% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_toDTranges}
\alias{vl_toDTranges}
\title{Convert character to bed file}
\usage{
vl_toDTranges(coor)
}
\arguments{
\item{coor}{Character vector of genomic coordinates, e.g "chr3R:2281049-2281297:*"}
}
\value{
DT ranges
}
\description{
Convert character to bed file
}
\examples{
vl_toDTranges("chr3R:2281049-2281297:*")
}
