% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoMotifs.R
\name{vl_trainLASSO}
\alias{vl_trainLASSO}
\title{LASSO regression using TF motifs}
\usage{
vl_trainLASSO(response, counts)
}
\arguments{
\item{response}{Variable to predict}

\item{counts}{Matrix of motif counts}
}
\value{
Return a LASSO model
}
\description{
LASSO regression using TF motifs
}
\examples{
test <- vl_trainLASSO(ATACSeq_FC, counts)

}
