% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vUpsetSplit.R
\name{vl_upset_split}
\alias{vl_upset_split}
\title{upset Split}
\usage{
vl_upset_split(dat_list)
}
\arguments{
\item{dat_list}{List containing the names to intersect (see examples)}
}
\description{
Split a list of objects in upset-plot lie classes
}
\examples{
test <- list(A= 1:1000, B= 1:1500, C= 1000:1750)
vl_upset_split(test)
}
