% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{vl_venn}
\alias{vl_venn}
\title{Title}
\usage{
vl_venn(
  dat,
  col = adjustcolor(c("cornflowerblue", "tomato", "gold"), 0.3),
  border = NA
)
}
\arguments{
\item{dat}{A named list of variable of length 2 or 3}

\item{col}{Circles colors. Default to c("cornflowerblue", "tomato", "gold") and alpha= 0.3}

\item{border}{circle borders}
}
\value{
Plots a venn diagram for 2/3 groups
}
\description{
Title
}
\examples{
par(mfrow= c(1,2))
dat <- list(A= c("A", "B", "C", "D", "E", "F", "G", "H"),
B= c("I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "A"),
C= c("A", "B", "C", "K", "L", "M", "U", "V", "W", "X", "Y", "Z"))
vl_venn(dat)
dat <- list(A= c("A", "B", "C", "D", "E", "F", "G", "H"),
B= c("A", "B", "C", "D", "K", "L", "M", "U", "V", "W", "X", "Y", "Z"))
vl_venn(dat)
}
