% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{plot.bw_heatmap}
\alias{plot.bw_heatmap}
\title{bw heatmap plot only}
\usage{
plot.bw_heatmap(obj, col = c("blue", "yellow", "white"),
  center_label = "TSS", max_FUN = function(x) quantile(x, 0.995, na.rm
  = T), order_col_idx = 1, order_FUN = function(x) max(x, na.rm = T))
}
\arguments{
\item{obj}{An object returned by the vl_heatmap_bw_track() function. The object can easily be modified to change colors and so on...}

\item{col}{Color vector to be used for heatmaps. col= c("blue", "yellow", "white")}

\item{center_label}{Label for the center of the heatmaps}

\item{max_FUN}{Function used to compute clipping max for each track. default= function(x) quantile(x, 0.995, na.rm =T)}
}
\description{
Plotting method for bw_heatmap retuned by vl_heatmap_bw_track output object
}
