% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_add_motifs}
\alias{vl_add_motifs}
\title{Add motifs to enrichment plots}
\usage{
vl_add_motifs(DT, cex.width = 1, cex.height = 1, lwd = 0.1)
}
\arguments{
\item{DT}{DT object output from vl_motif_enrich or vl_motif_cl_enrich}

\item{cex.width}{expansion factor for motif widths}

\item{cex.height}{expansion factor for motif heights}
}
\description{
Add motifs to enrichment plots
}
