% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_plot.R
\name{vl_alluvial_plot}
\alias{vl_alluvial_plot}
\title{bw Average tracks}
\usage{
vl_alluvial_plot(
  dat,
  col = vl_palette_categ2(length(unique(unlist(dat)))),
  ...
)
}
\arguments{
\item{dat}{Data.table containing the variables to plot}

\item{col}{Colors to use for bars. Wrapped with colorRampPalette}

\item{...}{Extra parameters to be passed to barplot}
}
\value{
An object that can be used with the vl_average_bw_track_plot_only() function.
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\examples{
test <- data.table(c("A", "A", "A", "A", "A", "B", "B", "B", "C", "C", "C"),
c("A", "A", "A", "B", "B", "B", "B", "C", "C", "C", "C"),
c("A", "A", "B", "B", "B", "B", "B", "B", "B", "C", "C"))
vl_alluvial_plot(test,
class_levels = c("C", "B", "A"))
}
