% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vAlluvialPlot.R
\name{vl_alluvial_plot}
\alias{vl_alluvial_plot}
\title{bw Average tracks}
\usage{
vl_alluvial_plot(
  dat,
  col = c("tomato", "lightgrey", "cornflowerblue"),
  connectionColDT = NULL
)
}
\arguments{
\item{dat}{Data.table containing the variables to plot}

\item{col}{Colors to use for bars. Wrapped with colorRampPalette}

\item{connectionColDT}{Data.table containing V1/V2/Cc columns with V1/V2 containing all possible transitions. See examples below}
}
\value{
An object that can be used with the vl_average_bw_track_plot_only() function.
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\examples{
if connectionColDT is set to null, the data.table is built internatlly using:
  computeConnectionCol <- function(dat, col)
{
  .c <- CJ(unlist(dat), unlist(dat), unique = T)
  .c[, Cc:= col]
  return(.c)
}
}
