% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_balloons_plot}
\alias{vl_balloons_plot}
\alias{vl_balloons_plot.matrix}
\title{Ballons plot}
\usage{
vl_balloons_plot(x, ...)

\method{vl_balloons_plot}{matrix}(x, color_var, x_breaks, color_breaks,
  col = c("cornflowerblue", "lightgrey", "tomato"), main = NA,
  balloon_size_legend = NA, balloon_col_legend = NA,
  cex.balloons = 4, auto_margins = T)
}
\arguments{
\item{x}{Matrix specifyiung size of the balloons.}

\item{color_var}{Matrix specifyiung color of the balloons.}

\item{color_breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{balloon_size_legend}{Title size legend}

\item{balloon_col_legend}{Title color legend}

\item{cex.balloons}{scaling factor for balloons size}

\item{auto_margins}{Use auto margins? Default= T}
}
\value{
Balloon plot
}
\description{
Starting from two matrices (dot size and colors, respectively), generates a balloon plot
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: matrix_method
}}

\examples{
mat <- matrix(0:11, ncol = 3)
col <- matrix(6:-5, ncol = 3)
vl_balloons_plot(mat)
vl_balloons_plot(x= mat, color_var = col)
vl_balloons_plot(x= mat,
color_var = col,
balloon_size_legend = "test",
balloon_col_legend = "test2")
vl_balloons_plot(x= mat,
color_var = col,
balloon_size_legend = "test",
balloon_col_legend = "test2", 
cex.balloons = 8,
main= "Check!")

}
