% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballons_plot.R
\name{vl_balloons_plot}
\alias{vl_balloons_plot}
\alias{vl_balloons_plot.matrix}
\title{Ballons plot}
\usage{
vl_balloons_plot(x, ...)

\method{vl_balloons_plot}{matrix}(
  x,
  color_var,
  x_breaks,
  color_breaks,
  col = c("cornflowerblue", "white", "tomato"),
  cex.balloons = 1,
  main = NA,
  balloon_size_legend = NA,
  balloon_col_legend = NA,
  auto_margins = T
)
}
\arguments{
\item{x}{Matrix specifyiung size of the balloons.}

\item{color_var}{Matrix specifyiung color of the balloons.}

\item{color_breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{cex.balloons}{Expansion factor for balloons. default= 1.}

\item{main}{Title. Default= NA}

\item{balloon_size_legend}{Title size legend}

\item{balloon_col_legend}{Title color legend}
}
\value{
Balloon plot
}
\description{
Starting from two matrices (dot size and colors, respectively), generates a balloon plot
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: matrix_method
}}

\examples{
x <- matrix(-3:5, ncol= 3)
cols <- matrix(-3:5, ncol= 3)
vl_balloons_plot(x, cols, balloon_size_legend = "Size", balloon_col_legend = "Color")

}
