% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{vl_barplot}
\alias{vl_barplot}
\title{Title}
\usage{
vl_barplot(
  height,
  bar.labels = NULL,
  bar.labels.cex = 0.7,
  xlim = NULL,
  ylim = NULL,
  tilt.names = T,
  individual.var = NULL,
  show.sd = !is.null(individual.var),
  arrow.lwd = 0.5,
  arrow.length = ifelse(length(height) > 1, diff(bar[c(1, 2)])/8, 0.2),
  ind.pch = 16,
  ind.col = adjustcolor("lightgrey", 0.7),
  ind.jitter = 0.2,
  ind.cex = 0.5,
  compute.diff = NULL,
  digits = 2,
  diff.cex = 0.8,
  xpd = NA,
  horiz = F,
  xaxt = "s",
  names.arg = names(height),
  ...
)
}
\arguments{
\item{height}{A vector of values describing the bars which make up the plot.}

\item{xlim}{}

\item{ylim}{}

\item{individual.var}{List of individual variables to be plotted as points.}

\item{arrow.lwd}{Line width of sd arrows Default= .5.}

\item{arrow.length}{Length of sd arrow length. defaults to 1/8 of distance between bar centers.}

\item{ind.pch}{pch value for individual variables}

\item{ind.col}{color value for individual variables}

\item{ind.jitter}{jitter value for individual variables}

\item{ind.cex}{cex factor for points}

\item{compute.diff}{List of pairwise bar pairs to be compared (V2/V11 ratio)}

\item{digits}{Number of digits for the computed diff. Default= 2}

\item{diff.cex}{cex value for pairwise comparisons}

\item{xpd}{xpd value for pairwaise comparisons}

\item{horiz}{Not supported atm}

\item{...}{Extra arguments to be passed to barplot}

\item{sd}{sd values to be plotted as whiskers around bars}
}
\value{
A simple barplot
}
\description{
Title
}
\examples{
vl_barplot(1:3, rep(.2, 3))
}
