% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_binBSgenome}
\alias{vl_binBSgenome}
\title{Bin BS genome}
\usage{
vl_binBSgenome(BSgenome, bin_size = 50)
}
\arguments{
\item{BSgenome}{BSgneome object to use.}

\item{bin_size}{bin size. default to 50bp}
}
\value{
data.table containing bin coordinates
}
\description{
Very fast using data.table
}
\examples{
vl_binBSgenome(BSgenome= BSgenome.Dmelanogaster.UCSC.dm3, bin_size= 50)
}
