% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_binBSgenome}
\alias{vl_binBSgenome}
\title{Bin BS genome}
\usage{
vl_binBSgenome(BSgenome, bins_width = 50, steps_width = bins_width)
}
\arguments{
\item{BSgenome}{BSgenome object to use.}

\item{bins_width}{bins width default to 50bp}

\item{steps_width}{steps width separating each bin. default set to bins_width}
}
\value{
data.table containing bin coordinates
}
\description{
Very fast using data.table
}
\examples{
vl_binBSgenome(BSgenome= BSgenome.Dmelanogaster.UCSC.dm3, bins_width= 50)
}
