% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSgenomeTools.R
\name{vl_binBSgenome}
\alias{vl_binBSgenome}
\title{Bin BS genome}
\usage{
vl_binBSgenome(genome, bins_width = 50L, steps_width = bins_width,
  restrict_seqnames = NULL)
}
\arguments{
\item{genome}{BSgenome object to use. Example "dm3"}

\item{bins_width}{bins width default to 50bp}

\item{steps_width}{steps width separating each bin. default set to bins_width}

\item{restrict_seqnames}{If specified, bins are restricted to provided seqnames}
}
\value{
data.table containing bin coordinates
}
\description{
Very fast using data.table
}
\examples{
vl_binBSgenome(genome= "dm3", bins_width= 50)
}
