% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vboxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\title{box plot}
\usage{
vl_boxplot(x, las.x = 1, las.y = 1, xlab = NA, ylab = NA, ...)
}
\arguments{
\item{x}{Matrix, data.atble or data.frame containing data}

\item{las.x}{las x axis labels}

\item{las.y}{las y axis labels}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\description{
This function is just a wrapper around default boxplot function with a nicer layout
}
\examples{
test <- data.table(test= c(1:10, 50, 50, 90:100),
test1= c(1:10, 50, 50, 90:100)+10)
vl_boxplot(test, xlab = "test", ylab = "test")
}
