% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\alias{vl_boxplot.formula}
\alias{vl_boxplot.data.table}
\title{violin plot}
\usage{
vl_boxplot(x, ...)

\method{vl_boxplot}{formula}(formula, x, ...)

\method{vl_boxplot}{data.table}(x, ...)
}
\arguments{
\item{x}{Matrix, data.atble or data.frame containing data}

\item{...}{Extra arguments set to plot function}

\item{x}{list data}

\item{compute_pval}{A list containing sublists of pairwise x indexes to compute pvalues from. example-> list(c(1,2), c(1,3))}

\item{outline}{show outliers? default=F}

\item{xlab}{}

\item{ylab}{}

\item{xlim}{}

\item{ylim}{}

\item{names}{group labels (x axis). By default, names= names(x). if set to false, no labels are ploted}

\item{boxcol}{box color}

\item{boxwex}{box width}

\item{violin}{If set to FALSE, the violin is not computed/shown}

\item{violcol}{violin color}

\item{violwex}{violin width}

\item{at}{x position for plotting}

\item{trim}{Should violin be trimmed?}

\item{add}{If TRUE, plot on the top of actual plot}
}
\description{
This function plots a violin plot from either matrix,
data.table, data.frame or using formula.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: method for formula

\item \code{data.table}: method for data.table
}}

