% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\alias{vl_boxplot.matrix}
\alias{vl_boxplot.formula}
\alias{vl_boxplot.data.table}
\alias{vl_boxplot.default}
\title{violin plot}
\usage{
vl_boxplot(x, ...)

\method{vl_boxplot}{matrix}(x, ...)

\method{vl_boxplot}{formula}(formula, data = NULL, ...)

\method{vl_boxplot}{data.table}(x, ...)

\method{vl_boxplot}{default}(x, compute_pval, outline = F, xlab = NA,
  ylab = NA, xlim, ylim, names = NULL, boxcol = "white",
  boxwex = 0.15, violin = F, violcol = "white", violwex = 0.35, at,
  trim = T, add = F, ...)
}
\arguments{
\item{x}{Matrix, data.atble or data.frame containing data}

\item{...}{Extra arguments set to plot function}

\item{compute_pval}{A list containing sublists of pairwise x indexes to compute pvalues from. example-> list(c(1,2), c(1,3))}

\item{outline}{show outliers? default=F}

\item{xlab}{}

\item{ylab}{}

\item{xlim}{}

\item{ylim}{}

\item{names}{group labels (x axis). By default, names= names(x). if set to false, no labels are ploted}

\item{boxcol}{box color}

\item{boxwex}{box width}

\item{violin}{If set to FALSE, the violin is not computed/shown}

\item{violcol}{violin color}

\item{violwex}{violin width}

\item{at}{x position for plotting}

\item{trim}{Should violin be trimmed?}

\item{add}{If TRUE, plot on the top of actual plot}

\item{x}{list data}
}
\description{
This function plots a violin plot from either matrix,
data.table, data.frame or using formula.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: method for matrices

\item \code{formula}: method for formula

\item \code{data.table}: method for data.table

\item \code{default}: default method
}}

\examples{
# Create test matrix
set.seed(1234)
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")

vl_boxplot(test, violin= T, outline= T, compute_pval= list(c(1,2),c(1,3),c(3,5)))

}
