% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\alias{vl_boxplot.default}
\alias{vl_boxplot.matrix}
\alias{vl_boxplot.formula}
\title{Boxplot}
\usage{
vl_boxplot(x, ...)

\method{vl_boxplot}{default}(x, ..., compute_pval = NULL,
  pval_offset = 0.08, tilt.names = F, range = 1.5, width = NULL,
  varwidth = FALSE, notch = FALSE, outline = FALSE, names,
  plot = TRUE, border = par("fg"), col = NULL, log = "",
  pars = list(boxwex = ifelse(violin, 0.2, 0.4), staplewex = NA, outwex =
  NA), horizontal = FALSE, add = FALSE, at = NULL, frame = F,
  whisklty = ifelse(violin, 2, 1), ylim = NULL, xaxt = "s",
  violin = T, viocol = NULL, viowex = 0.4)

\method{vl_boxplot}{matrix}(x, use.cols = TRUE, ...)

\method{vl_boxplot}{formula}(formula, data = NULL, ..., subset,
  na.action = NULL)
}
\arguments{
\item{x}{list of variables to be plotted}

\item{...}{Extra parameters for boxplot()}

\item{compute_pval}{list of vectors of length two containing pairwise x indexes to be compared}

\item{pval_offset}{offset for pval plotting. Defaults to 0.04 (fraction of ylim)}

\item{tilt.names}{Should names be tilted (ignored if horizontal= TRUE)}

\item{names}{Names to plot under boxplot. If function specified, applied to names before plotting}

\item{violin}{Should violins be added?}

\item{viocol}{Violin colors}

\item{viowex}{Expansion factor for violins}
}
\description{
Just a wrapper around boxplot that makes it nicer and allows to add wilcox pvals
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{matrix}: method for matrices

\item \code{formula}: method for formula
}}

\examples{
set.seed(1234)
vl_boxplot(formula= len~supp*dose, data=ToothGrowth, notch=TRUE,
col= c("gold","darkgreen"),
main="Tooth Growth", xlab="Suppliment and Dose",
compute_pval= list(c(1,2), c(5,6), c(1,6)))
}
