% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vboxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\title{violin plot}
\usage{
vl_boxplot(
  x,
  formula,
  outline = F,
  xlab = "",
  ylab = "",
  plot_labels = T,
  xlim = c(0.5, max(obj$at) + 0.5),
  ylim = if (outline) range(unlist(obj$value)) else range(c(obj$Q1, obj$Q5)),
  col = "white",
  at = NULL
)
}
\arguments{
\item{x}{Matrix, data.atble or data.frame containing data}

\item{formula}{Formula to apply to the data}

\item{outline}{plot outlier values? default= F}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{plot_labels}{Should box labels be plotted? Default= T}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}
}
\description{
This function plots a violin plot from either matrix,
data.table, data.frame or using formula.
}
