% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{vl_boxplot}
\alias{vl_boxplot}
\alias{vl_boxplot.default}
\alias{vl_boxplot.matrix}
\alias{vl_boxplot.formula}
\title{Boxplot}
\usage{
vl_boxplot(x, ...)

\method{vl_boxplot}{default}(
  x,
  ...,
  compute.pval = NULL,
  pval.FUN = function(x, y) wilcox.test(x, y)$p.value,
  pval.cex = 0.8,
  pval.stars = T,
  pval.values = F,
  tilt.names = F,
  srt = 45,
  range = 1.5,
  width = NULL,
  varwidth = FALSE,
  notch = FALSE,
  outline = FALSE,
  names,
  plot = TRUE,
  border = par("fg"),
  col = NULL,
  log = "",
  pars = list(boxwex = ifelse(violin, 0.2, 0.4), staplewex = NA, outwex = NA),
  horizontal = FALSE,
  add = FALSE,
  at = NULL,
  frame = F,
  whisklty = ifelse(violin, 2, 1),
  lwd = par("lwd"),
  ylim = NULL,
  xaxt = "s",
  violin = FALSE,
  viocol = "#FFFFFF00",
  viowex = 0.4
)

\method{vl_boxplot}{matrix}(x, use.cols = TRUE, ...)

\method{vl_boxplot}{formula}(formula, data = NULL, ..., subset, na.action = NULL)
}
\arguments{
\item{x}{list of variables to be plotted}

\item{...}{Extra parameters for boxplot()}

\item{compute.pval}{list of vectors of length two containing pairwise x indexes to be compared}

\item{pval.cex}{cex value for pval plotting. Default= .8.}

\item{pval.stars}{Should pvalues stars be shown?}

\item{pval.values}{Should pval values be shown?}

\item{tilt.names}{Should names be tilted (ignored if horizontal= TRUE)}

\item{srt}{rotation angle for titled names}

\item{names}{Names to plot under boxplot. If function specified, applied to names before plotting}

\item{violin}{Should violins be added?}

\item{viocol}{Violin colors. Default to transparent}

\item{viowex}{Expansion factor for violins}
}
\description{
Just a wrapper around boxplot that makes it nicer and allows to add wilcox pvals
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_boxplot(default)}: default method

\item \code{vl_boxplot(matrix)}: method for matrices

\item \code{vl_boxplot(formula)}: method for formula

}}
\examples{
set.seed(1234)
vl_boxplot(formula= len~supp*dose, data=ToothGrowth, notch=TRUE,
col= c("gold","darkgreen"),
main="Tooth Growth", xlab="Suppliment and Dose",
compute.pval= list(c(1,2), c(5,6), c(1,6)))
}
