% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_boxplot_pval}
\alias{vl_boxplot_pval}
\title{Plot pval}
\usage{
vl_boxplot_pval(
  x,
  compare,
  adj = 10,
  outline = F,
  las = 2,
  xlab = NA,
  ylim = c(min(box$stats, na.rm = T), max(obj$y) + inter),
  lwd = 0.75,
  staplewex = NA,
  whisklty = 1,
  whisklwd = 0.75,
  boxwex = 0.5,
  boxlwd = 0.75,
  medlwd = 1,
  ...
)
}
\arguments{
\item{x}{list corresponding to the different groups}

\item{compare}{A list of length 2 vectors specifying the groups to compare (can be either integers or group names)}

\item{adj}{Ratio of the boxplots' range that will be used to space pval segments}

\item{outline}{Should oultiers be ploted?}

\item{las}{boxplot las}

\item{xlab}{boxplot xlab}

\item{ylim}{boxplot ylim}

\item{staplewex}{boxplot staplewex}

\item{whisklty}{boxplot whisklty}

\item{boxwex}{boxplot boxwex}

\item{...}{Extra parameters to be passed to boxplot}
}
\value{
An object that can be used with the vl_average_bw_track_plot_only() function.
}
\description{
Plots formated pval data as text
}
\examples{
y <- split(InsectSprays$count, InsectSprays$spray)
vl_boxplot_pval(y, 
compare = list(c(1,2), c(1,3), c(1,5), c(4,5), c(5, 6)))
vl_boxplot_pval(y, 
compare = list(c("A","B"), c("A","C"), c("A","E"), c("D","E"), c("E", "F")))
}
