% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_bsub}
\alias{vl_bsub}
\title{Stark lab wrappers}
\usage{
vl_bsub(
  cmd,
  cores = 6L,
  m = 12,
  t = "08:00:00",
  name = "vl",
  o = NA,
  e = NA,
  wdir = getwd(),
  execute = T,
  d = NA
)
}
\arguments{
\item{cmd}{Command to be submitted}

\item{cores}{Number of cores. default= 12L}

\item{m}{Memory in Go. Default= 12}

\item{t}{Time parameter for slurm in the format hh:mm:ss. Default= '08:00:00'. For 2 days:'2-00:00:00'}

\item{name}{Name of the job. Default= "vl}

\item{o}{Std output directory}

\item{e}{Std error directory}

\item{wdir}{Working directory. defaults to getwd()}

\item{execute}{If FALSE, simply returns the full command. Default= TRUE}

\item{d}{Name of te dependent job}
}
\description{
bsub submits to the cluster
}
