% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vl_bsub}
\alias{vl_bsub}
\title{Stark lab wrappers}
\usage{
vl_bsub(
  cmd,
  cores = 1,
  name = "vl",
  d = NA,
  m = NA,
  o = NA,
  e = NA,
  t = NA,
  execute = T
)
}
\description{
bsub submits to the cluster
}
