% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_average_track}
\alias{vl_bw_average_track}
\alias{plot.vl_bw_average_track}
\title{bw Average tracks}
\usage{
vl_bw_average_track(
  bed,
  tracks,
  set_IDs = 1,
  upstream = 5000,
  downstream = 5000,
  stranded = F,
  nbins = 101L,
  names = gsub(".bw$", "", basename(tracks)),
  center_label = "Center",
  plot = T,
  xlab = "genomic distance",
  ylab = "Enrichment",
  ylim,
  col = c("#E69F00", "#68B1CB", "#15A390", "#96C954", "#77AB7A", "#4F6A6F", "#D26429",
    "#C57DA5", "#999999"),
  legend = T,
  legend.cex = 1
)

\method{plot}{vl_bw_average_track}(
  obj,
  xlab = "genomic distance",
  xlab.at,
  center_label = "Center",
  ylab = "Enrichment",
  ylim,
  legend = T,
  legend.cex = 1
)
}
\arguments{
\item{bed}{Regions to plot. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{set_IDs}{Set IDs specifying the groups as subsets of the bed file}

\item{upstream}{Upstream  extension of bed regions (centered on center)}

\item{downstream}{Downstream  extension of bed regions (centered on center)}

\item{stranded}{Should the average track be stranded?}

\item{nbins}{Number of bins spanning the extended regions. Default= 101L}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}

\item{plot}{Should the average track be ploted? default= T}

\item{xlab}{X label. default= "genomic distance"}

\item{ylab}{Y labels. default= "Enrichment"}

\item{ylim}{ylim for plotting. default= range(data)}

\item{col}{Color to be used for plotting.}

\item{legend}{Should the legeng be plotted? default to T}

\item{legend.cex}{Legend cex. defaults to 1}
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Method to plot average tracks
}}

\examples{
bed <- rbind(vl_SUHW_top_peaks, vl_STARR_DSCP_top_peaks, fill= T)
sets <- c(rep("suhw", 100), rep("STARR", 1000))
tracks <- c("../available_data_dm3/db/bw/GSE41354_SuHw_rep1_uniq.bw", "../gw_STARRSeq_bernardo/db/bw/DSCP_200bp_gw.UMI_cut_merged.bw")
vl_bw_average_track(bed, tracks, plot= T, upstream = 1000, downstream = 1000, set_IDs = sets)
}
