% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbw_coverage.R
\name{vl_bw_coverage}
\alias{vl_bw_coverage}
\title{bw coverage}
\usage{
vl_bw_coverage(bed, bw)
}
\arguments{
\item{bed}{Regions to quantify. Can be a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns"}

\item{bw}{Path to target bw file (character vector)}
}
\description{
Quantify bw file at a set of intervals. returns mean value/region
}
