% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_coverage_bins}
\alias{vl_bw_coverage_bins}
\title{Bins bed file and compute bw signal}
\usage{
vl_bw_coverage_bins(
  bed,
  set.IDs = 1L,
  tracks,
  names = gsub(".bw$", "", basename(tracks)),
  anchor = "center",
  upstream = 5000L,
  downstream = 5000L,
  nbins = if (anchor == "center") 501L else if (anchor == "region") c(100L, 301L, 100L),
  ignore.strand = FALSE,
  genome
)
}
\arguments{
\item{bed}{Regions to plot. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns.}

\item{set.IDs}{Set IDs specifying the groups as subsets of the bed file. Default= 1L (unique subset).}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{names}{Track names. Defaults to the basenames of .bw files.}

\item{anchor}{Anchor point to use. Must be one of 'center' or 'region' (meaning that the whole region will be used as anchor point, and extended by upstream/downstream values). Default= 'center'.}

\item{upstream}{Upstream  extension of bed regions (centered on center). Default= 5000L.}

\item{downstream}{Downstream  extension of bed regions (centered on center). Default= 5000L.}

\item{nbins}{A single integer value (anchor= 'center') or a vector of 3 integers (anchor= 'region') specifying the number of bins to be used. When anchor is set to 'regions', the 3 values will correspond to the number of bins before the region, the number of bins within the region and the number of bins after. Default= 501L (anchor= 'center') or c(100L, 301L, 100L) (anchor= 'region').}

\item{ignore.strand}{Should the strand be ignored? Default= FALSE.}

\item{genome}{BSgenome used to check limits of extended regions. Out of limit ranges will be resized accordingly}
}
\description{
Compared to vl_bw_coverage, this functions bins bed file before quantifying the signal. Useful to make heatmaps/average tracks
}
\examples{
bed <- rbind(vl_SUHW_top_peaks, vl_STARR_DSCP_top_peaks, fill= T)
sets <- c(rep("suhw", 100), rep("STARR", 1000))
tracks <- c("/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE41354_SuHw_rep1_uniq.bw",
"/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/bw/DSCP_200bp_gw.UMI_cut_merged.bw")
vl_bw_coverage_bins(bed, set.IDs = sets, tracks= tracks)[]
vl_bw_coverage_bins(bed, set.IDs = sets, tracks= tracks, anchor= "region")[]

}
