% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_heatmap}
\alias{vl_bw_heatmap}
\alias{plot.vl_bw_heatmap}
\title{bw Average heatmap}
\usage{
vl_bw_heatmap(
  bed,
  set.IDs = 1,
  tracks,
  upstream = 5000L,
  downstream = 5000L,
  ignore.strand = T,
  nbins = 101L,
  space = 20L,
  names = gsub(".bw$", "", basename(tracks)),
  plot = T,
  center.label = "Center",
  col = c("#440154FF", "#482878FF", "#3E4A89FF", "#31688EFF", "#26828EFF", "#1F9E89FF",
    "#35B779FF", "#6DCD59FF", "#B4DE2CFF", "#FDE725FF"),
  order.col = 1,
  max = NULL,
  fun.order = function(x) mean(x, na.rm = T),
  fun.max = function(x) quantile(x, 0.99, na.rm = T),
  na.col = "lightgrey",
  cex.labels = 0.6
)

\method{plot}{vl_bw_heatmap}(obj)
}
\arguments{
\item{bed}{Regions to plot. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{set.IDs}{Set IDs specifying the groups as subsets of the bed file. Default= 1L (no subgroups).}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid issues.}

\item{upstream}{Upstream  extension of bed regions (centered on center). Default= 5000L.}

\item{downstream}{Downstream  extension of bed regions (centered on center). Default= 5000L.}

\item{ignore.strand}{Should the strand information be ignored? Default= FALSE.}

\item{nbins}{Number of bins spanning the extended regions. Default= 101L.}

\item{names}{Factors used for ordering and naming of plots. By default, .bw basenames will be used (as factors).}

\item{plot}{Should the average track be ploted? Default= TRUE.}

\item{center.label}{Label center heatmap.}

\item{col}{Vector of colors to be used for heatmap.}

\item{order.col}{Index of the column(s) to be used for ordering. If set to FALSE, no ordering besides set.IDs. Default= 1L.}

\item{max}{Allows to manually specify max values. Otherwise, max is computed using the fun.max function for each track.}

\item{fun.order}{Function used to aggregated per region and order heatmap. Default= function(x) mean(x, na.rm= T)}

\item{fun.max}{Function to be used for clipping. Default= function(x) quantile(x, 0.995, na.rm= T). To avoid clipping, use: function(x) max(x, na.rm= T)}

\item{cex.labels}{cex paramter for labels}
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot(vl_bw_heatmap)}: Method to plot bw heatmaps

}}
\examples{
bed <- list(suhw= vl_SUHW_top_peaks,
            STARR= vl_STARR_DSCP_top_peaks)
bed <- rbindlist(bed,
                 idcol = T,
                 fill = T)
tracks <- c("/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE41354_SuHw_rep1_uniq.bw", 
            "/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/bw/DSCP_200bp_gw.UMI_cut_merged.bw")
test <- vl_bw_heatmap(bed,
                      set.IDs= bed$.id,
                      tracks,
                      plot= T,
                      upstream = 1000,
                      downstream = 1000,
                      fun.order = mean,
                      order.col= 2)
plot(test)

}
