% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_heatmap}
\alias{vl_bw_heatmap}
\alias{plot.vl_bw_heatmap}
\title{bw Average heatmap}
\usage{
vl_bw_heatmap(
  bed,
  tracks,
  set_IDs = 1,
  upstream = 5000,
  downstream = 5000,
  stranded = F,
  nbins = 101L,
  names = gsub(".bw$", "", basename(tracks)),
  plot = T,
  col = c("blue", "yellow"),
  order_cols = 1,
  order_FUN = function(x) mean(x, na.rm = T),
  max_FUN = function(x) quantile(x, 0.995, na.rm = T),
  na_col = "lightgrey"
)

\method{plot}{vl_bw_heatmap}(
  obj,
  col = c("blue", "yellow"),
  order_cols = 1,
  order_FUN = function(x) mean(x, na.rm = T),
  max_FUN = function(x) quantile(x, 0.995, na.rm = T),
  na_col = "lightgrey"
)
}
\arguments{
\item{bed}{Regions to plot. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{set_IDs}{Set IDs specifying the groups as subsets of the bed file}

\item{upstream}{Upstream  extension of bed regions (centered on center)}

\item{downstream}{Downstream  extension of bed regions (centered on center)}

\item{stranded}{Should the average track be stranded?}

\item{nbins}{Number of bins spanning the extended regions. Default= 100}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}

\item{plot}{Should the average track be ploted? default= T}

\item{col}{Vector of colors to be used for heatmap}

\item{order_cols}{Numeric vector giving the indexes of the tracks to be used for ordering}

\item{order_FUN}{Function used to aggregated per region and order heatmap. default= function(x) mean(x, na.rm= T)}

\item{max_FUN}{Function to be used for clipping. default= function(x) quantile(x, 0.995, na.rm= T). Using max -> no clipping}
}
\description{
Plots average tracks for a set bw files around (potentially) several sets of peaks
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot(vl_bw_heatmap)}: Method to plot bw heatmaps

}}
\examples{
bed <- rbind(vl_SUHW_top_peaks, vl_STARR_DSCP_top_peaks, fill= T)
tracks <- c("/groups/stark/vloubiere/projects/available_data_dm3/db/bw/GSE41354_SuHw_rep1_uniq.bw", 
"/groups/stark/vloubiere/projects/gw_STARRSeq_bernardo/db/bw/DSCP_200bp_gw.UMI_cut_merged.bw")
set_IDs <- c(rep("suhw", 100), rep("STARR", 1000))
test <- vl_bw_heatmap(bed, tracks, set_IDs= set_IDs, plot= T, upstream = 1000, downstream = 1000, order_FUN = mean, order_cols = 2)
plot(test, order_col= 2)
plot(test)
}
