% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_merge}
\alias{vl_bw_merge}
\title{Merge several bigwig files into 1}
\usage{
vl_bw_merge(
  tracks,
  genome,
  bins.width = 25L,
  output,
  restrict.seqnames = NULL,
  scoreFUN = NULL
)
}
\arguments{
\item{genome}{BSgenome. "dm3", "dm6", "mm10"}

\item{bins.width}{Bin size. default= 25L}

\item{output}{Output bw file path}

\item{restrict.seqnames}{If specified, bins are restricted to provided seqnames. Default= NULL}

\item{scoreFUN}{A function to be applied to the score column before return}

\item{bw}{A character vector containing the path of bw files to be merged}
}
\description{
Convenience function to merge directly into r
}
