% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_merge}
\alias{vl_bw_merge}
\title{Merge bw files}
\usage{
vl_bw_merge(x, output, BSgenome = BSgenome, bins = NA)
}
\arguments{
\item{x}{fastq file path (read1)}

\item{output}{Output files}

\item{BSgenome}{Used to extract bw}

\item{bins}{bin size used to simplify bw. Default to NA (no simplification)}
}
\description{
Merge bw files
}
