% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwTools.R
\name{vl_bw_merge}
\alias{vl_bw_merge}
\title{Merge several bigwig files into 1}
\usage{
vl_bw_merge(
  tracks,
  genome,
  scoreFUN = function(x) mean(x),
  bins.width = 1L,
  output,
  restrict.seqnames = NULL
)
}
\arguments{
\item{genome}{BSgenome. "dm3", "dm6", "mm10"}

\item{scoreFUN}{The function to be apply to scores. Default= mean.}

\item{bins.width}{Bin size. Default= 1L}

\item{output}{Output bw file path}

\item{restrict.seqnames}{If specified, bins are restricted to provided seqnames. Default= NULL}

\item{bw}{A character vector containing the path of bw files to be merged}
}
\description{
Convenience function to merge directly into r
}
