% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMergebw.R
\name{vl_bw_merge}
\alias{vl_bw_merge}
\title{Merge bw files}
\usage{
vl_bw_merge(x, output_folder, output_prefix)
}
\arguments{
\item{x}{fastq file path (read1)}

\item{output_folder}{Folder where to put processed files}

\item{output_prefix}{prefix for processed files}
}
\description{
Merge bw files
}
\examples{
vl_bw_merge(x= c("db/bw/cutnrun_reps/H3K27Ac_PH18_rep1.bw",
"db/bw/cutnrun_reps/H3K27Ac_PH18_rep2.bw"), 
output_folder = "db/bw/cutnrun_merge/",
output_prefix = "H3K27Ac_PH18_merge.bw")
}
