% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_closestBed}
\alias{vl_closestBed}
\title{Find closestBed regions}
\usage{
vl_closestBed(a, b = NULL, n = 1, min_dist = 0)
}
\arguments{
\item{a}{Granges or data.table FOR which closest features have to be found.}

\item{b}{Granges or data.table FROM which closest features have to be found. If set to NULL (default), then a is matched to itself.}

\item{n}{Number of closest features to be reported. Default to 1}

\item{min_dist}{Min distance for closest feature}
}
\value{
Return "a" coor and closeet "b" coordinates together with distance
}
\description{
For each line of a file, returns the closest lines in b
}
\examples{
To all closest features
vl_closestBed(vl_STARR_DSCP_top_peaks)
To find closet yet non touching features
vl_closestBed(vl_STARR_DSCP_top_peaks, min_dist = 1)
To n closest features
vl_closestBed(vl_STARR_DSCP_top_peaks, min_dist = 1, n= 2)

}
