% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_closestBed}
\alias{vl_closestBed}
\title{Find closestBed regions}
\usage{
vl_closestBed(a, b = NULL, n = Inf, min_dist = 0)
}
\arguments{
\item{a}{Granges or data.table FOR which closest features have to be found.}

\item{b}{Granges or data.table FROM which closest features have to be found. If set to NULL (default), then a is matched to itself.}

\item{n}{Number of closest features to be reported. Default to Inf}

\item{min_dist}{Min distance for closest feature}
}
\value{
Return "a" coor and closeet "b" coordinates together with distance
}
\description{
For each line of a file, returns the closest lines in b
}
\examples{
a <- data.table(seqnames= "chr2L", start= sample(10000, 1000))
a[, end:= start+1000]

To all closest features
vl_closestBed(a, min_dist= 0)

To find closet yet non touching features
vl_closestBed(a, min_dist= 1)

}
