% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vCollapseDTRanges.R
\name{vl_collapse_DT_ranges}
\alias{vl_collapse_DT_ranges}
\title{Collapse ranges with a certain min dist}
\usage{
vl_collapse_DT_ranges(bed, mingap = 1, stranded = F)
}
\arguments{
\item{bed}{GRanges or data.table file with "seqnames", "start", "end" (and optionally strand, set to * if absent)}

\item{mingap}{min gap ditance for merging. default is 1 (that is, touching coordinates)}

\item{stranded}{If set to true and strand column is provided, only merges coordinates with similar strand.}
}
\value{
Collapse coor data.table
}
\description{
Collapse ranges with a certain min dist between them
}
\examples{
ex <- GRanges(c("chr3L", "chr3L", "chr3L", "chr3L", "chr2R"), 
IRanges(c(1000, 2000, 2100, 2200, 2000), 
c(1500, 2099, 2199, 2299, 3000)),
strand= c("+","+","+","-","+"))
vl_collapse_DT_ranges(ex, mingap= 1, stranded= F)
vl_collapse_DT_ranges(ex, mingap= 1, stranded= T)

}
