% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_collapseBed}
\alias{vl_collapseBed}
\title{Collapse ranges with a certain min dist}
\usage{
vl_collapseBed(bed, mingap = 1, return_idx_only = F, ignore.strand = T)
}
\arguments{
\item{bed}{bed file to be collapsed. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames','start', 'end' columns. see ?vl_importBed()}

\item{mingap}{min gap ditance for merging. default is 1 (that is, touching coordinates)}

\item{return_idx_only}{If set to T, does not collapse regions but returns idx as an extra columns. default= F}

\item{ignore.strand}{Should strand be ignored? default= T}
}
\value{
Collapse coor data.table
}
\description{
Collapse ranges with a certain min dist between them
}
\examples{
gr <- GRanges(seqnames=Rle(paste("chr", c(1, 2, 1, 3), sep=""), c(1, 3, 2, 4)),
ranges=IRanges(1:10, width=10:1, names=letters[1:10]),
strand=Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)))
bed <- vl_importBed(gr)
vl_collapseBed(bed)
vl_collapseBed(bed, return_idx_only = T)
vl_collapseBed(bed, ignore.strand = F)
vl_collapseBed(bed, ignore.strand = F, return_idx_only = T)

# Mingap
bed[7, c("start", "end"):= .(start+10, end+10)]
vl_collapseBed(bed)
vl_collapseBed(bed, mingap = 10)
vl_collapseBed(bed, return_idx_only = T)

}
