% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_collapseBed}
\alias{vl_collapseBed}
\title{Collapse ranges with a certain min dist}
\usage{
vl_collapseBed(bed, ming.gap = 1, return.idx.only = F, ignore.strand = T)
}
\arguments{
\item{bed}{bed file to be collapsed. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames','start', 'end' columns. see ?vl_importBed()}

\item{ming.gap}{min gap ditance for merging. default is 1 (that is, touching coordinates)}

\item{return.idx.only}{If set to T, does not collapse regions but returns idx as an extra columns. default= F}

\item{ignore.strand}{Should strand be ignored? default= T}
}
\value{
Collapse bed data.table
}
\description{
Collapse ranges with a certain min dist between them
}
\examples{
bed <- data.table(seqnames= "chr3R",
                  start= c(10e3, 11e3, 12e3, 20e3, 21e3),
                  end= c(12e3, 13e3, 14e3, 22e3, 23e3),
                  strand= c("+", "+", "-", "+", "+"))
                  
vl_collapseBed(bed)
# Return mergin index only
vl_collapseBed(bed, return.idx.only = T)
# Only merge if strand is similar
vl_collapseBed(bed, ignore.strand = F)
# Allow a certain gap for merging
vl_collapseBed(bed, ming.gap = 1000)
vl_collapseBed(bed, ming.gap = 10000)

}
