% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_contrib_addMotif}
\alias{vl_contrib_addMotif}
\title{Add motif columns to contribution object}
\usage{
vl_contrib_addMotif(
  contrib,
  sel = vl_motifs_DB_v2[collection == "jaspar", motif_ID],
  motifDB = vl_motifs_DB_v2,
  p.cutoff = 5e-05,
  genome,
  bg = "genome"
)
}
\arguments{
\item{contrib}{A contribution object generated by ?vl_importContrib()}

\item{sel}{A selection of motif_IDs. Default= vl_motifs_DB_v2[collection=="jaspar", motif_ID]}

\item{motifDB}{The motifDB to be used. Default= vl_motifs_DB_v2}

\item{p.cutoff}{The cutoff to be used for calling motifs. Default= 5e-5}

\item{genome}{The genome to be used}

\item{bg}{background. Default= "genome".}
}
\value{
A contribution data.table containing, for each motif_ID, a vector corresponding to its matches.
}
\description{
Add motif columns to contribution object
}
