% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcontrol_regions_BSgenome.R
\name{vl_control_regions_BSgenome}
\alias{vl_control_regions_BSgenome}
\title{Generate random control regions}
\usage{
vl_control_regions_BSgenome(BSgenome, n, width, restrict_seqnames = NULL)
}
\arguments{
\item{BSgenome}{BSgneome object to use.}

\item{n}{Number of regions to sample}

\item{width}{Widths of regions to sample (length should either be 1 or equal n)}

\item{restrict_seqnames}{If specified, only the providsed seqnames will be used}
}
\value{
data.table containing random control regions
}
\description{
Generate control regions with given width distribution
}
\examples{
vl_control_regions_BSgenome(BSgenome = BSgenome.Dmelanogaster.UCSC.dm3, 
n= 5000, 
width= 1000, 
restrict_seqnames = c("chr2L", "chr2R", "chr3L", "chr3R", "chr4", "chrX"))

}
