% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIP_CUT&RUN_pipeline.R
\name{vl_CUTNRUN_peakCalling}
\alias{vl_CUTNRUN_peakCalling}
\title{Title}
\usage{
vl_CUTNRUN_peakCalling(
  processed_metadata,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  extsize = 300,
  cores = 8,
  mem = 64,
  overwrite = F,
  submit = F,
  wdir = getwd(),
  logs = "db/logs/CUTNRUN/peak_calling",
  time = "1-00:00:00"
)
}
\arguments{
\item{processed_metadata}{Processed metadata file generated by vl_CUTNRUN_processing}

\item{Rpath}{The path to the Rscript executable to use, on which the latest version of the vlfunction package should be installed. Default= /software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript}

\item{cores}{Number of cores per job. Default= 8}

\item{mem}{Memory per job (in Go). Default= 32.}

\item{overwrite}{Should existing files be overwritten?}

\item{submit}{Should the command be submitted? default= G}

\item{wdir}{The working directory to use. defaut= getwd().}

\item{logs}{Path to save logs. Default= "db/logs/CUTNRUN/peak_calling"}

\item{time}{The time required for the SLURM scheduler. Default= '1-00:00:00'}

\item{extsise}{The extsize to be used, meaning that the building of the peak model will be skipped (--nomodel). To build the model, set extize= NA. Default= 300}
}
\value{
Command lines for peak calling.
}
\description{
Title
}
\examples{
# Peak calling example dataset ----
library(vlfunctions)

check <- readRDS("Rdata/metadata_CutNRun_processed.rds") # Check all processed files exist

vl_CUTNRUN_peakCalling(processed_metadata = "Rdata/metadata_CutNRun_processed.rds",
                       extsize = 300,
                       cores = 8,
                       mem = 64,
                       overwrite = FALSE,
                       submit = TRUE)

}
