% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIP_CUT&RUN_pipeline.R
\name{vl_CUTNRUN_processing}
\alias{vl_CUTNRUN_processing}
\alias{vl_CUTNRUN_processing.character}
\alias{vl_CUTNRUN_processing.default}
\title{ORFtag pipeline}
\usage{
vl_CUTNRUN_processing(metadata, ...)

\method{vl_CUTNRUN_processing}{character}(
  metadata,
  processed_metadata_output = gsub(".xlsx$", "_processed.rds", metadata),
  ...
)

\method{vl_CUTNRUN_processing}{default}(
  metadata,
  processed_metadata_output = gsub(".txt$", "_processed.rds", metadata),
  scratch_folder = "/scratch/stark/vloubiere",
  cores = 8,
  mem = 64,
  overwrite = F,
  submit = F,
  wdir = getwd(),
  logs = "db/logs/CUTNRUN/processing",
  time = "1-00:00:00"
)
}
\arguments{
\item{metadata}{The path to a .txt, tab-separated metadata file containing at least 12 columns. See vl_metadata_CUTNRUN for an example.}

\item{processed_metadata_output}{An .rds path where to save the processed metadata file (containing the paths of output files). By default, when importing the metadata from an excel sheet, "_processed.rds" will be appended to the excel file path.}

\item{scratch_folder}{Folder where intermediate bam and fastq files will be saved}

\item{cores}{Number of cores per job. Default= 8}

\item{mem}{Memory per job (in Go). Default= 32.}

\item{overwrite}{Should existing files be overwritten?}

\item{submit}{Should the command be submitted? default= FALSE.}

\item{wdir}{The working directory to use. defaut= getwd().}

\item{logs}{Path to save logs. Default= "db/logs/CUTNRUN/processing"}

\item{time}{The time required for the SLURM scheduler. Default= '1-00:00:00'}
}
\value{
Return a data.table containing, for each sampleID, the concatenated commands that are required to process the data. These commands can then be submitted using ?vl_bsub().
}
\description{
Takes as input a (correctly formatted) metadata file, saves the processed metadata file and returns the command lines to: \cr
1/ extract reads from VBC bam file \cr
2/ trim the reads and align to mouse/human genome (see 'genome' column of the metadata table) \cr
3/ and return alignment statistics as well as collapsed reads and
6/ assign insertions to closest downstream genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_CUTNRUN_processing(character)}: for excel files path

\item \code{vl_CUTNRUN_processing(default)}: default method

}}
\examples{
# Process example dataset ----
library(vlfunctions)
meta <- vl_metadata_CUTNRUN
vl_CUTNRUN_processing(metadata = meta,
                      processed_metadata_output = "Rdata/metadata_CutNRun_processed.rds",
                      cores= 8,
                      mem= 64,
                      overwrite= FALSE,
                      submit= TRUE)
check <- readRDS("Rdata/metadata_CutNRun_processed.rds")

# Peak calling ----
vl_CUTNRUN_peakCalling(processed_metadata = "Rdata/metadata_CutNRun_processed.rds",
                       extsize = 300,
                       cores = 8,
                       mem = 64,
                       overwrite = FALSE,
                       submit = TRUE)

}
