% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_density.R
\name{vl_densScatterplot}
\alias{vl_densScatterplot}
\title{Title}
\usage{
vl_densScatterplot(
  x,
  y,
  label,
  xlim = range(x, na.rm = T),
  ylim = range(y, na.rm = T),
  col = rainbow(length(unique(label))),
  xlab = deparse1(substitute(x)),
  ylab = deparse1(substitute(y)),
  dens.lw = 1,
  frame = F,
  ...
)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{label}{Labels to split the data}

\item{xlim}{x axis lims}

\item{ylim}{y axis lims}

\item{col}{Colors to be used (must match number of levels in label)}

\item{xlab}{x axis labels}

\item{ylab}{y axis labels}

\item{dens.lw}{Width of the density lines to be drawn on the top and right of the scatterplot}

\item{...}{Extra arguments to be passed to plot().}
}
\value{
A scatterplot with density lines on the side
}
\description{
Title
}
\examples{
set.seed(1)
dat <- data.table(x= rnorm(1000, 1, 0.5),
y= rnorm(1000, 1, 0.5),
label= sample(rep(c("A", "B"), each= 500), 1000))
par(mai= c(0.9, 0.9, 0.9, 0.9),
tcl= -0.1,
mgp= c(2, .5, 0),
las= 1)
vl_densScatterplot(dat$x, dat$y, dat$label)

}
