% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_fb_go_table_dm6_FB2020_05}
\alias{vl_fb_go_table_dm6_FB2020_05}
\title{Dm6 GO database}
\format{
An object containing Motifs and related metadata
\describe{
  \item{GO}{GO ID}
  \item{name}{GO name}
  \item{type}{GO category}
  \item{partents}{parent GOs}
  \item{children}{children GOs}
  \item{Fbgn}{Gene flybase ID}
  \item{symbol}{Gene symbol}
  ...
}
}
\source{
{"../../vlfunction/data/vl_fb_go_table_dm6_FB2020_05.RData"}
}
\usage{
Can be used with vl_GO_cluster. see ?vl_GO_cluster()
}
\description{
Contains GO annotation from FB2020_05
}
\examples{
Object was generated using the followin code:
go <- fread("../../genomes/dm6/FB2020_05_gene_association.fb",
skip= 5,
select = c(2,3,5))
colnames(go) <- c("FBgn", "Symbol", "GO")
go_details <- get_ontology("../../genomes/dm6/FB2020_05_go-basic.obo",
extract_tags = "everything")
go_details <- data.table(GO= go_details$id,
name= go_details$name,
type= go_details$namespace,
parents= go_details$parents,
children= go_details$children)
vl_fb_go_table_dm6_FB2020_05 <- go_details
save(vl_fb_go_table_dm6_FB2020_05, file= "../../vlfunction/data/vl_fb_go_table_dm6_FB2020_05.RData")

}
\keyword{datasets}
