% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_genes_set}
\alias{vl_genes_set}
\title{Set of genes}
\format{
FBgn, symbol and GO (RpL/HOX)
}
\usage{
see ?vl_GO_enrich()
}
\description{
Example set that can be used for GO enrichment. Contain RpL and HOX genes
}
\keyword{datasets}
