% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vGOcluster.R
\name{vl_GO_clusters}
\alias{vl_GO_clusters}
\title{GO analysis}
\usage{
vl_GO_clusters(FBgn_list, go_object = vl_fb_go_table_dm6_FB2020_05,
  FBgn_universe = "all", go_type = "all", cex = 1,
  padj_cutoff = 1e-05, N_top = Inf)
}
\arguments{
\item{FBgn_list}{A nameds list of FBgn IDs (1 sublist/cluster)}

\item{go_object}{object containing GO data. see ?vl_fb_go_table_dm6_FB2020_05}

\item{cex}{cex ballons (usefull to adjust size)}

\item{padj_cutoff}{padjust cutoff applied to GOs}
}
\description{
This function compares GO enrichments for several groups
}
