% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vGOcluster.R
\name{vl_GO_clusters}
\alias{vl_GO_clusters}
\title{GO analysis}
\usage{
vl_GO_clusters(
  FBgn_list,
  go_object = vl_fb_go_table_dm6_FB2020_05,
  FBgn_universe = "all",
  cex = 1
)
}
\arguments{
\item{FBgn_list}{A nameds list of FBgn IDs (1 sublist/cluster)}

\item{go_object}{object containing GO data. see ?vl_fb_go_table_dm6_FB2020_05}

\item{cex}{cex ballons (usefull to adjust size)}
}
\description{
This function compares GO enrichments for several groups
}
