% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_clusters}
\alias{vl_GO_clusters}
\title{GO analysis}
\usage{
vl_GO_clusters(
  FBgn_list,
  go_object = vl_Dmel_GO_FB2020_05,
  FBgn_universe = "all",
  go_type = "all",
  plot = T,
  padj_cutoff = 1e-05,
  log2OR_cutoff = 0,
  top_enrich = Inf,
  x_breaks,
  color_breaks,
  cex.balloons = 1,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  auto_margins = T
)
}
\arguments{
\item{FBgn_list}{A nameds list of FBgn IDs (1 sublist/cluster)}

\item{go_object}{object containing GO data. see ?vl_fb_go_table_dm6_FB2020_05}

\item{go_type}{The type of go to be considered. Default "all" means "biological_process" AND "cellular_component" AND "molecular_function".}

\item{plot}{Should the result be plot using balloons plot?}

\item{padj_cutoff}{cutoff for ballons to be ploted}

\item{log2OR_cutoff}{cutoff for ballons to be ploted}

\item{top_enrich}{Select top enriched motifs/cluster}

\item{x_breaks}{Breaks used for ballon's sizes}

\item{color_breaks}{Color breaks used for coloring}

\item{cex.balloons}{Expansion factor for balloons}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{auto_margins}{Use auto margins? Default= T}

\item{all_FBgns}{Vector of FBgns to be used in the universe, typically all the genes tested with DESeq. Default= "all" FBgns present in go_object.}
}
\description{
This function compares GO enrichments for several groups
}
\examples{
FBgn_list <- split(vl_genes_set$FBgn, vl_genes_set$GO)
par(mar= c(5,25,2,5), las= 1)
vl_GO_clusters(FBgn_list, auto_margins = F, cex.balloons = 0.3)
}
