% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_clusters}
\alias{vl_GO_clusters}
\title{GO analysis}
\usage{
vl_GO_clusters(FBgn_list, go_object = vl_fb_go_table_dm6_FB2020_05,
  all_FBgns = "all", go_type = "all", cex = 1, padj_cutoff = 1e-05,
  N_top = Inf, auto_margin = T)
}
\arguments{
\item{FBgn_list}{A nameds list of FBgn IDs (1 sublist/cluster)}

\item{go_object}{object containing GO data. see ?vl_fb_go_table_dm6_FB2020_05}

\item{all_FBgns}{Vector of FBgns to be used in the universe, typically all the genes tested with DESeq. Default= "all" FBgns present in go_object.}

\item{go_type}{The type of go to be considered. Default "all" means "biological_process" AND "cellular_component" AND "molecular_function".}

\item{cex}{cex ballons (usefull to adjust size)}

\item{padj_cutoff}{padjust cutoff applied to GOs}

\item{N_top}{Select only N top GOs/cluster}

\item{auto_margin}{Compute and apply optimal margins. Default= T}
}
\description{
This function compares GO enrichments for several groups
}
