% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_enrich}
\alias{vl_GO_enrich}
\title{GO analysis}
\usage{
vl_GO_enrich(
  geneIDs,
  geneUniverse.IDs = NULL,
  species,
  select = c("BP", "CC", "MF"),
  plot = F,
  padj.cutoff = 1e-05,
  top.enrich = Inf,
  order = "padj",
  x.breaks,
  color.breaks,
  cex.balloons = 1,
  col = c("blue", "red"),
  main = NA,
  plot.empty.clusters = T
)
}
\arguments{
\item{geneIDs}{A vector (barplot) of a list (clusters) of gene IDs}

\item{geneUniverse.IDs}{Vector of FBgn IDs to which the analysis will be restricted. If NULL, uses all genes.}

\item{species}{Chose between "Dm" and "Mm"}

\item{select}{Which annotations should be considered? Default to c("BP", "CC", "MF")}

\item{plot}{Plot result?}

\item{padj.cutoff}{cutoff for plotting}

\item{top.enrich}{Show only n top enriched motifs. Default= 10L}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{x.breaks}{Breaks used for balloons sizes}

\item{color.breaks}{Balloons color breaks}

\item{cex.balloons}{Balloons size expansion factor}

\item{col}{Color vector}

\item{main}{Title}

\item{plot.empty.clusters}{Should empty clusters be plotted? Default to TRUE}
}
\description{
This function compute GO enrichment in a group of genes
}
\examples{
RpL <- vl_genes_set[GO=="RpL_genes", FBgn]
Hox <- vl_genes_set[GO=="HOX_genes", FBgn]

# Compute and plot enrichments of RpL genes
vl_GO_enrich(RpL, species= "Dm", plot= T)

# Separately compute and plot enrichments of HOX genes
DT <- vl_GO_enrich(Hox, species= "Dm")
plot(DT, padj.cutoff= 1e-10)

# Do it for the two groups at once (useful to compare clusters...)
DT <- vl_GO_enrich(list(RpL= RpL,
                        Hox= Hox),
                   species= "Dm",
                   cex.balloons = 0.3,
                   padj.cutoff= 1e-5,
                   top.enrich = 20,
                   plot= T)
plot(DT,
     top.enrich= 5)

}
