% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_enrich}
\alias{vl_GO_enrich}
\title{GO analysis}
\usage{
vl_GO_enrich(
  FBgn_vector,
  go_object = vl_fb_go_table_dm6_FB2020_05,
  FBgn_universe = "all",
  go_type = "all",
  main = ""
)
}
\arguments{
\item{go_object}{object containing GO data. see ?vl_fb_go_table_dm6_FB2020_05}

\item{FBgn_universe}{Vector of FBgn IDs to which the analysis will be restricted. default= "all" (i.e. all the FBGns present in the table)}

\item{main}{title}

\item{FBgn_list}{A vector of FBgn IDs}
}
\description{
This function compute GO enrichment in a group of genes
}
