% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOTools.R
\name{vl_GO_enrich}
\alias{vl_GO_enrich}
\title{GO analysis}
\usage{
vl_GO_enrich(
  geneIDs,
  geneUniverse_IDs = NULL,
  species,
  select = c("BP", "CC", "MF"),
  plot = F,
  padj_cutoff = 1e-05,
  top_enrich = NA,
  order = "padj",
  x_breaks,
  color_breaks,
  cex.balloons = 1,
  col = c("blue", "red"),
  main = NA,
  plot_empty_clusters = T
)
}
\arguments{
\item{geneIDs}{A vector (barplot) of a list (clusters) of gene IDs}

\item{geneUniverse_IDs}{Vector of FBgn IDs to which the analysis will be restricted. If NULL, uses all genes.}

\item{species}{Chose between "Dm" and "Mm"}

\item{select}{Which annotations should be considered? Default to c("BP", "CC", "MF")}

\item{plot}{Plot result?}

\item{padj_cutoff}{cutoff for plotting}

\item{top_enrich}{Show only n top enriched motifs}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{x_breaks}{Breaks used for balloons sizes}

\item{color_breaks}{Balloons color breaks}

\item{cex.balloons}{Balloons size expansion factor}

\item{col}{Color vector}

\item{main}{Title}

\item{plot_empty_clusters}{Should empty clusters be plotted? Default to TRUE}
}
\description{
This function compute GO enrichment in a group of genes
}
\examples{
RpL <- vl_genes_set[GO=="RpL_genes", FBgn]
Hox <- vl_genes_set[GO=="HOX_genes", FBgn]
par(mar= c(4,20,2,6), las= 1)
vl_GO_enrich(RpL, species= "Dm", plot= T)
DT <- vl_GO_enrich(Hox, species= "Dm")
plot(DT, padj_cutoff= 1e-10)
DT <- vl_GO_enrich(list(RpL, Hox), species= "Dm", cex.balloons = 0.3, padj_cutoff= 1e-5, top_enrich = 20, plot= T)
plot(DT, top_enrich= 5)
}
