% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_heatkey}
\alias{vl_heatkey}
\title{Plots a heatkey}
\usage{
vl_heatkey(
  breaks,
  col,
  continuous = T,
  left = par("usr")[2],
  top = par("usr")[4],
  height = strheight("M") * 6,
  width = strwidth("M"),
  ticks.cex = 0.6,
  main.cex = 1,
  main = NA,
  border = "black"
)
}
\arguments{
\item{breaks}{Legend breaks}

\item{col}{Color vectors with the same length as breaks}

\item{continuous}{Are the breaks supposed to represent continuous variable? If not, draws discrete categories. Default= T}

\item{left}{left pos}

\item{top}{top pos}

\item{height}{height}

\item{width}{width}

\item{ticks.cex}{cex expansion factor for ticks labels}

\item{main.cex}{cex expansion factor for legend title}

\item{main}{Title}

\item{border}{Border color. default= "black"}
}
\value{
Plots heatkey
}
\description{
Plots a heatkey
}
