% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_heatkey}
\alias{vl_heatkey}
\title{Heatkey plot}
\usage{
vl_heatkey(
  breaks,
  col,
  left = par("usr")[2],
  top = par("usr")[4],
  height = strheight("M") * 6,
  width = strwidth("M"),
  main = NA
)
}
\arguments{
\item{breaks}{Legend breaks}

\item{col}{Color vectors with the same length as breaks}

\item{left}{left pos}

\item{top}{top pos}

\item{height}{height}

\item{width}{width}

\item{main}{Title}
}
\value{
Plots heatkey
}
\description{
Heatkey plot
}
