% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{vl_heatmap}
\alias{vl_heatmap}
\alias{vl_heatmap.data.table}
\alias{vl_heatmap.matrix}
\title{Heatmap}
\usage{
vl_heatmap(x, ...)

\method{vl_heatmap}{data.table}(x, rownames = NULL, rownames.value = NULL, ...)

\method{vl_heatmap}{matrix}(
  x,
  cluster.rows = TRUE,
  clustering.distance.rows = "euclidean",
  cutree.rows = 1L,
  row.clusters = 1L,
  kmeans.k = NA,
  cluster.cols = TRUE,
  clustering.distance.cols = "euclidean",
  cutree.cols = 1,
  col.clusters = 1,
  clustering.method = "complete",
  plot = TRUE,
  breaks = seq(min(x, na.rm = T), max(x, na.rm = TRUE), length.out = length(col)),
  col = c("cornflowerblue", "white", "red"),
  na.col = "lightgrey",
  useRaster = TRUE,
  grid = FALSE,
  main = NA,
  show.rownames = TRUE,
  show.colnames = TRUE,
  tilt.colnames = FALSE,
  show.row.dendrogram = TRUE,
  show.row.clusters = length(unique(rows$cl)) > 1,
  row.clusters.pos = "right",
  row.clusters.colors = grDevices::gray.colors(100),
  row.cluster.line.color = "black",
  show.col.dendrogram = TRUE,
  show.col.clusters = length(unique(cols$cl)) > 1,
  col.clusters.colors = grDevices::gray.colors(100),
  show.legend = TRUE,
  legend.title = NA,
  legend.cex = 1,
  display.numbers = FALSE,
  display.numbers.matrix = x,
  display.numbers.cex = 1,
  display.numbers.FUN = function(x) round(x, 2),
  box.lwd = 0.25
)
}
\arguments{
\item{x}{Data to plot. Can be one of matrix, data.table, or formula.}

\item{cluster.rows}{Should rows be clustered? Default= TRUE.}

\item{clustering.distance.rows}{Method for clustering distance rows. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall".}

\item{cutree.rows}{Number of cuts for rows tree. Default= 1L.}

\item{row.clusters}{Vector of row clusters (overwritten by clustering). Default= 1L (no clusters).}

\item{kmeans.k}{Number of row kmean clusters. If specified, takes over row clustering. Default= NA.}

\item{cluster.cols}{Should columns be clustered? Default= TRUE.}

\item{clustering.distance.cols}{Method for clustering distance cols. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall".}

\item{cutree.cols}{Number of cuts for cols tree. Default= 1L.}

\item{col.clusters}{Vector of col clusters (overwritten by clustering). Default= 1L (no clusters).}

\item{clustering.method}{Clustering method to use. Must be one of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{plot}{Should the heatmap be plotted? Default= T.}

\item{breaks}{Heatmap breaks, which should be the same length as the color vector.}

\item{col}{Color scale. default= c("cornflowerblue", "white", "red").}

\item{na.col}{Color for na values. Default= "lightgrey".}

\item{useRaster}{Should rasterImage be used? Default= TRUE.}

\item{grid}{Should grid be drawn? Default= FALSE.}

\item{main}{Title. Default= NA.}

\item{show.rownames}{Plot row names? Default= TRUE.}

\item{show.colnames}{Plot col names? Default= TRUE.}

\item{tilt.colnames}{Should colnames be tilted? default= FALSE.}

\item{show.row.dendrogram}{Plot row dendrogram? Default= TRUE.}

\item{show.row.clusters}{Plot row cluster names/lines? By default, they will only be showed if different clusters exist.}

\item{row.clusters.pos}{One of "left" or "right".}

\item{row.clusters.colors}{Row clusters colors (default is to use grey palette).}

\item{row.cluster.line.color}{Color of the line separating row clusters.}

\item{show.col.dendrogram}{Plot columns dendrogram? Default= TRUE.}

\item{show.col.clusters}{Plot column cluster names/lines?  By default, they will only be showed if different clusters exist.}

\item{col.clusters.colors}{Column cluster colors (default is to use grey palette).}

\item{show.legend}{Should the legend be plotted?}

\item{legend.title}{Title of the legend. Default= NA (no title).}

\item{legend.cex}{cex expansion factor applying to the whole legend.}

\item{display.numbers}{Display numbers on heatmap? Default= FALSE.}

\item{display.numbers.matrix}{Matrix of numbers to be displayed. By default, the x matrix will be used.}

\item{display.numbers.cex}{cex display numbers.}

\item{display.numbers.FUN}{Function to be applies to display.numbers.matrix before plotting. Default= function(x) round(x, 2).}

\item{box.lwd}{Line width of the box around the heatmap.}
}
\value{
An object of class vl_heatmap_obj containing clustering data and can be used for plotting using plot(obj, ...)
}
\description{
This function plots a heatmap and returns an S3 object of type vl_heatmap_obj which can be modified and used for later plotting.
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_heatmap(data.table)}: Just a wrapper that collapses DT to a suitable matrix

\item \code{vl_heatmap(matrix)}: Default function using matrix as an input

}}
\examples{
# Create test matrix
set.seed(1234)
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")

# Cluster using hierarhical clustering
par(mfrow=c(2,2))
vl_heatmap(test,
           cutree.rows = 3,
           cutree.cols = 2,
           show.col.clusters = F,
           legend.title = "log2FoldChange",
           main= "test")
# Cluster using kmeans clustering
res <- vl_heatmap(test,
                  kmeans.k = 4,
                  cutree.rows = 3,
                  cutree.cols = 5,
                  legend.title = "log2FoldChange",
                  main= "test")

# Play with plotting parameters
plot(res,
     col= c("blue", "yellow"),
     display.numbers= T,
     tilt.colnames= T,
     col.clusters.colors= rainbow(9))

}
