% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{vl_heatmap}
\alias{vl_heatmap}
\alias{vl_heatmap.data.table}
\alias{vl_heatmap.matrix}
\title{Heatmap}
\usage{
vl_heatmap(x, ...)

\method{vl_heatmap}{data.table}(x, rownames = NULL, rownames.value = NULL, ...)

\method{vl_heatmap}{matrix}(
  x,
  cluster.rows = TRUE,
  row.clusters = 1,
  row.clusters.col = NULL,
  kmeans.k = NA,
  clustering.distance.rows = "euclidean",
  cutree.rows = 1,
  cluster.cols = TRUE,
  col.clusters = 1,
  col.clusters.col = NULL,
  clustering.distance.cols = "euclidean",
  cutree.cols = 1,
  clustering.method = "complete",
  plot = TRUE,
  breaks = seq(min(x, na.rm = T), max(x, na.rm = TRUE), length.out = length(col)),
  col = c("cornflowerblue", "white", "red"),
  na.col = "lightgrey",
  grid = FALSE,
  main = NA,
  legend.title = NA,
  legend.cex = 1,
  show.rownames = TRUE,
  show.colnames = TRUE,
  tilt.colnames = FALSE,
  show.row.clusters = length(unique(rows$cl)) > 1,
  row.clusters.pos = "left",
  row.cluster.line.col = "black",
  show.col.clusters = length(unique(cols$cl)) > 1,
  show.row.dendrogram = TRUE,
  show.col.dendrogram = TRUE,
  show.legend = TRUE,
  display.numbers = FALSE,
  display.numbers.matrix,
  display.numbers.cex = 1,
  box.lwd = 0.25
)
}
\arguments{
\item{x}{Data to plot. Can be one of matrix, data.table, or formula.}

\item{cluster.rows}{Should rows be clustered? Default= T}

\item{row.clusters}{Vector of row clusters (overwritten by clustering, Default= 1)}

\item{row.clusters.col}{Row clusters colors (default is to use grey palette)}

\item{kmeans.k}{Number of row kmean clusters. If specified, takes over row clustering.}

\item{clustering.distance.rows}{Method for clustering distance rows. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall"}

\item{cutree.rows}{Number of cuts for rows tree. Default= 1L}

\item{cluster.cols}{Should columns be clustered? Default= T}

\item{col.clusters}{Vector of col clusters (overwritten by clustering, Default= 1)}

\item{col.clusters.col}{Col clusters colors (default is to use grey palette)}

\item{clustering.distance.cols}{Method for clustering distance cols. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall"}

\item{cutree.cols}{Number of cuts for cols tree. default= 1L}

\item{clustering.method}{Clustering method to use. Must be one of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{plot}{Should the heatmap be plotted? Default= T}

\item{breaks}{Heatmap breaks, which should be the same length as the color vector}

\item{col}{Color scale. default= c("cornflowerblue", "white", "red")}

\item{na.col}{Color for na values. Default= "lightgrey"}

\item{grid}{Should grid be drawn? Default= F}

\item{main}{Title. Default= NA}

\item{legend.title}{Character to plot as title legend}

\item{legend.cex}{cex expansion factor applying to the whole legend}

\item{show.rownames}{Plot row names? Default= T}

\item{show.colnames}{Plot col names? Default= T}

\item{tilt.colnames}{Should colnames be tilted? default= F}

\item{show.row.clusters}{Plot row cluster names/lines?}

\item{row.clusters.pos}{One of "left" or "right"}

\item{row.cluster.line.col}{Color of the line separting row clusters}

\item{show.col.clusters}{Plot col cluster names/lines?}

\item{show.row.dendrogram}{Plot row dendrogram?}

\item{show.col.dendrogram}{Plot col dendrogram?}

\item{show.legend}{Should the legend be plotted?}

\item{display.numbers}{Display numbers on heatmap? Default= F}

\item{display.numbers.matrix}{Matrix of numbers to be displayed. useful for fine tuning. Default= x}

\item{display.numbers.cex}{cex display numbers}

\item{box.lwd}{Line width of the box around the heatmap}
}
\value{
An object of class vl_heatmap_obj containing clustering data and can be used for plotting using plot(obj, ...)
}
\description{
This function plots a heatmap and returns an S3 object of type vl_heatmap_obj which can be modified and used for later plotting.
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_heatmap(data.table)}: Just a wrapper that collapses DT to a suitable matrix

\item \code{vl_heatmap(matrix)}: Default function using matrix as an input

}}
\examples{
# Create test matrix
set.seed(1234)
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")

par(mfrow=c(2,2))
res <- vl_heatmap(test,
                  kmeans.k = 4,
                  cutree.rows = 3,
                  cutree.cols = 5,
                  legend.title = "log2FoldChange",
                  main= "test")
vl_heatmap(test,
           cutree.rows = 3,
           cutree.cols = 2,
           show.col.clusters = F,
           legend.title = "log2FoldChange",
           main= "test")

}
