% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{vl_heatmap}
\alias{vl_heatmap}
\alias{vl_heatmap.data.table}
\alias{vl_heatmap.matrix}
\alias{vl_heatmap.vl_heatmap}
\title{Heatmap}
\usage{
vl_heatmap(x, ...)

\method{vl_heatmap}{data.table}(x, rownames = NULL,
  rownames.value = NULL, ...)

\method{vl_heatmap}{matrix}(x, cluster_rows = T, kmeans_k = NA,
  clustering_distance_rows = "euclidean", cutree_rows = 1,
  cluster_cols = T, clustering_distance_cols = "euclidean",
  cutree_cols = 1, clustering_method = "complete", plot = T,
  auto_margins = T, breaks = seq(min(x, na.rm = T), max(x, na.rm = T),
  length.out = length(col)), col = c("cornflowerblue", "white", "red"),
  na_col = "lightgrey", main = NA, legend_title = NA,
  show_rownames = T, show_colnames = T, show_row_clusters = T,
  show_col_clusters = T, show_row_dendrogram = T,
  show_col_dendrogram = T, show_legend = T, display_numbers = F,
  display_numbers_FUN = function(x) round(x, 2),
  display_numbers_cex = 0.5)

\method{vl_heatmap}{vl_heatmap}(obj, add, inherit_row_cl = T,
  inherit_col_cl = T, ...)
}
\arguments{
\item{x}{Data to plot. Can be one of matrix, data.table, or formula.}

\item{...}{Extra parameters to be passed to ?vl_heatmap().}

\item{cluster_rows}{Should rows be clustered? Default= T}

\item{kmeans_k}{Number of row kmean clusters. If specified, takes over row clustering.}

\item{clustering_distance_rows}{Method for clustering distance rows. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall"}

\item{cutree_rows}{Number of cuts for rows tree. Default= 1L}

\item{cluster_cols}{Should columns be clustered? Default= T}

\item{clustering_distance_cols}{Method for clustering distance cols. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall"}

\item{cutree_cols}{Number of cuts for cols tree. default= 1L}

\item{clustering_method}{Clustering method to use. Must be one of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{plot}{Should the heatmap be plotted? Default= T}

\item{auto_margins}{Should auto margins be computed and used?}

\item{breaks}{Heatmap breaks, which should be the same length as the color vector}

\item{col}{Color scale. default= c("cornflowerblue", "white", "red")}

\item{na_col}{Color for na values. Default= "lightgrey"}

\item{main}{Title. Default= NA}

\item{legend_title}{Character to plot as title legend}

\item{show_rownames}{Plot row names? Default= T}

\item{show_colnames}{Plot col names? Default= T}

\item{show_row_clusters}{Plot row cluster names/lines?}

\item{show_col_clusters}{Plot col cluster names/lines?}

\item{show_row_dendrogram}{Plot row dendrogram?}

\item{show_col_dendrogram}{Plot col dendrogram?}

\item{show_legend}{Should the legend be plotted?}

\item{display_numbers}{Display numbers on heatmap? Default= F}

\item{display_numbers_FUN}{Function to apply before displaying numbers on heatmap.}

\item{display_numbers_cex}{cex display numbers}

\item{obj}{vl_heatmap object containing the parent clustering}

\item{add}{matrix that will inhiherit obj clustering}

\item{inherit_row_cl}{Should row clustering be inherited? Default is TRUE if rownames match}

\item{inherit_col_cl}{Should col clustering be inherited? Default is TRUE if colnames match}

\item{x}{}

\item{...}{Extra arguments to be passed to vl_heatmap}
}
\value{
An object of class vl_heatmap_obj containing clustering data and can be used for plotting using plot(obj, ...)

a vl_heatmap object whose clustering is inherited from obj
}
\description{
This function plots a heatmap and returns an S3 object of type vl_heatmap_obj which can be modified and used for later plotting.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.table}: Just a wrapper that collapses DT to a suitable matrix

\item \code{matrix}: Default function using matrix as an input

\item \code{vl_heatmap}: Inherit clustering to another matrix
}}

\examples{
# Create test matrix
set.seed(1234)
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")

par(mfrow=c(2,2))
res <- vl_heatmap(test, kmeans_k = 4, cutree_rows = 3, cutree_cols = 5, legend_title = "log2FoldChange", main= "test")
vl_heatmap(test, cutree_rows = 3, cutree_cols = 2, show_col_clusters = F, legend_title = "log2FoldChange", main= "test")

}
