% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vheatmap.R
\name{vl_heatmap}
\alias{vl_heatmap}
\title{Heatmap}
\usage{
vl_heatmap(mat, newdata = NULL, breaks = NULL, show_rownames = T,
  show_colnames = T, col = c("cornflowerblue", "white", "red"),
  cluster_rows = T, clustering_distance_rows = "euclidean",
  cutree_rows = 1, cluster_cols = T,
  clustering_distance_cols = "euclidean", cutree_cols = 1,
  clustering_method = "complete", kmeans_k = NA, display_numbers = F,
  display_numbers_FUN = function(x) round(x, 2),
  display_numbers_cex = 0.5, legend_title = "legend",
  legend_pos = c(0.8, 1.15, 1, 1.2), scale = "none")
}
\arguments{
\item{mat}{Matrix to plot. Row names and col names will be reported in return}

\item{newdata}{If specified, clusters using matrix and plots new data instead}

\item{breaks}{Heatmap breaks, which should be the same length as the color vector}

\item{cluster_rows}{Should matrix rows be clustered? default is T}

\item{clustering_distance_rows}{Method for clustering distance rows. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall"}

\item{cutree_rows}{Number of cuts for rows tree. default= 1z}

\item{cluster_cols}{Should matrix cols be clustered? default is T}

\item{clustering_distance_cols}{Method for clustering distance cols. Can be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman", "kendall"}

\item{cutree_cols}{Number of cuts for cols tree. default= 1z}

\item{clustering_method}{Clustering method to use. Must be one of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{kmeans_k}{Number of row kmean clusters. If specified, takes over row clustering.}

\item{display_numbers}{Display numbers on heatmap? Default= F}

\item{display_numbers_FUN}{Function to apply before displaying numbers on heatmap.}

\item{display_numbers_cex}{cex display numbers}

\item{legend_title}{Character to plot as title legend}

\item{legend_pos}{vector(bottom, left, top, right). default= c(0.8, 1.15, 1, 1.2),}

\item{scale}{Scale matrix before clustering? Can be 'none' (default), 'column' or 'row'.}
}
\value{
A data.table containing clustering data!
}
\description{
This function plots a violin plot from either matrix,
data.table, data.frame or using formula.
}
