% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_importBed}
\alias{vl_importBed}
\alias{vl_importBed.character}
\alias{vl_importBed.GRanges}
\title{Import bed file as data.table}
\usage{
vl_importBed(bed)

\method{vl_importBed}{character}(bed)

\method{vl_importBed}{GRanges}(bed)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}
}
\value{
Imported bed
}
\description{
Imports bed as data.table and check formats
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: for bed paths

\item \code{GRanges}: for GRanges
}}

