% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_importContrib}
\alias{vl_importContrib}
\title{Import contributions object}
\usage{
vl_importContrib(
  h5,
  bed = list.files(dirname(h5), ".bed$", full.names = TRUE),
  fa = list.files(dirname(h5), ".fa$", full.names = TRUE)
)
}
\arguments{
\item{h5}{A vector of paths to h5 file(s) containing the contribution scores.}

\item{bed}{Bed file(s) containing the regions for which contributions were computed. By default, the file is searched in the same folder as the h5 file.}

\item{fa}{fasta file(s) containing the sequences for which contributions were computed. By default, the file is searched in the same folder as the h5 file.}
}
\value{
A contribution data.table containing, for each region, the contribution scores (as list) and the corresponding sequence.
}
\description{
Import contributions object
}
\examples{
dat <- vl_importContrib(h5= list.files("db/model_PH18/contributions/", "h5$", recursive = TRUE, full.names = TRUE))

}
