% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_intersectBed}
\alias{vl_intersectBed}
\title{Find closestBed regions}
\usage{
vl_intersectBed(a, b)
}
\arguments{
\item{a}{Granges or data.table from which regions overlapping b have to be returned}

\item{b}{Set of regions of interest}
}
\value{
Return "a" coor and closeet "b" coordinates together with distance
}
\description{
Return regions "a" overlapping with regions "b"
}
\examples{
a <- data.table(seqnames= "chr2L", start= c(1000,2000), end= c(2000, 3000))
b <- data.table(seqnames= "chr2L", start= 1500, end= 1600)
vl_intersectBed(a, b) 
}
