% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_isDTranges}
\alias{vl_isDTranges}
\title{Check if data.table file likely corresponds to a bed}
\usage{
vl_isDTranges(x)
}
\arguments{
\item{x}{Object to be testd}
}
\value{
boolean
}
\description{
Check if data.table file likely corresponds to a bed
}
