% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_metadata_ORFtag}
\alias{vl_metadata_ORFtag}
\title{Example metadata ORFtag}
\format{
data.table
#' \describe{
  \item{user}{User name}
  \item{batch}{batch}
  \item{screen}{screen name}
  \item{condition}{Typically one of 'sort' or 'input'}
  \item{replicate}{replicate}
  \item{barcodes}{Typically two barcodes, e.g. 'GCCTCTTC|ATTGATTC'}
  \item{sampleID}{sampleID. Should correspond to the catenation of screen, condtion and replicate; and be unique for each biological samples: all data will be merged per sampleID}
  \item{expName}{Experiment name}
  \item{species}{One of 'mouse' or 'human'}
  \item{layout}{'PAIRED' or 'SINGLE'. Will be used in combination with 'sequence' to know in which column barcodes should be found}
  \item{sequencer}{'NextSeq or HiSeq. Will be used in combination with 'layout' to know in which column barcodes should be found'}
  \item{bam_path}{The path to the VBC bam file containing the reads}
  ...
}
}
\usage{
See ?ORFtag_pipeline()
}
\description{
Metadata table example for the ORFtag pipeline. See ?ORFtag_pipeline() for further details
}
\keyword{datasets}
