% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R,
%   R/plot_methods_enrichment_analyses.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\alias{plot.vl_enr_cl}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich(
  counts.list,
  control.cl = NULL,
  names = NULL,
  plot = F,
  padj.cutoff = 1e-05,
  top.enrich = NA,
  order = "padj",
  x.breaks,
  color.breaks,
  cex.balloons = 1,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  plot.empty.clusters = T,
  add.motifs = F,
  cex.width = 1,
  cex.height = 1
)

\method{plot}{vl_enr_cl}(
  obj,
  x.breaks,
  padj.cutoff = 0.05,
  top.enrich = Inf,
  order = "padj",
  color.breaks,
  cex.balloons = 1,
  col = c("blue", "red"),
  main = NA,
  plot.empty.clusters = T
)
}
\arguments{
\item{counts.list}{List of data.table containing counts for the regions of interest and potentially control regions (see next argument)}

\item{control.cl}{IDs of clusters to be used as background. default to NULL, meaning all clusters are used except the one being tested}

\item{names}{Convenient names to be used for plotting and so on... Default (NULL) returns vl_Dmel_motifs_DB_full$motif_cluster names}

\item{plot}{Should the result be plot using balloons plot? Default to FALSE}

\item{padj.cutoff}{cutoff for ballons to be plotted}

\item{top.enrich}{Select top enriched motifs/cluster. Default to NA (All)}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{x.breaks}{Breaks used for ballon's sizes}

\item{color.breaks}{Color breaks used for coloring}

\item{cex.balloons}{Expansion factor for balloons}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{plot.empty.clusters}{Should empty clusters be plotted? Default= TRUE}

\item{add.motifs}{Should motif pwms be plotted?}

\item{cex.width}{expansion factor for motif widths}

\item{cex.height}{expansion factor for motif heights}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
\section{Functions}{
\itemize{
\item \code{plot(vl_enr_cl)}: method to plot enrichment objects (containing variable, log2OR and padj)

}}
\examples{
sel <- vl_Dmel_motifs_DB_full[collection=="jaspar", motif_ID]
top_SUHW <- vl_resizeBed(vl_SUHW_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
top_STARR <- vl_resizeBed(vl_STARR_DSCP_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
counts <- vl_motif_counts(top_SUHW, genome= "dm3", sel= sel)
control.counts <- vl_motif_counts(top_STARR, genome= "dm3", sel= sel)
counts.list <- list(SUHW= counts, ATAC= control.counts)
par(mar= c(3,20,2,5), las= 1)
DT <- vl_motif_cl_enrich(counts.list, padj.cutoff = 1e-3, plot= T, add.motifs= T)
DT <- plot(DT, padj.cutoff= 1e-5)
vl_add_motifs(DT)

}
