% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMotifClusterEnrichment.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich(obj, cl_column)
}
\arguments{
\item{obj}{A motif count object similar to the output of ?vl_motif_counts() (colnames:'motif', 'motif_counts', 'motif_name') + one cluster column!}

\item{cl_column}{name of the cluster column}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
\examples{

}
