% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R,
%   R/plot_methods_enrichment_analyses.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\alias{plot.vl_enr_cl}
\title{Compute motif enrichment for several groups}
\usage{
vl_motif_cl_enrich(
  counts.list,
  control.cl = NULL,
  names = vl_Dmel_motifs_DB_full[colnames(counts), motif_cluster, on = "motif_ID"],
  plot = F,
  padj.cutoff = 1e-05,
  top.enrich = Inf,
  min.counts = 3L,
  order = "padj",
  x.breaks,
  color.breaks,
  cex.balloons = 1,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  plot.empty.clusters = T,
  add.motifs = F,
  cex.width = 1,
  cex.height = 1
)

\method{plot}{vl_enr_cl}(
  obj,
  x.breaks,
  padj.cutoff = 0.05,
  top.enrich = Inf,
  min.counts = 3L,
  order = "padj",
  color.breaks,
  cex.balloons = 1,
  col = c("blue", "red"),
  main = NA,
  plot.empty.clusters = T
)
}
\arguments{
\item{counts.list}{List of data.table containing counts for the regions of interest and potentially control regions (see next argument).}

\item{control.cl}{IDs of clusters to be used as background. Default= NULL, meaning all clusters are used except the one being tested.}

\item{names}{Convenient names to be used for aggregating and plotting. By default, returns motif_cluster matching motif_ID in vl_Dmel_motifs_DB_full.}

\item{plot}{Should the result be plotted using balloons plot? Default to FALSE}

\item{padj.cutoff}{Cutoff for balloons to be plotted.}

\item{top.enrich}{Select top enriched motifs/cluster. Default= Inf (All).}

\item{min.counts}{The minimum number of counts required to call a hit. Default= 3L.}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj" (the default) and "log2OR".}

\item{x.breaks}{Breaks used for balloon's sizes.}

\item{color.breaks}{Color breaks used for coloring.}

\item{cex.balloons}{Expansion factor for balloons.}

\item{col}{Vector of colors used for coloring.}

\item{main}{Title. Default= NA.}

\item{plot.empty.clusters}{Should empty clusters be plotted? Default= TRUE.}

\item{add.motifs}{Should motif pwms be plotted?}

\item{cex.width}{Expansion factor for motif widths.}

\item{cex.height}{Expansion factor for motif heights.}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for each element of a list (clusters, groups....)
}
\section{Functions}{
\itemize{
\item \code{plot(vl_enr_cl)}: method to plot enrichment objects (containing variable, log2OR and padj)

}}
\examples{
# Resize example peaks
SUHW <- vl_resizeBed(vl_SUHW_top_peaks, genome = "dm3")
STARR <- vl_resizeBed(vl_STARR_DSCP_top_peaks, genome = "dm3")

# Generate same number of random regions
random <- vl_control_regions_BSgenome(bed= STARR, genome= "dm3")

# Combine the three sets of regions
combined <- rbindlist(list(SUHW= SUHW,
                           STARR= STARR,
                           random= random),
                      idcol = "cluster",
                      fill = T)

# Count JAPSPAR motifs
counts <- vl_motif_counts(combined, genome= "dm3", sel= pwm_log_odds= vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_log_odds])

# Motifs can also be counted using a custom PWMatrixList, for example for promoter motifs:
prom_db <- readRDS("/groups/stark/almeida/data/motifs/CP_motifs/CP_motifs_PWM.rds")
prom <- prom_db$Pwms_log_odds
for(i in seq(prom)) # This set needs a fix
  prom[[i]]@profileMatrix <- vl_pwm_perc_to_log2(prom_db$Pwms_perc[[i]]@profileMatrix)

prom_motifs <- vl_motif_counts(combined,
                               pwm_log_odds= prom,
                               genome= "dm3")

# Enrichment per cluster
enr <- vl_motif_cl_enrich(split(counts, combined$cluster),
                          control.cl = "random")

# Plot
vl_par()
plot(enr,
     padj.cutoff= 1e-5)

}
