% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich(
  counts_matrix,
  cl_IDs,
  control_cl = unique(cl_IDs),
  plot = F,
  padj_cutoff = 1e-05,
  log2OR_cutoff = 0,
  N_top = Inf,
  x_breaks,
  color_breaks,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  auto_margins = T
)
}
\arguments{
\item{counts_matrix}{A matrix containing motif counts (rows= sequences, cols= motifs)}

\item{cl_IDs}{vector of cluster IDs (used to split the data.table)}

\item{control_cl}{IDs of clusters to be used as background. default to unique(cl_IDs), meaning all clusters are used except the one being tested}

\item{plot}{Should the result be plot using balloons plot?}

\item{padj_cutoff}{cutoff for ballons to be ploted}

\item{log2OR_cutoff}{cutoff for ballons to be ploted}

\item{N_top}{Select top enriched motifs/cluster}

\item{x_breaks}{Breaks used for ballon's sizes}

\item{color_breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{auto_margins}{Use auto margins? Default= T}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
