% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMotifClusterEnrichment.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\title{Counts motifs}
\usage{
vl_motif_cl_enrich(obj, cl_column)
}
\arguments{
\item{bed}{Either a GRanges object or a data.table that can be coerced to it}

\item{genome}{Genome to be used for coordinates ("dm6, "dm3")}

\item{resize}{Should the regions be resize according to extend arg? Default= T}

\item{extend}{Vector containing two positive integers (e.g c(500,500) indicating how the regions should be extend around their center}

\item{sel}{Either "Dmel" (convenient set of Dmel motifs), NULL (all motifs) or a list of IDs existing in vl_Dmel_motifs_DB$metadata$motif_name}
}
\value{
Network plot.
}
\description{
Counts motif occurences in a set of regions
}
\examples{
test <- cl_motif_counts(bed = GRanges("chr3R", IRanges(c(2e6, 3e6), c(2e6, 3e6)+1e3)),
genome= "dm6",
resize= T, 
extend = c(500, 500),
sel= vl_Dmel_motifs_DB$metadata$motif_name[1:10])

}
