% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R,
%   R/plot_methods_enrichment_analyses.R
\name{vl_motif_cl_enrich}
\alias{vl_motif_cl_enrich}
\alias{plot.vl_enr_cl}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich(
  counts_matrix,
  cl_IDs,
  control_cl = unique(cl_IDs),
  plot = F,
  padj_cutoff = 1e-05,
  log2OR_cutoff = 0,
  N_top = Inf,
  x_breaks,
  color_breaks,
  cex.balloons = 1,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  auto_margins = T
)

\method{plot}{vl_enr_cl}(
  obj,
  x_breaks,
  padj_cutoff = 0.05,
  log2OR_cutoff = 0,
  N_top = Inf,
  color_breaks,
  cex.balloons = 1,
  col = c("cornflowerblue", "lightgrey", "tomato"),
  main = NA,
  auto_margins = T
)
}
\arguments{
\item{counts_matrix}{A matrix containing motif counts (rows= sequences, cols= motifs)}

\item{cl_IDs}{vector of cluster IDs (used to split the data.table)}

\item{control_cl}{IDs of clusters to be used as background. default to unique(cl_IDs), meaning all clusters are used except the one being tested}

\item{plot}{Should the result be plot using balloons plot?}

\item{padj_cutoff}{cutoff for ballons to be ploted}

\item{log2OR_cutoff}{cutoff for ballons to be ploted}

\item{N_top}{Select top enriched motifs/cluster}

\item{x_breaks}{Breaks used for ballon's sizes}

\item{color_breaks}{Color breaks used for coloring}

\item{col}{Vector of colors used for coloring}

\item{main}{Title. Default= NA}

\item{auto_margins}{Use auto margins? Default= T}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: method to plot cluster enrichment objects (containing variable, log2OR and padj)
}}

\examples{
# Sets of peaks
top_SUHW <- vl_resizeBed(vl_SUHW_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
top_STARR <- vl_resizeBed(vl_STARR_DSCP_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
rdm <- vl_random_regions_BSgenome("dm3", 2000, width= 500)

# Compute enrichment SUHW vs STARR-Seq
set <- rbind(top_SUHW,
             top_STARR, fill= T)
counts <- vl_motif_counts(set, "dm3")
enr <- vl_motif_cl_enrich(counts, 
                          cl_IDs = c(rep(1, nrow(top_SUHW)),
                                     rep(2, nrow(top_STARR))),
                          plot=F)
plot(enr, padj_cutoff= 1e-2)

# Compute enrichment of SUHW & STARR-Seq over random regions
counts <- vl_motif_counts(rbind(top_SUHW,
                                top_STARR, 
                                rdm, fill= T), "dm3")
cl_IDs <- c(rep("suhw", nrow(top_SUHW)),
            rep("STARR", nrow(top_STARR)),
            rep("rdm", nrow(rdm)))
enr <- vl_motif_cl_enrich(counts, 
                          cl_IDs = cl_IDs,
                          control_cl = "rdm",
                          plot=F)
par(las= 1)
plot(enr, padj_cutoff= 1e-20)
}
