% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_motif_cl_enrich_plot_only}
\alias{vl_motif_cl_enrich_plot_only}
\title{Compute motif enrichment}
\usage{
vl_motif_cl_enrich_plot_only(obj, padj_cutoff = 1e-05,
  log2OR_cutoff = 0, N_top = Inf, auto_margin = T)
}
\arguments{
\item{obj}{A fisher test matrix similar to ?vl_motif_cl_enrich() ouput.}

\item{padj_cutoff}{cutoff for ballons to be ploted}

\item{log2OR_cutoff}{cutoff for ballons to be ploted}

\item{N_top}{Select top enriched motifs/cluster}

\item{auto_margin}{computes optimal plotting margins}
}
\value{
Fisher test data.table.
}
\description{
Compute motif enrichment for the cluster in cl_columns, using all the lines as background
}
\examples{

}
