% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMotifCounts.R
\name{vl_motif_counts}
\alias{vl_motif_counts}
\title{Counts motifs}
\usage{
vl_motif_counts(bed, genome = "dm6",
  sel = vl_Dmel_motifs_DB_full[!is.na(vl_Dmel_motifs_DB_full$FBgn),
  motif])
}
\arguments{
\item{bed}{Either a GRanges object or a data.table that can be coerced to it}

\item{genome}{Genome to be used for coordinates ("dm6, "dm3")}

\item{sel}{List of motifs to compute. see vl_Dmel_motifs_DB_full$motif}
}
\value{
Network plot.
}
\description{
Counts motif occurences in a set of regions
}
\examples{
test <- cl_motif_counts(bed = GRanges("chr3R", IRanges(c(2e6, 3e6), c(2e6, 3e6)+1e3)),
genome= "dm6",
resize= T, 
extend = c(500, 500),
sel= vl_Dmel_motifs_DB$metadata$motif_name[1:10])

}
