% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R,
%   R/plot_methods_enrichment_analyses.R
\name{vl_motif_enrich}
\alias{vl_motif_enrich}
\alias{plot.vl_enr}
\title{Motif enrichment analysis}
\usage{
vl_motif_enrich(counts, control_counts, names = NULL, plot = F,
  padj_cutoff = 0.05, top_enrich = NA, order = "padj",
  breaks = NULL, col = c("blue", "red"), xlab = "Odd Ratio (log2)",
  add_motifs = F, cex.width = 1, cex.height = 1)

plot.vl_enr(obj, padj_cutoff = 0.05, top_enrich = NA, order = "padj",
  xlab = "Odd Ratio (log2)", col = c("blue", "red"), breaks = NULL)
}
\arguments{
\item{counts}{data.table containing counts for the regions of interest}

\item{control_counts}{data.table containing counts for control regions (data.table)}

\item{names}{Convenient names to be used for plotting and so on... Default (NULL) returns to vl_Dmel_motifs_DB_full$motif_cluster}

\item{plot}{Plot result?}

\item{padj_cutoff}{cutoff for plotting. Default to FALSE}

\item{top_enrich}{Show only n top enriched motifs}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{breaks}{Color breaks to be used. Defaults to range of filtered padj.}

\item{col}{Colors vector for bars}

\item{xlab}{x label}

\item{add_motifs}{Should motif pwms be plotted?}

\item{cex.width}{expansion factor for motif widths}

\item{cex.height}{expansion factor for motif heights}

\item{obj}{An object of class vl_enr}

\item{padj_cutoff}{padjust cutoff to be applied before plotting}

\item{top_enrich}{Top enrichments to plot (based on padj)}

\item{order}{Value to be used for ordering before selecting top enriched. Possible values are "padj", "log2OR". defaut= "padj"}

\item{xlab}{Default to "Odd Ratio (log2)"}

\item{col}{Color scale}

\item{breaks}{Color breaks to be used. Defaults to range of filtered padj.}
}
\value{
DT of enrichment values

Returns a table of enrichment which can be plot using ?plot.vl_GO_enr
}
\description{
Compute motif enrichment between a set of regions and control regions
}
\section{Functions}{
\itemize{
\item \code{plot.vl_enr}: method to plot enrichment objects (containing variable, log2OR and padj)
}}

\examples{
# Example run
sel <- vl_Dmel_motifs_DB_full[collection=="jaspar", motif_ID]
top_SUHW <- vl_resizeBed(vl_SUHW_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
top_STARR <- vl_resizeBed(vl_STARR_DSCP_top_peaks, upstream = 250, downstream = 250, genome = "dm3")
counts <- vl_motif_counts(top_SUHW, genome= "dm3", sel= sel)
control_counts <- vl_motif_counts(top_STARR, genome= "dm3", sel= sel)
par(mar= c(4,15,4,6))
DT <- vl_motif_enrich(counts, control_counts, add_motifs= T, padj_cutoff= 1e-5, plot= T)
DT <- plot(DT, padj_cutoff= 1e-5)
vl_add_motifs(DT)
}
