% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R,
%   R/plot_methods_enrichment_analyses.R
\name{vl_motif_enrich}
\alias{vl_motif_enrich}
\alias{plot.vl_enr}
\title{Motif enrichment analysis}
\usage{
vl_motif_enrich(counts, control_counts,
  collapse_clusters = colnames(counts), plot = T, padj_cutoff = 0.05,
  top_enrich = Inf, plot_motifs = F)

plot.vl_enr(obj, padj_cutoff = 0.05, top_enrich = Inf,
  xlab = "Odd Ratio (log2)", col = c("blue", "red"), plot_motifs = F,
  names, width = 1, cex.width = 1, cex.height = 1, ...)
}
\arguments{
\item{counts}{matrix of counts for the regions of interest}

\item{control_counts}{matrix of counts for control regions}

\item{collapse_clusters}{A vector of clusters of lenght ncol(counts) used to collapse motifs (only the most enriched motif/cluster will be returned). Default= colnames(counts), meaning no collapsing}

\item{plot}{Plot result?}

\item{padj_cutoff}{cutoff for plotting}

\item{top_enrich}{Show only n top enriched motifs}

\item{plot_motifs}{Should PWMs be ploted?}

\item{obj}{An object of class vl_enr}

\item{xlab}{Default to "Odd Ratio (log2)"}

\item{col}{Color scale}

\item{names}{names for bars}

\item{width}{bar width (barplot parameter)}

\item{cex.width}{expansion factor for motif widths}

\item{cex.height}{expansion factor for motif heights}

\item{...}{Extra args to be passed to barplot}

\item{padj_cutoff}{padjust cutoff to be applied before plotting}

\item{top_enrich}{Top enrichments to plot (based on padj)}
}
\value{
DT of enrichment values

Returns a table of enrichment which can be plot using ?plot.vl_GO_enr
}
\description{
Compute motif enrichment between a set of regions and control regions
}
\section{Functions}{
\itemize{
\item \code{plot.vl_enr}: method to plot enrichment objects (containing variable, log2OR and padj)
}}

\examples{
# Example run
suhw <- vl_motif_counts(vl_SUHW_top_peaks, "dm3")
ctls_regions <- vl_control_regions_BSgenome(bed= vl_SUHW_top_peaks, "dm3")
ctl <- vl_motif_counts(ctls_regions, "dm3")
pl <- vl_motif_enrich(suhw, ctl, plot= F)
par(mar= c(4,8,4,6))
plot(pl, padj_cutoff= 1e-3)
}
