% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_motif_pos}
\alias{vl_motif_pos}
\alias{vl_motif_pos.data.table}
\alias{vl_motif_pos.character}
\title{Find motif positions}
\usage{
vl_motif_pos(sequences, ...)

\method{vl_motif_pos}{data.table}(bed, genome, ...)

\method{vl_motif_pos}{character}(
  sequences,
  pwm_log_odds = vl_Dmel_motifs_DB_full[collection == "jaspar", pwms_log_odds],
  genome,
  bg = "genome",
  p.cutoff = 0.00005,
  collapse.overlapping = TRUE,
  scratch = "/scratch/stark/vloubiere/",
  sub.folder = tempdir(),
  overwrite = FALSE
)
}
\arguments{
\item{sequences}{Named character vector of sequences to analyse. If provided takes over bed argument (in the case where both are specified).}

\item{bed}{Either a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start' and 'end' columns.}

\item{genome}{Genome to be used for coordinates ("dm6, "dm3") and as background for counting motifs when bg= "genome".}

\item{bg}{Background used to find motifs. Possible values include "genome" and "even". Default= "genome"}

\item{p.cutoff}{p.value cutoff used for motif detection. For enrichment analyses based on presence/absence of a motif, high cutoff might perform better (1e-4 or 5e-5) while for regression analyses, lower cutoffs might be prefered (5e-4). Default= 5e-5 (stringent).}

\item{collapse.overlapping}{Should overlapping motifs be merged? If TRUE (default), motif instances that overlap more than 70 percent of their width are collapsed.}

\item{scratch}{Temp folder where temporary files will be saved. Useful to resume long jobs, see next argument}

\item{sub.folder}{Temp sub-folder where temporary files will be saved. Explicitly setting a sub.folder will allow the function to resume where it stoped at previous iteration. By default, sub.folder= tempdir().}

\item{overwrite}{If set to FALSE (default), motifs for which positions have already been computed and saved will be skipped.}

\item{pwms_log_odds}{A PWMatrixList (in log2 odd ratio format) for which motif matches should be mapped. Overrides sel and motifDB arguments (see above).}
}
\value{
A list of positions of length = length(sequences)
}
\description{
Compute motif positions
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_motif_pos(data.table)}: Compute motifs position within

\item \code{vl_motif_pos(character)}: Identify motif positions within sequences

}}
\examples{
# Find position of 3 different motifs within two regions
pos <- vl_motif_pos(vl_SUHW_top_peaks[1:2],
                    genome= "dm3",
                    pwm_log_odds= vl_Dmel_motifs_DB_full[c("cisbp__M2328", "flyfactorsurvey__suHw_FlyReg_FBgn0003567", "jaspar__MA0533.1"), pwms_log_odds, on= "motif_ID"])

# Starting from sequence
pos <- vl_motif_pos(sequence= "TGAGTTGTGTCTGAAATTGGGATTGCTGTTGCGACAATGCCTGTCTGACAGCATTGTCGATAAGAGCTTGAATCTGATTGGGGTCCATGGTAATATCTACCGTGGCACTATCTAACGGCCGACCTAATGCTTGGCCTACTTGCTCCTCCTCCCAGCTATCCTCGCTTTCGTATTCGACCTTAACCTTTCTGTAGTT#' ATGTGCCCAACTCATTGGTTGTTGGTTGGCACACCACAAATATACTGTTGCCGAGCACAATTGATCGGCTAAATGGTATGGCAAGAAAAGGTATGCAATATAATAATCTTTTATTGGGTATGCAACGAAAATTTGTTTCGTCAACGTATGCAATATTTTTTATTAAAAGAGGGTATGCAATGTATTTTATTAAAAACGGGTATGCAATATAATAATCTTTTATTGGG#' TATGCAACGAAAATTTGTTTCGTCAAAGTATGCAATATTTTTTATTAAAAGAGGGTATGCAATGTATTTTATTAAAAACGGGTATGCAATAAAAAATTATTTGGTTTCTCTAAAAAGTATGCAGCACTTATTTTTTGATAAGGTATGCAACAAAATTTTACTTTGCCGAAAATATGCAATGTTTTTGCGAATAAATTCAACGCACACTTATTACGTGGCCAGATACA#' CAACTTTTTTTTTTTTTTTTCACTCGTAAATTTCTTGATTGCGTCAAAGA",
                    genome= "dm3",
                    pwm_log_odds= vl_Dmel_motifs_DB_full[c("cisbp__M2328", "flyfactorsurvey__suHw_FlyReg_FBgn0003567", "jaspar__MA0533.1"), pwms_log_odds, on= "motif_ID"])

# Motifs can also be mapped using a custom PWMatrixList, for example for promoter motifs
prom_db <- readRDS("/groups/stark/almeida/data/motifs/CP_motifs/CP_motifs_PWM.rds")
prom <- prom_db$Pwms_log_odds
for(i in seq(prom))
  prom[[i]]@profileMatrix <- vl_pwm_perc_to_log2(prom_db$Pwms_perc[[i]]@profileMatrix)

# Select Ribo Protein gene promoters
proms <- rtracklayer::import("/groups/stark/annotations/dm3/dmel-all-filtered-r5.57_genes_and_transcripts_only.gff")
proms <- as.data.table(proms)
proms <- vl_resizeBed(proms[grepl("ribosomal protein", fullname)], "start", 150, 50)
proms[, seqnames:= paste0("chr", seqnames)]

prom_motifs <- vl_motif_pos(proms,
                            pwm_log_odds= prom,
                            genome= "dm3")
# Many instances of TCT motif are found, as expected
prom_motifs$TC_17_Zabidi

}
