% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAnalysisTools.R
\name{vl_motif_pos}
\alias{vl_motif_pos}
\alias{vl_motif_pos.data.table}
\alias{vl_motif_pos.character}
\title{Find motif positions}
\usage{
vl_motif_pos(sequences, ...)

\method{vl_motif_pos}{data.table}(bed, genome, ...)

\method{vl_motif_pos}{character}(
  sequences,
  sel,
  genome,
  bg = "genome",
  p.cutoff = 5e-04,
  collapse_overlapping = TRUE,
  motifDB = vl_Dmel_motifs_DB_full
)
}
\arguments{
\item{sequences}{Named character vector of sequences to analyse. If provided takes over bed argument (in the case where both are specified)}

\item{bed}{Either a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{genome}{Genome to be used for coordinates ("dm6, "dm3") and as background for counting motifs when bg= "genome"}

\item{sel}{List of motifs to compute. see vl_Dmel_motifs_DB_full$motif}

\item{bg}{Background used to find motifs. Possible values include "genome" and "even". Default= "genome"}

\item{p.cutoff}{Pval cutoff used for motif detection}

\item{collapse_overlapping}{Should overlapping motifs be merged? If TRUE (default), motif instances that overlap more than 70 percent of their width are collapsed.}
}
\value{
A list of positions of length = length(sequences)
}
\description{
Compute motif positions
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_motif_pos(data.table)}: Compute motifs position within

\item \code{vl_motif_pos(character)}: Identify motif positions within sequences

}}
\examples{
vl_motif_pos.data.table(vl_SUHW_top_peaks[1:2], genome= "dm3", sel= c("cisbp__M2328", "flyfactorsurvey__suHw_FlyReg_FBgn0003567", "jaspar__MA0533.1"))
}
