% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFeome_pipeline.R
\name{vl_ORFeome_MAGeCK}
\alias{vl_ORFeome_MAGeCK}
\title{test}
\usage{
vl_ORFeome_MAGeCK(
  sample.counts,
  input.counts,
  screen_name,
  cdition_name,
  sample.names = gsub("^(.*?_rep\\\\d+).*", "\\\\1", basename(sample.counts)),
  input.names = gsub("^(.*?_rep\\\\d+).*", "\\\\1", basename(input.counts)),
  output_folder = "db/FC_tables/ORFeome/",
  sort = "pos",
  input.cutoff.FUN = function(x) sum(x) >= 0,
  sample.cutoff.FUN = function(x) sum(x) >= 3,
  row.cutoff.FUN = function(x) sum(x) >= 0,
  input.pseudocount = 0,
  sample.pseudocount = 0,
  paired = length(sample.names) == length(input.names),
  logFC.cutoff = 1,
  FDR.cutoff = 0.05,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 2,
  mem = 8,
  overwrite = FALSE,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/ORFeome/MAGeCK",
  time = "01:00:00"
)
}
\description{
test
}
