% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFeome_pipeline.R
\name{vl_ORFeome_MAGeCK_auto}
\alias{vl_ORFeome_MAGeCK_auto}
\title{Automatic wrapper MAGeCK analysis ORFeome screen}
\usage{
vl_ORFeome_MAGeCK_auto(
  processed_metadata,
  FC_metadata_output = gsub("_processed.rds", "_FC_tables.rds"),
  screen_group_columns = c("screen", "cell_line"),
  condition_group_columns = c("condition", "sort"),
  paired = "auto",
  sort_column = "MAGeCK_sort",
  output_folder = "db/FC_tables/ORFeome/",
  input.cutoff.FUN = function(x) sum(x) >= 0,
  sample.cutoff.FUN = function(x) sum(x) >= 3,
  row.cutoff.FUN = function(x) sum(x) >= 0,
  pseudocount = 0,
  logFC.cutoff = 1,
  FDR.cutoff = 0.05,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 2,
  mem = 8,
  overwrite = FALSE,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/ORFeome/MAGeCK",
  time = "01:00:00"
)
}
\arguments{
\item{processed_metadata}{Path to the metadata file generated by vl_ORFeome_processing (in .rds or .txt format), or the corresponding data.table.}

\item{FC_metadata_output}{An .rds path where to save the FC metadata file, which contains the path to MAGeCK output files. By default, when the processed_metadata is a path to a processed_metadata files, "_FC_tables.rds" will be appended processed_metadata file path.}

\item{screen_group_columns}{Columns to be used to identify individual screens. Default= c("screen", "cell_line"),}

\item{condition_group_columns}{Columns to be used to identify individual conditions. Default= c("screen", "cell_line"),}

\item{paired}{Should samples be analyzee as paired? Possible value are "auto" (which will consider samples paired if the number of sample count files matches the number of input count files), TRUE or FALSE. Default= 'auto'.}

\item{sort_column}{The name of the column containing sort options ('pos' or 'neg'). Default= 'MAGeCK_sort'.}

\item{output_folder}{Output folder where all output files should be saved. Default= "db/FC_tables/ORFeome/"}

\item{input.cutoff.FUN}{Function to be applied to filter input columns. Default= function(x) sum(x)>=0}

\item{sample.cutoff.FUN}{Function to be applied to filter sample columns. Default= function(x) sum(x)>=3}

\item{row.cutoff.FUN}{Function to be applied to filter all columns. Default= function(x) sum(x)>=3}

\item{pseudocount}{The pseudocount to be added to 0 values (only!) in sample and input count columns (note that this value will be further normalized for sequencing depth). Default= 0}

\item{logFC.cutoff}{logFC cutoff to call hits for the MA plot. Default= 1}

\item{FDR.cutoff}{FDR cutoff to call hits for the MA plot. Default= 0.05}

\item{Rpath}{Path to an Rscript executable. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}

\item{cores}{Number of cores per job. Default= 2.}

\item{mem}{Memory per job (in Go). Default= 8.}

\item{overwrite}{Should existing files be overwritten?}

\item{submit}{Should the command be submitted? Default= FALSE.}

\item{wdir}{The working directory to use. Default= getwd(), meaning current working directory will be used.}

\item{logs}{Output folder for log files. Default= "db/logs/ORFeome/MAGeCK".}

\item{time}{The time required for the SLURM scheduler. Default= '1:00:00'.}
}
\value{
Command lines to run MAGeCK for a single screen.
}
\description{
This function is a wrapper around ?vl_ORFeome_MAGeCK() that used the provided metadata to run all comparisons.
}
\examples{
# Processed metadata ----
processed <- readRDS("Rdata/metadata_ORFeome_processed.rds")

# Call hits for all conditions using MAgECK ----
vl_ORFeome_MAGeCK_auto(processed,
                       #' FC_metadata_output = "Rdata/metadata_ORFeome_FC_tables.rds",
                       overwrite = FALSE, # Do not re-process existing files
                       submit = TRUE)
                       
# FC tables metadata ----
FC_tables <- readRDS("Rdata/metadata_ORFeome_FC_tables.rds")

}
