% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFeome_pipeline.R
\name{vl_ORFeome_processing}
\alias{vl_ORFeome_processing}
\alias{vl_ORFeome_processing.character}
\alias{vl_ORFeome_processing.default}
\title{ORFeome pipeline}
\usage{
vl_ORFeome_processing(metadata, ...)

\method{vl_ORFeome_processing}{character}(
  metadata,
  processed_metadata_output = gsub(".xlsx$", "_processed.rds", metadata),
  ...
)

\method{vl_ORFeome_processing}{default}(
  metadata,
  processed_metadata_output,
  fq_output_folder = "/scratch/stark/vloubiere/ORFeome/fq/",
  bam_output_folder = "/scratch/stark/vloubiere/ORFeome/bam/",
  alignment_stats_output_folder = "db/alignment_stats/ORFeome/",
  BC_count_output_folder = "db/counts/ORFeome/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 6,
  mem = 32,
  overwrite = FALSE,
  counts.exists = TRUE,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/ORFeome/processing",
  time = "12:00:00"
)
}
\description{
ORFeome pipeline
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_ORFeome_processing(character)}: for excel files path

\item \code{vl_ORFeome_processing(default)}: default method

}}
