% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFtag_pipeline.R
\name{vl_ORFtrap_call_hits}
\alias{vl_ORFtrap_call_hits}
\title{ORFtrap_call_hits}
\usage{
vl_ORFtrap_call_hits(
  sorted.forward.counts,
  unsorted.forward.counts,
  exons.start.gtf,
  name,
  output.suffix = "_vs_unsorted.txt",
  output.folder.FC.file
)
}
\arguments{
\item{sorted.forward.counts}{Sorted forward counts file (see bamToBed_and_assign_insertions.R function for further details)}

\item{unsorted.forward.counts}{Unsorted (input) forward counts file (see bamToBed_and_assign_insertions.R function for further details)}

\item{exons.start.gtf}{gtf exon file. Used for consistent ordering of output FC table}

\item{name}{Name to be appended at the beginning of output file}

\item{output.suffix}{Suffix to be appended at the end of output file. Default to "_vs_unsorted.txt"}

\item{output.folder.FC.file}{Output folder for FC files}
}
\value{
Returns FC tables containing DESeq2-like columns
}
\description{
Function to call hits from a given screen. Receives sorted and unsorted counts as input, computes FC table
}
\examples{
vl_ORFtrap_call_hits(sorted.forward.counts = c("db/gene_assignment/ORFtag/Activator2_sort_rep1_same_strand.txt",
                                               "db/gene_assignment/ORFtag/Activator2_sort_rep2_same_strand.txt"),
                     unsorted.forward.counts = c("db/gene_assignment/ORFtag/Activator2_input_rep1_same_strand.txt",
                                                 "db/gene_assignment/ORFtag/Activator2_input_rep2_same_strand.txt"),
                     exons.start.gtf = "/groups/stark/vloubiere/projects/ORFTRAP_1/db/gtf/exons_start_mm10.gtf",
                     name = "Activator_2",
                     output.suffix = "_vs_input.txt",
                     output.folder.FC.file = "db/FC_tables/ORFtag/Activator_2")
}
