% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFtag_pipeline.R
\name{vl_ORFtrap_call_hits_strandBias}
\alias{vl_ORFtrap_call_hits_strandBias}
\title{vl_ORFtrap_call_hits_strandBias}
\usage{
vl_ORFtrap_call_hits_strandBias(
  sorted.forward.counts,
  sorted.reverse.counts,
  unsorted.forward.counts,
  unsorted.reverse.counts,
  genome,
  name,
  output.suffix = "_vs_revStrand",
  output.folder.FC.file = "db/FC_tables/ORFtag"
)
}
\arguments{
\item{sorted.forward.counts}{Sorted forward counts file (see bamToBed_and_assign_insertions.R function for further details)}

\item{sorted.reverse.counts}{Sorted reverse counts file (see bamToBed_and_assign_insertions.R function for further details)}

\item{unsorted.forward.counts}{Unsorted (input) forward counts file (see bamToBed_and_assign_insertions.R function for further details)}

\item{unsorted.reverse.counts}{Unsorted (input) reverse counts file (see bamToBed_and_assign_insertions.R function for further details)}

\item{genome}{The genome annotation to use. For now, only "mm10" and "hg38" are supported, and will be used to retrieve non-first exon gtf file and sort output file. See at the bottom of the ?vl_ORFtag_pipeline help page to see how to generate custom non-first exon start gtf files.}

\item{name}{Name to be appended at the beginning of output file}

\item{output.suffix}{Suffix to be appended at the end of output file. Default to "_vs_revStrand.txt"}

\item{output.folder.FC.file}{Output folder for FC files. Defaults to "db/FC_tables/ORFtag"}
}
\value{
Returns FC tables containing DESeq2-like columns
}
\description{
Function to call hits from a given screen. Receives sorted and unsorted counts as input, computes FC table
}
\examples{
library(vlfunctions)

# 63 hits should be called with this dataset
vl_ORFtrap_call_hits_strandBias(sorted.forward.counts = c("db/gene_assignment/ORFtag/Activator2_sort_rep1_same_strand.txt",
                                                          "db/gene_assignment/ORFtag/Activator2_sort_rep2_same_strand.txt"),
                                sorted.reverse.counts = c("db/gene_assignment/ORFtag/Activator2_sort_rep1_rev_strand.txt",
                                                          "db/gene_assignment/ORFtag/Activator2_sort_rep2_rev_strand.txt"),
                                unsorted.forward.counts = c("db/gene_assignment/ORFtag/Activator2_input_rep1_same_strand.txt",
                                                            "db/gene_assignment/ORFtag/Activator2_input_rep2_same_strand.txt"),
                                unsorted.reverse.counts = c("db/gene_assignment/ORFtag/Activator2_input_rep1_rev_strand.txt",
                                                            "db/gene_assignment/ORFtag/Activator2_input_rep2_rev_strand.txt"),
                                genome = "mm10",
                                name = "Activator_2",
                                output.suffix = "_vs_input_strandBias.txt")
                                
}
