% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{vl_palette_blueWhiteRed}
\alias{vl_palette_blueWhiteRed}
\title{Title}
\usage{
vl_palette_blueWhiteRed(x, rep_white = 1L)
}
\arguments{
\item{x}{number of colours ro return}

\item{rep_white}{Number of times the central white color should be repeated. Higher values means further extending white. Default= 1L}
}
\value{
vector of colors of length x
}
\description{
Title
}
