% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_par}
\alias{vl_par}
\title{Sets convenient par env}
\usage{
vl_par(
  mai = c(0.9, 0.9, 0.9, 0.9),
  las = 1,
  tcl = -0.1,
  mgp = c(1.5, 0.35, 0),
  cex = 1,
  cex.lab = 9/12,
  cex.axis = 7/12,
  bty = "n",
  ...
)

vl_par(
  mai = c(0.9, 0.9, 0.9, 0.9),
  las = 1,
  tcl = -0.1,
  mgp = c(1.5, 0.35, 0),
  cex = 1,
  cex.lab = 9/12,
  cex.axis = 7/12,
  bty = "n",
  ...
)
}
\arguments{
\item{mai}{See ?par()}

\item{las}{}

\item{tcl}{}

\item{mgp}{}

\item{cex}{}

\item{cex.lab}{}

\item{cex.axis}{}

\item{bty}{}

\item{...}{Extra arguments to be passed to ?par()}

\item{bottom_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{left_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{top_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{right_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}
}
\value{
Improves default par env

Set up nice plotting parameters
}
\description{
Sets convenient par env

Sets up my favorite par parameters. Optimized for 3x3 inch square.
}
