% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_par}
\alias{vl_par}
\title{Sets convenient par env}
\usage{
vl_par(
  bottom_strings,
  left_strings,
  top_strings,
  right_strings,
  mgp = c(1.5, 0.5, 0),
  las = 1,
  tcl = -0.2,
  ...
)
}
\arguments{
\item{bottom_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{left_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{top_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{right_strings}{Character vector. If specific, margin will be adjusted to afford longest string.}

\item{mgp}{}

\item{las}{}

\item{tcl}{}

\item{...}{}
}
\value{
Improves default par env
}
\description{
Sets convenient par env
}
