% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_peakCalling}
\alias{vl_peakCalling}
\title{Compute peak calling}
\usage{
vl_peakCalling(ChIP_bed, Input_bed = NULL, BSgenome, binsize = 100)
}
\arguments{
\item{ChIP_bed}{ChIP bed file. Should be a vector of bed file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{Input_bed}{Input bed file. Default to NULL, meaning shuffle ChIP bed is used a background}

\item{BSgenome}{A BSgenome object used for gw binning}

\item{binsize}{binsize used for peak calling, Default to 100 (narrow Peaks). Use larger bins to call domains}
}
\value{
significantly enriched peaks
}
\description{
Compute peak calling using ChIP and Input
}
