% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCalling.R
\name{vl_peakCalling}
\alias{vl_peakCalling}
\title{Compute peak calling}
\usage{
vl_peakCalling(
  ChIP,
  Input,
  read.length = 50,
  fun.average = function(x) ceiling(mean(x)),
  gaussian.blur = F,
  BSgenome,
  bins.width = 100,
  steps.width = round(bins.width/2),
  bins.pval.cutoff = 0.05,
  bins.OR.cutoff = 1
)
}
\arguments{
\item{ChIP}{ChIP bed file. Should be a vector of or bw file paths, a GRange object or a data.table containing 'seqnames', 'start', 'end' columns. see ?vl_importBed()}

\item{Input}{Input file (same format as ChIP).}

\item{read.length}{If ChIP and Input are bigwig files, this argument will be used to estimate total reads. default to 50}

\item{fun.average}{Function applied to ChIP read counts to define candidate peaks}

\item{gaussian.blur}{Applies gaussian blur to ChIP signal to identify peak candidates. Useful for noisy signal. default= FALSE}

\item{BSgenome}{A BSgenome object used for gw binning}

\item{bins.width}{bins width used for peak calling, Default to 100L (narrow Peaks). Use larger bins to call domains}

\item{steps.width}{Step between bins. default to round(bins.width/2)}

\item{bins.pval.cutoff}{The cutoff that will be applied to find candidate bins and merge them into peaks. Less stringent cutoffs means broader regions. default to 0.05}

\item{bins.OR.cutoff}{Enrichment cutoff that will be applied to find candidate bins and merge them into peaks. Less stringent cutoffs means broader regions. default to 1}
}
\value{
significantly enriched peaks
}
\description{
IN DEVELOPMENT!! -> Computes peak calling using ChIP and Input
}
