% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPeakCalling.R
\name{vl_peakCalling}
\alias{vl_peakCalling}
\title{Peak calling}
\usage{
vl_peakCalling(
  ChIP_bed,
  Input_bed = NULL,
  binsize = 100,
  BSgenome = BSgenome.Dmelanogaster.UCSC.dm3,
  collapse_touching_peaks = T,
  collapse_mingap = 3 * binsize + 1,
  min_N_replicates = length(ChIP_bed),
  cutoff_input_counts = 5
)
}
\arguments{
\item{ChIP_bed}{Vector of CHIP bed files path}

\item{Input_bed}{Vector of INPUT bed files path. default= NULL uses local enrichment. Otherwise, 1 input file can be specified or as many files as for ChIP}

\item{binsize}{size of the bins used to call peaks}

\item{BSgenome}{BSgenome object used for binning the data}

\item{collapse_touching_peaks}{If set to FALSE, return all bins with related padj and OR. Else, returns collapsed reads with max OR and -log10(padj) (Default).}

\item{collapse_mingap}{mingap for collapsing peaks. see ?vl_collapse_DT_ranges().}

\item{min_N_replicates}{minimum number of replicates required to retain peak. Default= all replicates}

\item{cutoff_input_counts}{minimum number of input reads for the bin to be considered for testing. default= 5}
}
\value{
peaks
}
\description{
Compute peaks from bed files
}
\examples{
ChIP_bed <- c("../available_data_dm3/db/bed//GSE119708_ATAC_rep1_uniq.bed", "../available_data_dm3/db/bed//GSE119708_ATAC_rep2_uniq.bed")
peaks <- vl_peakCalling(ChIP_bed)

}
