% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{vl_pie}
\alias{vl_pie}
\title{Title}
\usage{
vl_pie(x, col = adjustcolor(rainbow(length(unique(x))), 0.3), ...)
}
\arguments{
\item{x}{Variable to be summarized}

\item{col}{colors to be used}

\item{...}{extra arguments passed to pie(x, ...)}
}
\value{
Plots a pie chart
}
\description{
Title
}
\examples{
vl_pie(rep(1:3, 1:3))
}
