% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepLearning_tools.R
\name{vl_plot_contrib_logo}
\alias{vl_plot_contrib_logo}
\title{Plot contribution scores}
\usage{
vl_plot_contrib_logo(
  bed,
  bw,
  genome,
  mot,
  mot.name.column = "motif_ID",
  xlab = "nt",
  ylab = "Contribution",
  xlim = c(0, length(seq)),
  ylim = range(contrib)
)
}
\arguments{
\item{bed}{A bed file containing a unique region for which contrib scores will be plotted.}

\item{bw}{A bw file containing the contribution scores.}

\item{genome}{The genome to be used.}

\item{mot}{An optional bed file containing motifs to be added.}

\item{mot.name.column}{Name of the column containing the motif name.}

\item{xlab}{Default= "nt"}

\item{ylab}{Default= "Contribution"}

\item{xlim}{Default= sequence length}

\item{ylim}{Default= range(contrib)}
}
\value{
contrib plot
}
\description{
Plot contribution scores
}
