% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_pval_text}
\alias{vl_plot_pval_text}
\title{Plot pval}
\usage{
vl_plot_pval_text(
  x,
  y,
  pval,
  stars = T,
  values = F,
  pos = 3,
  offset = ifelse(values, -0.2, -0.35),
  cex = 0.6,
  ...
)
}
\arguments{
\item{x}{X position for plotting}

\item{y}{Y position for plotting}

\item{pval}{Pval to be plotted}

\item{stars}{Should stars be printed?}

\item{values}{Should p values be printed?}

\item{pos}{pos argument (label position). see ?text()}

\item{offset}{P values offest (default= 0)}

\item{cex}{cex expansion factor passed to text}

\item{...}{Extra arguments passed to test function}
}
\description{
Plots formated pval data as text
}
\examples{
pval <- c(1e-10, 1e-5, 1e-3, 1e-2, 1e-1, 1)
plot(NA, xlim= c(0,7), ylim= c(-1,1))
vl_plot_pval(1:6, 0.5, pval)
}
