% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_pval_text}
\alias{vl_plot_pval_text}
\title{Plot pval}
\usage{
vl_plot_pval_text(
  x,
  y,
  pval,
  stars.only = F,
  show.NS = T,
  pos = 3,
  offset = 0,
  offset.star = offset - 0.1,
  cex = 0.6,
  digits = 1,
  ...
)
}
\arguments{
\item{x}{X position for plotting}

\item{y}{Y position for plotting}

\item{pval}{Pval to be plotted}

\item{stars.only}{If set to TRU, then only plots */N.S. Default= FALSE}

\item{show.NS}{Should NS values be shown?}

\item{pos}{pos argument (label position). see ?text()}

\item{offset}{Offset parameter passed to text to print N.S values}

\item{offset.star}{Adjust star (signif) offest relative to offset (used for N.S values)}

\item{cex}{cex expansion factor passed to text}

\item{...}{Extra arguments passed to test function}
}
\description{
Plots formated pval data as text
}
\examples{
pval <- c(1e-10, 1e-5, 1e-3, 1e-2, 1e-1, 1)
plot(NA, xlim= c(0,7), ylim= c(-1,1))
vl_plot_pval(1:6, 0.5, pval)
}
