% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_pval_text}
\alias{vl_plot_pval_text}
\title{Plot pval}
\usage{
vl_plot_pval_text(x, y, pval, stars_only = F, pos = 3, offset = 0, ...)
}
\arguments{
\item{x}{X position for plotting}

\item{y}{Y position for plotting}

\item{pval}{Pval to be plotted}

\item{stars_only}{If set to TRU, then only plots */N.S. Default= FALSE}

\item{pos}{pos argument (label position). see ?text()}

\item{...}{Extra arguments passed to test function}

\item{srt}{srt argument (rotation). see ?text()}
}
\description{
Plots formated pval data as text
}
\examples{
pval <- c(1e-10, 1e-5, 1e-3, 1e-2, 1e-1, 1)
plot(NA, xlim= c(0,7), ylim= c(-1,1))
vl_plot_pval(1:6, 0.5, pval)
}
