% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_plot_table}
\alias{vl_plot_table}
\title{Table plot}
\usage{
vl_plot_table(DT, wrap_text = 20)
}
\arguments{
\item{DT}{data.table}

\item{wrap_text}{Number of characters after which text can be wrapped}
}
\value{
Plots table
}
\description{
Plots a data.table as text
}
