% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROseq_pipeline.R
\name{vl_PROseq_DESeq2}
\alias{vl_PROseq_DESeq2}
\alias{vl_PROseq_DESeq2.character}
\alias{vl_PROseq_DESeq2.default}
\title{DESeq2 analysis PROseq pipeline}
\usage{
vl_PROseq_DESeq2(processed_metadata, ...)

\method{vl_PROseq_DESeq2}{character}(processed_metadata, ...)

\method{vl_PROseq_DESeq2}{default}(
  processed_metadata,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  FC_output_folder = "db/FC_tables/PROseq/",
  PDF_output_folder = "pdf/PROseq/",
  cores = 4,
  mem = 16,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/PROseq/DESeq2",
  time = "1-00:00:00"
)
}
\arguments{
\item{processed_metadata}{Path to the metadata file generated by vl_PROseq_processing, or the corresponding data.table.}

\item{Rpath}{The path to the Rscript executable to use, on which the latest version of the vlfunction package should be installed. Default= /software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript}

\item{FC_output_folder}{Folder where FC_table should be saved. Default= "db/FC_tables/PROseq/".}

\item{PDF_output_folder}{Folder where pdf files (MA plots and stats) should be saved. Default= "pdf/PROseq/".}

\item{cores}{Number of cores per job. Default= 4.}

\item{mem}{Memory per job (in Go). Default= 16.}

\item{submit}{Should the command be submitted? default= G}

\item{wdir}{The working directory to use. defaut= getwd().}

\item{logs}{Path to save logs. Default= "db/logs/CUTNRUN/peak_calling"}

\item{time}{The time required for the SLURM scheduler. Default= '1-00:00:00'}
}
\value{
Differential analyses using DESeq2
}
\description{
DESeq2 analysis PROseq pipeline
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_PROseq_DESeq2(character)}: for processed_metadata file path

\item \code{vl_PROseq_DESeq2(default)}: default method

}}
\examples{
# Make sure that all processed files exist:
processed <- readRDS("Rdata/metadata_PROseq_processed.rds")
all(file.exists(na.omit(unlist(processed[, fq1:bwNS]))))

# Differential analysis (of note, no "overwrite" option is available in this case)
vl_PROseq_DESeq2(processed, # Here, you could also provide the path to the processed metadata table ("Rdata/metadata_PROseq_processed.rds").
                 submit = TRUE)
                 
}
