% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROseq_pipeline.R
\name{vl_PROseq_DESeq2}
\alias{vl_PROseq_DESeq2}
\alias{vl_PROseq_DESeq2.character}
\alias{vl_PROseq_DESeq2.default}
\title{DESeq2 analysis PROseq pipeline}
\usage{
vl_PROseq_DESeq2(processed_metadata, ...)

\method{vl_PROseq_DESeq2}{character}(
  processed_metadata,
  FC_metadata_output = gsub("processed.rds$", "FC_tables.rds", processed_metadata),
  ...
)

\method{vl_PROseq_DESeq2}{default}(
  processed_metadata,
  FC_metadata_output,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  dds_output_folder = "db/dds/PROseq/",
  FC_output_folder = "db/FC_tables/PROseq/",
  PDF_output_folder = "pdf/PROseq/",
  cores = 4,
  mem = 16,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/PROseq/DESeq2",
  time = "1-00:00:00"
)
}
\arguments{
\item{processed_metadata}{Path to the metadata file generated by vl_PROseq_processing, or the corresponding data.table.}

\item{FC_metadata_output}{An .rds path where to save the metadata file, which contains the directories containing .dds (DESeq2 objects) and FC table files and will be used to manage them.
By default, when the processed_metadata is a path to a processed_metadata files, "_FC_tables.rds" will be appended to the excel file path.}

\item{Rpath}{Path to an Rscript executable. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}

\item{dds_output_folder}{Output folder for .dds files (DESeq2 objects). Default= "db/dds/PROseq/".}

\item{FC_output_folder}{Output folder for FC tables. Default= "db/FC_tables/PROseq/".}

\item{PDF_output_folder}{Output folder for .pdf files of MA plots and statistics. Default= "pdf/PROseq/".}

\item{cores}{Number of cores per job. Default= 4.}

\item{mem}{Memory per job (in Go). Default= 16.}

\item{submit}{Should the command be submitted? default= FALSE.}

\item{wdir}{The working directory to use. Default= getwd(), meaning current working directory will be used.}

\item{logs}{Output folder for log files. Default= "db/logs/CUTNRUN/peak_calling".}

\item{time}{The time required for the SLURM scheduler. Default= '1-00:00:00'.}
}
\value{
Differential analyses using DESeq2.
}
\description{
DESeq2 analysis PROseq pipeline
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_PROseq_DESeq2(character)}: for processed_metadata file path

\item \code{vl_PROseq_DESeq2(default)}: default method

}}
\examples{
# Make sure that all processed files exist:
processed <- readRDS("Rdata/metadata_PROseq_processed.rds")
file.exists(na.omit(unlist(processed[, fq1:bwNS])))

# Differential analysis (of note, no "overwrite" option is available in this case)
vl_PROseq_DESeq2(processed, # Here, you could also provide the path to the processed metadata table ("Rdata/metadata_PROseq_processed.rds").
                 FC_metadata_output = "Rdata/metadata_PROseq_FC_tables.rds",
                 submit = TRUE)
                 
# Once the commands ran, you can retrieve output folder using
FC <- readRDS("Rdata/metadata_PROseq_FC_tables.rds")

# For example, to retrieve spikein normalized FC tables:
files <- FC[, list.files(fcDir, paste0(DESeq2_condition, ".*", "spike")), .(fcDir, DESeq2_condition)]$V1
                 
}
