% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{vl_resizeBed}
\alias{vl_resizeBed}
\title{Find closestBed regions}
\usage{
vl_resizeBed(
  bed,
  center = "center",
  upstream = 500,
  downstream = 500,
  ignore.strand = F
)
}
\arguments{
\item{bed}{Bed file to resize. Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{center}{From where should the region be centered before extension. Either "center", "start" or "end"}

\item{upstream}{Upstream extension. default= 500L}

\item{downstream}{Downstream extension. default= 500L}

\item{ignore.strand}{Should the strand be considered when defininng start or end centering? Default= F}
}
\value{
Resized DT ranges
}
\description{
For each line of a file, returns the closest lines in b
}
