% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_pipeline.R
\name{vl_RNAseq_DESeq2}
\alias{vl_RNAseq_DESeq2}
\alias{vl_RNAseq_DESeq2.character}
\alias{vl_RNAseq_DESeq2.default}
\title{DESeq2 analysis RNAseq pipeline}
\usage{
vl_RNAseq_DESeq2(processed_metadata, ...)

\method{vl_RNAseq_DESeq2}{character}(
  processed_metadata,
  FC_metadata_output = gsub("processed.rds$", "FC_tables.rds", processed_metadata),
  ...
)

\method{vl_RNAseq_DESeq2}{default}(
  processed_metadata,
  FC_metadata_output,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  dds_output_folder = "db/dds/RNAseq/",
  FC_output_folder = "db/FC_tables/RNAseq/",
  PDF_output_folder = "pdf/RNAseq/",
  cores = 2,
  mem = 8,
  overwrite = FALSE,
  submit = FALSE,
  wdir = getwd(),
  logs = "db/logs/RNAseq/DESeq2",
  time = "01:00:00"
)
}
\arguments{
\item{processed_metadata}{Path to the metadata file generated by vl_PROseq_processing, or the corresponding data.table.}

\item{FC_metadata_output}{An .rds path where to save the metadata file, which contains the directories containing .dds (DESeq2 objects) and FC table files and will be used to manage them.
By default, when the processed_metadata is a path to a processed_metadata files, "_FC_tables.rds" will be appended to the excel file path.}

\item{Rpath}{Path to an Rscript executable. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}

\item{dds_output_folder}{Output folder for .dds files (DESeq2 objects). Default= "db/dds/PROseq/".}

\item{FC_output_folder}{Output folder for FC tables. Default= "db/FC_tables/PROseq/".}

\item{PDF_output_folder}{Output folder for .pdf files of MA plots and statistics. Default= "pdf/PROseq/".}

\item{cores}{Number of cores per job. Default= 4.}

\item{mem}{Memory per job (in Go). Default= 16.}

\item{overwrite}{Should the count tables be overwritten? Note that FC tables will always be overwritten. Default= FALSE}

\item{submit}{Should the command be submitted? Default= FALSE.}

\item{wdir}{The working directory to use. Default= getwd(), meaning current working directory will be used.}

\item{logs}{Output folder for log files. Default= "db/logs/CUTNRUN/peak_calling".}

\item{time}{The time required for the SLURM scheduler. Default= '1-00:00:00'.}

\item{count_table_output_folder}{Output folder for count tables. Default= "db/count_tables/PROseq/".}

\item{normalization}{A vector of normalization approaches to be used. Possible values are are "default" (DESeq2 default), "libSize", "spikeIn" and "combined" (Vanja's method: note that changing the control sample will change the outcome). Default= c("libSize", "spikeIn").}
}
\value{
Differential analyses using DESeq2.
}
\description{
DESeq2 analysis RNAseq pipeline
}
\section{Methods (by class)}{
\itemize{
\item \code{vl_RNAseq_DESeq2(character)}: for processed_metadata file path

\item \code{vl_RNAseq_DESeq2(default)}: default method

}}
\examples{
# List annotation files for features of interest
annot <- data.table(genome= "mm10",
                    feature= c("promoter", "gene_body", "transcript"),
.                    annotation.file= c("/groups/stark/vloubiere/projects/PROseq_pipeline/db/annotations/mm10_promoters.rds",
                                       "/groups/stark/vloubiere/projects/PROseq_pipeline/db/annotations/mm10_genebody.rds",
                                       "/groups/stark/vloubiere/projects/PROseq_pipeline/db/annotations/mm10_transcript.rds"))

# Import processed metadata file output by vl_PROseq_processing()
processed <- readRDS("Rdata/metadata_PROseq_processed.rds")

# Differential analysis
vl_PROseq_DESeq2(processed,
                 FC_metadata_output = "Rdata/metadata_PROseq_FC_tables.rds",
                 annotation.files= annot,
                 overwrite = FALSE,
                 submit = TRUE)

# Once the commands ran, you can use the FC metadata to retrieve the corresponding files:
FC <- readRDS("Rdata/metadata_PROseq_FC_tables.rds")

# For example, to retrieve spikeIn normalized files
FC[norm=="spikeIn"]

# Or for a specific feature
FC[feature=="promoter"]
                 
}
