% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshotSimple.R
\name{vl_screenshot}
\alias{vl_screenshot}
\title{Simplified version. less options, faster}
\usage{
vl_screenshot(
  bed,
  tracks,
  highlight.bed = NULL,
  col = "black",
  highlight.col = "lightgrey",
  bg.col = "white",
  density = "track",
  density.col = c("#440154FF", "#482878FF", "#3E4A89FF", "#31688EFF", "#26828EFF",
    "#1F9E89FF", "#35B779FF", "#6DCD59FF", "#B4DE2CFF", "#FDE725FF"),
  space = 30,
  widths = c(100L, 20L),
  names = NULL,
  min = 0,
  max = as.numeric(NA),
  genome,
  min.symbol = 1,
  add = F
)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{bw or bed (peaks) files.}

\item{highlight.bed}{Regions to be highlighted}

\item{col}{Track colors}

\item{highlight.col}{Color used for highlighted region}

\item{bg.col}{Color used for background. default= "white"}

\item{density}{Type of track to plot. either "track" (default) or "density"}

\item{density.col}{Color palette used for density bw. default= viridis::viridis(10)}

\item{space}{space between regions. Default is 30 px}

\item{widths}{Length 2 integer vector specifying the width of bw and bed tracks, respectively. Default= c(100L, 20L)}

\item{names}{names for bw/bed files}

\item{min}{Min value for tracks. default to 0! (Setting to NA will compute min internally)}

\item{max}{Max value for tracks (will be internally reset to 1 for non-bw files)}

\item{genome}{Genome used to plot transcripts. Available: "dm3", "dm6", "mm10", "hg19}

\item{min.symbol}{Set the min size for symbols that will be plotted. Decreasing= more symbols. default= 1}

\item{add}{Should only the tracks be added on top of existing plot? default= F}

\item{space=}{30,}

\item{widths=}{c(100L, 20L),}
}
\description{
Simplified version. less options, faster
}
\examples{
bed <- data.table(seqnames= "chr3R",
                  start= c(2166691, 12049006),
                  end= c(2949651, 12932873),
                  name= c("ANTP-C", "BX-C"))

bw <-c("/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/cutnrun/H3K27me3_PH18_merge.bw",
       "/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/cutnrun/H3K27Ac_PH18_merge.bw")

highlight_regions <- data.table(seqnames= "chr3R",
                                start= 12586652,
                                end= 12630183)

peaks <- "/groups/stark/vloubiere/projects/epigenetic_cancer/db/peaks/cutnrun/H3K27Ac_PH18_confident_peaks.narrowPeak"

vl_screenshot(bed = bed, 
              tracks = c(bw, peaks), 
              highlight.bed = highlight_regions, 
              col = c("black", "blue", "red"),
              genome = "dm6")

}
