% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshotSimple.R
\name{vl_screenshot}
\alias{vl_screenshot}
\title{Simplified version. less options, faster}
\usage{
vl_screenshot(bed, tracks, highlight_bed = NULL, col = "black",
  highlight_col = "lightgrey", space = 30, widths = c(100L, 20L),
  names = NULL, max = NULL, genome)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{bw of bed (peaks)}

\item{highlight_bed}{Regions to be highlighted}

\item{col}{Track colors}

\item{highlight_col}{Color used for highlighted region}

\item{space}{space between regions. Default is 30 px}

\item{widths}{Length 2 integer vector specifying the width of bw and bed tracks, respectively. Default= c(100L, 20L)}

\item{names}{names for bw/bed files}

\item{max}{Max value for bw tracks}

\item{genome}{Genome used to plot transcripts. Available: "dm3", "dm6", "mm10"}
}
\description{
Simplified version. less options, faster
}
\examples{
regions <- data.table(seqnames= "chr3R",
start= c(2166691, 12049006),
end= c(2949651, 12932873),
name= c("ANTP-C", "BX-C"))
vl_isDTranges(regions)

bw <-c("../available_data_dm3/db/bw/GSE41440_H3K27ac_rep1_uniq.bw",
"../available_data_dm3/db/bw/GSE41440_H3K27me3_rep1_uniq.bw")

highlight_regions <- data.table(seqnames= "chr3R",
start= 12586652,
end= 12630183)
vl_isDTranges(highlight_regions)

peaks <- "db/peaks/ATAC_peaks.txt"

vl_screenshot(bed = regions, 
tracks = c(bw, peaks), 
highlight_bed = highlight_regions, 
col = c("black", "blue", "red"),
genome = "dm3")
}
