% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vscreenshot.R
\name{vl_screenshot}
\alias{vl_screenshot}
\title{bw screenshot}
\usage{
vl_screenshot(bed, tracks, highlight_regions = NULL,
  highlight_col = "lightgrey", names = NULL, max = NULL,
  col = NULL, n_genes = 4, gband = 0.1)
}
\arguments{
\item{bed}{Regions to plot. Can be a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns"}

\item{tracks}{Vector of bw files to plot. Use full paths to avoid pbs.}

\item{highlight_regions}{Granges or data.table object specifying regions to highlight}

\item{highlight_col}{Color to use for highlighting regions. default is "lightgrey"}

\item{names}{Track names to plot. If specified, must be the same length as bw vector. By default, bw basenames will be used.}

\item{max}{Max values to clip the data. If specified, must be the same length as bw vector. By default, uses max values/track.}

\item{col}{Ploting colors. If specified, must be the same length as bw vector. By default, uses "black"}

\item{n_genes}{Max number of genes to plot.}

\item{gband}{gband ratio. default is 0.1}
}
\description{
This function plots a screenshot of bw files for a set of regions.
}
