% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshotSimple.R
\name{vl_screenshot_simple}
\alias{vl_screenshot_simple}
\title{Simplified version. less options, faster}
\usage{
vl_screenshot_simple(bed, tracks, space = 30, names = NULL, gtf = NULL)
}
\arguments{
\item{bed}{Either a vector of bed file paths, a GRanges object or a data.table containing 'seqnames', 'start', 'end' columns}

\item{tracks}{bw of bed (peaks)}

\item{space}{space between regions. Default is 30 px}

\item{names}{names for bw/bed files}

\item{gtf}{Optional gtf file(plot genes)}
}
\value{

}
\description{
Simplified version. less options, faster
}
