% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vseqlogo.R
\name{vl_seqlogo}
\alias{vl_seqlogo}
\title{plot seqlogo rasterImage}
\usage{
vl_seqlogo(pwm, xleft = 0, ybottom = 0, xright = 1, ytop = 1,
  add = T)
}
\arguments{
\item{pwm}{pwm matrix}

\item{xleft}{left plot limit. Default= 0}

\item{ybottom}{bottom plot limit. Default= 0}

\item{xright}{right plot limit. Default= 1}

\item{ytop}{top plot limit. Default= 1}

\item{add}{Should the pwm be plot on the top of opened device? Default= T}
}
\description{
plot seqlogo from pwm matrix
}
\examples{
pwm <- matrix(c(0.33,0.21,0.33,0.13,0.1,0.42,0.38,0.1,0.26,0.26,0.27,0.21,0,0.03,0.19,0.78,0.1,0.05,0.1,0.75,0.24,0.05,0.18,0.53,0.8,0.04,0,0.16,0.13,0.16,0.02,0.69,0.04,0.05,0.7,0.21,0.24,0.09,0.57,0.1,0.02,0.8,0.15,0.03,0.22,0.28,0.31,0.19,0.35,0.26,0.26,0.13,0.19,0.33,0.26,0.22), nrow= 4)
plot.new()
vl_seqlogo(pwm)
}
