% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRINGtools.R
\name{vl_STRING_interaction}
\alias{vl_STRING_interaction}
\title{get STRING intrations DT}
\usage{
vl_STRING_interaction(
  symbols = NULL,
  size = NULL,
  col = NULL,
  cex.label = NULL,
  score_cutoff = 200,
  top_N = NA,
  db_path = "/mnt/d/_R_data/genomes/dm6/STRING/dmel_7227_v11.5_vl_db.rds"
)
}
\arguments{
\item{symbols}{vector of Dmel gene symbols}

\item{size}{For each gene symbol, the size of corresponding vertice (i.e, absolute FC)}

\item{col}{For each gene symbol, the color of corresponding vertice (i.e, tomato for up, cornflowerblue for down)}

\item{cex.label}{cex factor to be applied to correponding vertices labels}

\item{score_cutoff}{Interaction score cutoff. Default is 200, max value in DB is 999. Also used to compute edge widths!}

\item{db_path}{Path to an existing db, or filename where to create it. See examples}
}
\value{
An object that can be used with the vl_STRING_network() function.
}
\description{
Extract interactions form STRIN db (Dmel only for now)
}
\examples{
#USAGE
symbols <- c("Pc", "Psc", "E(z)", "dgt1", "Rcd1")
interactions <- vl_STRING_interaction(symbols)
vl_STRING_network(interactions)


#------------#
# Code used to create DB
#------------#
# Download STRING db v11.5 fro Dmel
link <- "https://stringdb-static.org/download/protein.links.detailed.v11.5/7227.protein.links.detailed.v11.5.txt.gz"
tmp <- tempfile(fileext = ".txt.gz")
download.file(link, tmp)
DB <- fread(tmp)
# Download correspondance Dmel symbols
link <- "https://stringdb-static.org/download/protein.info.v11.5/7227.protein.info.v11.5.txt.gz"
download.file(link, tmp)
sym <- fread(tmp)
DB[sym, protein1_symbol:= i.preferred_name, on= "protein1==`#string_protein_id`"]
DB[sym, protein2_symbol:= i.preferred_name, on= "protein2==`#string_protein_id`"]
# Collapse unique interactions
DB[, merge:= paste0(sort(c(protein1_symbol, protein2_symbol)), collapse= ""), .(protein1_symbol, protein2_symbol)]
DB <- DB[, .SD[1], merge]
# SAVE
saveRDS(DB[, .(protein1_symbol, protein2_symbol, combined_score)], 
file = db_path)

}
