% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRINGtools.R
\name{vl_STRING_interaction}
\alias{vl_STRING_interaction}
\title{get STRING interactions Object}
\usage{
vl_STRING_interaction(
  STRINGdb,
  symbols,
  score_cutoff = 400,
  top_N = Inf,
  remove_non_connected = T,
  plot = T,
  size = 15,
  size2 = 15,
  color = "tomato",
  frame.color = "black",
  frame.width = 1,
  shape = "circle",
  label = symbols,
  label.family = "serif",
  label.font = 1,
  label.cex = 1,
  label.dist = 0,
  label.degree = -pi/4,
  label.color = "black"
)
}
\arguments{
\item{STRINGdb}{A STRINGdb object. See ?vl_STRING_getDB}

\item{symbols}{vector of Dmel gene symbols}

\item{score_cutoff}{If specified, only plot interactions with score>=cutoff}

\item{top_N}{If specified, only plot top N interactions}

\item{remove_non_connected}{Remove proteins that have 0 connections. Default= T}

\item{plot}{Should the graph be plotted?}

\item{size}{Vertex size. Default= 15}

\item{size2}{Vertex size2 for certain shapes. Default= 15}

\item{color}{Vertex color. Default= "tomato"}

\item{frame.color}{Frame color. Default= "black"}

\item{frame.width}{Frame width. Default= 1}

\item{shape}{Shape of the vertex. Default= "circle"}

\item{label}{Label of the vertex. Default= symbols}

\item{label.family}{Default= "serif"}

\item{label.font}{Default= 1}

\item{label.cex}{Default= 1}

\item{label.dist}{The distance of the label from the center of the vertex. If it is 0 then the label is centered on the vertex. If it is 1 then the label is displayed beside the vertex. Default= 0}

\item{label.degree}{It defines the position of the vertex labels, relative to the center of the vertices. It is interpreted as an angle in radian. Default= -pi/4.}

\item{label.color}{Default= "black"}
}
\value{
An igraph network. See ?plot.igraph()
}
\description{
Extract interactions form STRINGdb
}
\examples{
# Build database
db <- vl_STRING_getDB(species= "Dm", network_type = "full", version= "11")

# Extract interactions
symbols <- c("Pc", "Psc", "E(z)", "RpL10", "RpL11", "RpL12")
vl_STRING_interaction(STRINGdb = db,
                      symbols = symbols)
.i <- vl_STRING_interaction(STRINGdb = db,
                            symbols = symbols,
                            plot= F)
plot(.i)

#Modify ploting parameters
print(paste0("param= ", paste0(colnames(igraph::as_data_frame(.i, what= "vertices")), collapse = ", ")))
igraph::V(.i)$color <- "cornflowerblue"
igraph::V(.i)$label.cex <- c(1,1,2,2,3,3)
plot(.i)


.i$V$color <- "cornflowerblue"
.i$V$label.cex <- seq(1, 3, length.out= 6)
.i$V$size <- seq(10, 60, 10)
plot(.i)

}
