% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRINGtools.R
\name{vl_STRING_interaction}
\alias{vl_STRING_interaction}
\alias{plot.vl_STRING}
\title{get STRING intrations DT}
\usage{
vl_STRING_interaction(symbols, plot = T, score_cutoff = 900,
  top_N = NA, col = "tomato", size = 10, cex.label = 1)

plot.vl_STRING(obj, score_cutoff = 900, top_N = NA, col, size,
  cex.label)
}
\arguments{
\item{symbols}{vector of Dmel gene symbols}

\item{plot}{Should the graph be plotted?}

\item{score_cutoff}{If specified, only plot interactions with score>=cutoff}

\item{top_N}{If specified, only plot top N interactions}

\item{col}{For each gene symbol, the color of corresponding vertice (i.e, tomato for up, cornflowerblue for down)}

\item{size}{For each gene symbol, the size of corresponding vertice (i.e, absolute FC)}

\item{cex.label}{cex factor to be applied to correponding vertices labels}
}
\value{
An object that can be used with the plot.vl_STRING() method
}
\description{
Extract interactions form STRIN db (Dmel only for now)
}
\section{Functions}{
\itemize{
\item \code{plot.vl_STRING}: Method to plot STRING interaction igraphs
}}

\examples{
#USAGE
symbols <- c("Pc", "Psc", "E(z)", "RpL10", "RpL11", "RpL12")
vl_STRING_interaction(symbols)
test <- vl_STRING_interaction(symbols, col= vl_palette_categ1(6))
plot(test)
plot(test, size= 10*1:6)
plot(test, col= "red")
}
