% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vStringPlotNetwork.R
\name{vl_STRING_network}
\alias{vl_STRING_network}
\title{plot STRING interactions}
\usage{
vl_STRING_network(obj, cex.vertices = 1, label.cex = 1)
}
\arguments{
\item{cex.vertices}{Scaling factor vertices}

\item{label.cex}{Scaling factor vertices labels}

\item{symbols}{vector of Dmel gene symbols}
}
\value{
Network plot.
}
\description{
Network plotting funciton for vl_STRING_interaction output
}
\examples{
#USAGE
symbols <- c("Pc", "Psc", "E(z)", "dgt1", "Rcd1")
interactions <- vl_STRING_interaction(symbols)
vl_STRING_network(interactions)

}
