% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_thermofisher_restriction_enzymes_table}
\alias{vl_thermofisher_restriction_enzymes_table}
\title{Thermofisher enzymes}
\format{
A table containing thermofisher enzymes and related info
\describe{
  \item{name1}{enzyme name}
  \item{name2}{Alternative names}
  \item{fastdigest}{FastDigest?}
  \item{buffer}{Buffer to be used}
  \item{temperature}{Temeprature for digest}
  \item{cutsite}{cutsite pattern}
  \item{consensus_F}{consensus motif}
  ...
}
}
\source{
{"https://www.thermofisher.com/at/en/home/brands/thermo-scientific/molecular-biology/thermo-scientific-restriction-modifying-enzymes/restriction-enzymes-thermo-scientific/conventional-restriction-enzymes-thermo-scientific.html"}
}
\usage{
Can be used with vl_digest. see ?vl_digest()
}
\description{
Contains cutting sites for most commercial enzymes
}
\examples{
The table was generated using the following code:
setwd("/groups/stark/vloubiere/projects/z_miscellaneous/thermofisher_restriction_enzymes_table/")
require(rvest)
html <- read_html("https://www.thermofisher.com/at/en/home/brands/thermo-scientific/molecular-biology/thermo-scientific-restriction-modifying-enzymes/restriction-enzymes-thermo-scientific/conventional-restriction#' #' #' -enzymes-thermo-scientific.html")
dat <- as.data.table(matrix(html \%>\%
                              html_nodes("#quickchart a") \%>\%
                              html_text(), ncol= 3, byrow = T))
colnames(dat) <- c("name1", "name2", "fastdigest")
dat[, buffer := html \%>\%
      html_nodes(".optimalbuffer") \%>\%
      html_text()]
dat[, temperature:= html \%>\%
      html_nodes(".otherconditions tr:nth-child(1) td") \%>\%
      html_text()]
dat[, cutsite := html \%>\%
html_nodes("code") \%>\%
      html_text()]
dat[, consensus_F := {
  .c <- unlist(strsplit(cutsite, ""))
  .c <- .c[.c \%in\% DNA_ALPHABET[1:15]]
  paste0(.c, collapse= "")}, cutsite]
}
\keyword{datasets}
