% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlPlots.R
\name{vl_upset_plot}
\alias{vl_upset_plot}
\title{upset Plot}
\usage{
vl_upset_plot(dat_list, ylab = "Intersection size",
  intersection_cutoff = 0)
}
\arguments{
\item{dat_list}{List containing the names to intersect (see examples)}

\item{ylab}{Ylab main barplot}

\item{intersection_cutoff}{min cutoff for selections to be shown}
}
\description{
Plots an upset plot convenient to visualize intersections
}
\examples{
test <- list(A= 1:1000, B= 1:1500, C= 1000:1750)
vl_upset_plot(test)
}
